Struct octorest::apis::ActionsApi

pub struct ActionsApi<'t> { /* fields omitted */ }

This is supported on feature="gh-actions" only.

Actions API

Implementations

impl<'client> ActionsApi<'client>

pub fn list_runner_applications_for_repo(
&self
) -> ActionsListRunnerApplicationsForRepoBuilder

List runner applications for a repository

Lists binaries for the runner application that you can download and run. You must authenticate using an access token with the repo scope to use this endpoint.

See also

GitHub Developer Guide

pub fn set_selected_repos_for_org_secret(
&self
) -> ActionsSetSelectedReposForOrgSecretBuilder

Set selected repositories for an organization secret

Replaces all repositories for an organization secret when the visibility for repository access is set to selected. The visibility is set when you Create or update an organization secret. You must authenticate using an access token with the admin:org scope to use this endpoint. GitHub Apps must have the secrets organization permission to use this endpoint.

See also

GitHub Developer Guide

pub fn list_selected_repos_for_org_secret(
&self
) -> ActionsListSelectedReposForOrgSecretBuilder

List selected repositories for an organization secret

Lists all repositories that have been selected when the visibility for repository access to a secret is set to selected. You must authenticate using an access token with the admin:org scope to use this endpoint. GitHub Apps must have the secrets organization permission to use this endpoint.

See also

GitHub Developer Guide

pub fn get_workflow_run_usage(&self) -> ActionsGetWorkflowRunUsageBuilder

Get workflow run usage

Warning: This GitHub Actions usage endpoint is currently in public beta and subject to change. For more information, see "GitHub Actions API workflow usage."

Gets the number of billable minutes and total run time for a specific workflow run. Billable minutes only apply to workflows in private repositories that use GitHub-hosted runners. Usage is listed for each GitHub-hosted runner operating system in milliseconds. Any job re-runs are also included in the usage. The usage does not include the multiplier for macOS and Windows runners and is not rounded up to the nearest whole minute. For more information, see "Managing billing for GitHub Actions" in the GitHub Help documentation.

Anyone with read access to the repository can use this endpoint. If the repository is private you must use an access token with the repo scope. GitHub Apps must have the actions:read permission to use this endpoint.

See also

GitHub Developer Guide

pub fn download_artifact(&self) -> ActionsDownloadArtifactBuilder

Download an artifact

Gets a redirect URL to download an archive for a repository. This URL expires after 1 minute. Look for Location: in the response header to find the URL for the download. The :archive_format must be zip. Anyone with read access to the repository can use this endpoint. If the repository is private you must use an access token with the repo scope. GitHub Apps must have the actions:read permission to use this endpoint.

Call this endpoint using the -v flag, which enables verbose output and allows you to see the download URL in the header. To download the file into the current working directory, specify the filename using the -o flag.

See also

GitHub Developer Guide

pub fn list_org_secrets(&self) -> ActionsListOrgSecretsBuilder

List organization secrets

Lists all secrets available in an organization without revealing their encrypted values. You must authenticate using an access token with the admin:org scope to use this endpoint. GitHub Apps must have the secrets organization permission to use this endpoint.

See also

GitHub Developer Guide

pub fn list_workflow_run_artifacts(
&self
) -> ActionsListWorkflowRunArtifactsBuilder

List workflow run artifacts

Lists artifacts for a workflow run. Anyone with read access to the repository can use this endpoint. If the repository is private you must use an access token with the repo scope. GitHub Apps must have the actions:read permission to use this endpoint.

See also

GitHub Developer Guide

pub fn list_self_hosted_runners_for_repo(
&self
) -> ActionsListSelfHostedRunnersForRepoBuilder

List self-hosted runners for a repository

Lists all self-hosted runners for a repository. You must authenticate using an access token with the repo scope to use this endpoint.

See also

GitHub Developer Guide

pub fn list_repo_workflows(&self) -> ActionsListRepoWorkflowsBuilder

List repository workflows

Lists the workflows in a repository. Anyone with read access to the repository can use this endpoint. If the repository is private you must use an access token with the repo scope. GitHub Apps must have the actions:read permission to use this endpoint.

See also

GitHub Developer Guide

pub fn create_registration_token_for_org(
&self
) -> ActionsCreateRegistrationTokenForOrgBuilder

Create a registration token for an organization

Warning: The self-hosted runners API for organizations is currently in public beta and subject to change.

Returns a token that you can pass to the config script. The token expires after one hour. You must authenticate using an access token with the admin:org scope to use this endpoint.

Configure your self-hosted runner, replacing TOKEN with the registration token provided by this endpoint.

See also

GitHub Developer Guide

pub fn get_self_hosted_runner_for_org(
&self
) -> ActionsGetSelfHostedRunnerForOrgBuilder

Get a self-hosted runner for an organization

Warning: The self-hosted runners API for organizations is currently in public beta and subject to change.

Gets a specific self-hosted runner for an organization. You must authenticate using an access token with the admin:org scope to use this endpoint.

See also

GitHub Developer Guide

pub fn delete_self_hosted_runner_from_org(
&self
) -> ActionsDeleteSelfHostedRunnerFromOrgBuilder

Delete a self-hosted runner from an organization

Warning: The self-hosted runners API for organizations is currently in public beta and subject to change.

Forces the removal of a self-hosted runner from an organization. You can use this endpoint to completely remove the runner when the machine you were using no longer exists. You must authenticate using an access token with the admin:org scope to use this endpoint.

See also

GitHub Developer Guide

pub fn get_workflow(&self) -> ActionsGetWorkflowBuilder

Get a workflow

Gets a specific workflow. You can also replace :workflow_id with :workflow_file_name. For example, you could use main.yml. Anyone with read access to the repository can use this endpoint. If the repository is private you must use an access token with the repo scope. GitHub Apps must have the actions:read permission to use this endpoint.

See also

GitHub Developer Guide

pub fn remove_selected_repo_from_org_secret(
&self
) -> ActionsRemoveSelectedRepoFromOrgSecretBuilder

Remove selected repository from an organization secret

Removes a repository from an organization secret when the visibility for repository access is set to selected. The visibility is set when you Create or update an organization secret. You must authenticate using an access token with the admin:org scope to use this endpoint. GitHub Apps must have the secrets organization permission to use this endpoint.

See also

GitHub Developer Guide

pub fn add_selected_repo_to_org_secret(
&self
) -> ActionsAddSelectedRepoToOrgSecretBuilder

Add selected repository to an organization secret

Adds a repository to an organization secret when the visibility for repository access is set to selected. The visibility is set when you Create or update an organization secret. You must authenticate using an access token with the admin:org scope to use this endpoint. GitHub Apps must have the secrets organization permission to use this endpoint.

See also

GitHub Developer Guide

pub fn create_remove_token_for_org(
&self
) -> ActionsCreateRemoveTokenForOrgBuilder

Create a remove token for an organization

Warning: The self-hosted runners API for organizations is currently in public beta and subject to change.

Returns a token that you can pass to the config script to remove a self-hosted runner from an organization. The token expires after one hour. You must authenticate using an access token with the admin:org scope to use this endpoint.

To remove your self-hosted runner from an organization, replace TOKEN with the remove token provided by this endpoint.

See also

GitHub Developer Guide

pub fn create_registration_token_for_repo(
&self
) -> ActionsCreateRegistrationTokenForRepoBuilder

Create a registration token for a repository

Returns a token that you can pass to the config script. The token expires after one hour. You must authenticate using an access token with the repo scope to use this endpoint.

Configure your self-hosted runner, replacing TOKEN with the registration token provided by this endpoint.

See also

GitHub Developer Guide

pub fn list_workflow_runs_for_repo(
&self
) -> ActionsListWorkflowRunsForRepoBuilder

List workflow runs for a repository

Lists all workflow runs for a repository. You can use parameters to narrow the list of results. For more information about using parameters, see Parameters.

Anyone with read access to the repository can use this endpoint. If the repository is private you must use an access token with the repo scope. GitHub Apps must have the actions:read permission to use this endpoint.

See also

GitHub Developer Guide

pub fn get_workflow_usage(&self) -> ActionsGetWorkflowUsageBuilder

Get workflow usage

Warning: This GitHub Actions usage endpoint is currently in public beta and subject to change. For more information, see "GitHub Actions API workflow usage."

Gets the number of billable minutes used by a specific workflow during the current billing cycle. Billable minutes only apply to workflows in private repositories that use GitHub-hosted runners. Usage is listed for each GitHub-hosted runner operating system in milliseconds. Any job re-runs are also included in the usage. The usage does not include the multiplier for macOS and Windows runners and is not rounded up to the nearest whole minute. For more information, see "Managing billing for GitHub Actions" in the GitHub Help documentation.

You can also replace :workflow_id with :workflow_file_name. For example, you could use main.yml. Anyone with read access to the repository can use this endpoint. If the repository is private you must use an access token with the repo scope. GitHub Apps must have the actions:read permission to use this endpoint.

See also

GitHub Developer Guide

pub fn list_self_hosted_runners_for_org(
&self
) -> ActionsListSelfHostedRunnersForOrgBuilder

List self-hosted runners for an organization

Warning: The self-hosted runners API for organizations is currently in public beta and subject to change.

Lists all self-hosted runners for an organization. You must authenticate using an access token with the admin:org scope to use this endpoint.

See also

GitHub Developer Guide

pub fn get_org_public_key(&self) -> ActionsGetOrgPublicKeyBuilder

Get an organization public key

Gets your public key, which you need to encrypt secrets. You need to encrypt a secret before you can create or update secrets. You must authenticate using an access token with the admin:org scope to use this endpoint. GitHub Apps must have the secrets organization permission to use this endpoint.

See also

GitHub Developer Guide

pub fn create_remove_token_for_repo(
&self
) -> ActionsCreateRemoveTokenForRepoBuilder

Create a remove token for a repository

Returns a token that you can pass to remove a self-hosted runner from a repository. The token expires after one hour. You must authenticate using an access token with the repo scope to use this endpoint.

Remove your self-hosted runner from a repository, replacing TOKEN with the remove token provided by this endpoint.

See also

GitHub Developer Guide

pub fn get_job_for_workflow_run(&self) -> ActionsGetJobForWorkflowRunBuilder

Get a job for a workflow run

Gets a specific job in a workflow run. Anyone with read access to the repository can use this endpoint. If the repository is private you must use an access token with the repo scope. GitHub Apps must have the actions:read permission to use this endpoint.

See also

GitHub Developer Guide

pub fn list_artifacts_for_repo(&self) -> ActionsListArtifactsForRepoBuilder

List artifacts for a repository

Lists all artifacts for a repository. Anyone with read access to the repository can use this endpoint. If the repository is private you must use an access token with the repo scope. GitHub Apps must have the actions:read permission to use this endpoint.

See also

GitHub Developer Guide

pub fn re_run_workflow(&self) -> ActionsReRunWorkflowBuilder

Re-run a workflow

Re-runs your workflow run using its id. You must authenticate using an access token with the repo scope to use this endpoint. GitHub Apps must have the actions:write permission to use this endpoint.

See also

GitHub Developer Guide

pub fn cancel_workflow_run(&self) -> ActionsCancelWorkflowRunBuilder

Cancel a workflow run

Cancels a workflow run using its id. You must authenticate using an access token with the repo scope to use this endpoint. GitHub Apps must have the actions:write permission to use this endpoint.

See also

GitHub Developer Guide

pub fn list_workflow_runs(&self) -> ActionsListWorkflowRunsBuilder

List workflow runs

List all workflow runs for a workflow. You can also replace :workflow_id with :workflow_file_name. For example, you could use main.yml. You can use parameters to narrow the list of results. For more information about using parameters, see Parameters.

Anyone with read access to the repository can use this endpoint. If the repository is private you must use an access token with the repo scope.

See also

GitHub Developer Guide

pub fn delete_artifact(&self) -> ActionsDeleteArtifactBuilder

Delete an artifact

Deletes an artifact for a workflow run. You must authenticate using an access token with the repo scope to use this endpoint. GitHub Apps must have the actions:write permission to use this endpoint.

See also

GitHub Developer Guide

pub fn get_artifact(&self) -> ActionsGetArtifactBuilder

Get an artifact

Gets a specific artifact for a workflow run. Anyone with read access to the repository can use this endpoint. If the repository is private you must use an access token with the repo scope. GitHub Apps must have the actions:read permission to use this endpoint.

See also

GitHub Developer Guide

pub fn get_repo_public_key(&self) -> ActionsGetRepoPublicKeyBuilder

Get a repository public key

Gets your public key, which you need to encrypt secrets. You need to encrypt a secret before you can create or update secrets. Anyone with read access to the repository can use this endpoint. If the repository is private you must use an access token with the repo scope. GitHub Apps must have the secrets repository permission to use this endpoint.

See also

GitHub Developer Guide

pub fn create_or_update_org_secret(
&self
) -> ActionsCreateOrUpdateOrgSecretBuilder

Create or update an organization secret

Creates or updates an organization secret with an encrypted value. Encrypt your secret using LibSodium. You must authenticate using an access token with the admin:org scope to use this endpoint. GitHub Apps must have the secrets organization permission to use this endpoint.

Encrypt your secret using the tweetsodium library.

Encrypt your secret using pynacl with Python 3.

Encrypt your secret using the Sodium.Core package.

Encrypt your secret using the rbnacl gem.

See also

GitHub Developer Guide

pub fn delete_org_secret(&self) -> ActionsDeleteOrgSecretBuilder

Delete an organization secret

Deletes a secret in an organization using the secret name. You must authenticate using an access token with the admin:org scope to use this endpoint. GitHub Apps must have the secrets organization permission to use this endpoint.

See also

GitHub Developer Guide

pub fn get_org_secret(&self) -> ActionsGetOrgSecretBuilder

Get an organization secret

Gets a single organization secret without revealing its encrypted value. You must authenticate using an access token with the admin:org scope to use this endpoint. GitHub Apps must have the secrets organization permission to use this endpoint.

See also

GitHub Developer Guide

pub fn delete_workflow_run_logs(&self) -> ActionsDeleteWorkflowRunLogsBuilder

Delete workflow run logs

Deletes all logs for a workflow run. You must authenticate using an access token with the repo scope to use this endpoint. GitHub Apps must have the actions:write permission to use this endpoint.

See also

GitHub Developer Guide

pub fn download_workflow_run_logs(
&self
) -> ActionsDownloadWorkflowRunLogsBuilder

Download workflow run logs

Gets a redirect URL to download an archive of log files for a workflow run. This link expires after 1 minute. Look for Location: in the response header to find the URL for the download. Anyone with read access to the repository can use this endpoint. If the repository is private you must use an access token with the repo scope. GitHub Apps must have the actions:read permission to use this endpoint.

Call this endpoint using the -v flag, which enables verbose output and allows you to see the download URL in the header. To download the file into the current working directory, specify the filename using the -o flag.

See also

GitHub Developer Guide

pub fn get_repo_secret(&self) -> ActionsGetRepoSecretBuilder

Get a repository secret

Gets a single repository secret without revealing its encrypted value. You must authenticate using an access token with the repo scope to use this endpoint. GitHub Apps must have the secrets repository permission to use this endpoint.

See also

GitHub Developer Guide

pub fn create_or_update_repo_secret(
&self
) -> ActionsCreateOrUpdateRepoSecretBuilder

Create or update a repository secret

Creates or updates a repository secret with an encrypted value. Encrypt your secret using LibSodium. You must authenticate using an access token with the repo scope to use this endpoint. GitHub Apps must have the secrets repository permission to use this endpoint.

Encrypt your secret using the tweetsodium library.

Encrypt your secret using pynacl with Python 3.

Encrypt your secret using the Sodium.Core package.

Encrypt your secret using the rbnacl gem.

See also

GitHub Developer Guide

pub fn delete_repo_secret(&self) -> ActionsDeleteRepoSecretBuilder

Delete a repository secret

Deletes a secret in a repository using the secret name. You must authenticate using an access token with the repo scope to use this endpoint. GitHub Apps must have the secrets repository permission to use this endpoint.

See also

GitHub Developer Guide

pub fn list_repo_secrets(&self) -> ActionsListRepoSecretsBuilder

List repository secrets

Lists all secrets available in a repository without revealing their encrypted values. You must authenticate using an access token with the repo scope to use this endpoint. GitHub Apps must have the secrets repository permission to use this endpoint.

See also

GitHub Developer Guide

pub fn delete_self_hosted_runner_from_repo(
&self
) -> ActionsDeleteSelfHostedRunnerFromRepoBuilder

Delete a self-hosted runner from a repository

Forces the removal of a self-hosted runner from a repository. You can use this endpoint to completely remove the runner when the machine you were using no longer exists. You must authenticate using an access token with the repo scope to use this endpoint.

See also

GitHub Developer Guide

pub fn get_self_hosted_runner_for_repo(
&self
) -> ActionsGetSelfHostedRunnerForRepoBuilder

Get a self-hosted runner for a repository

Gets a specific self-hosted runner. You must authenticate using an access token with the repo scope to use this endpoint.

See also

GitHub Developer Guide

pub fn download_job_logs_for_workflow_run(
&self
) -> ActionsDownloadJobLogsForWorkflowRunBuilder

Download job logs for a workflow run

Gets a redirect URL to download a plain text file of logs for a workflow job. This link expires after 1 minute. Look for Location: in the response header to find the URL for the download. Anyone with read access to the repository can use this endpoint. If the repository is private you must use an access token with the repo scope. GitHub Apps must have the actions:read permission to use this endpoint.

Call this endpoint using the -v flag, which enables verbose output and allows you to see the download URL in the header. To download the file into the current working directory, specify the filename using the -o flag.

See also

GitHub Developer Guide

pub fn get_workflow_run(&self) -> ActionsGetWorkflowRunBuilder

Get a workflow run

Gets a specific workflow run. Anyone with read access to the repository can use this endpoint. If the repository is private you must use an access token with the repo scope. GitHub Apps must have the actions:read permission to use this endpoint.

See also

GitHub Developer Guide

pub fn list_runner_applications_for_org(
&self
) -> ActionsListRunnerApplicationsForOrgBuilder

List runner applications for an organization

Warning: The self-hosted runners API for organizations is currently in public beta and subject to change.

Lists binaries for the runner application that you can download and run. You must authenticate using an access token with the admin:org scope to use this endpoint.

See also

GitHub Developer Guide

pub fn list_jobs_for_workflow_run(&self) -> ActionsListJobsForWorkflowRunBuilder

List jobs for a workflow run

Lists jobs for a workflow run. Anyone with read access to the repository can use this endpoint. If the repository is private you must use an access token with the repo scope. GitHub Apps must have the actions:read permission to use this endpoint. You can use parameters to narrow the list of results. For more information about using parameters, see Parameters.

See also

GitHub Developer Guide