Struct octorest::apis::ActivityApi

pub struct ActivityApi<'t> { /* fields omitted */ }

This is supported on feature="gh-activity" only.

Activity API

Implementations

impl<'client> ActivityApi<'client>

pub fn list_public_events(&self) -> ActivityListPublicEventsBuilder

List public events

We delay the public events feed by five minutes, which means the most recent event returned by the public events API actually occurred at least five minutes ago.

See also

GitHub Developer Guide

pub fn list_repo_notifications_for_authenticated_user(
&self
) -> ActivityListRepoNotificationsForAuthenticatedUserBuilder

List repository notifications for the authenticated user

List all notifications for the current user.

See also

GitHub Developer Guide

pub fn mark_repo_notifications_as_read(
&self
) -> ActivityMarkRepoNotificationsAsReadBuilder

Mark repository notifications as read

Marks all notifications in a repository as "read" removes them from the default view on GitHub. If the number of notifications is too large to complete in one request, you will receive a 202 Accepted status and GitHub will run an asynchronous process to mark notifications as "read." To check whether any "unread" notifications remain, you can use the List repository notifications for the authenticated user endpoint and pass the query parameter all=false.

See also

GitHub Developer Guide

pub fn list_notifications_for_authenticated_user(
&self
) -> ActivityListNotificationsForAuthenticatedUserBuilder

List notifications for the authenticated user

List all notifications for the current user, sorted by most recently updated.

The following example uses the since parameter to list notifications that have been updated after the specified time.

See also

GitHub Developer Guide

pub fn mark_notifications_as_read(
&self
) -> ActivityMarkNotificationsAsReadBuilder

Mark notifications as read

Marks all notifications as "read" removes it from the default view on GitHub. If the number of notifications is too large to complete in one request, you will receive a 202 Accepted status and GitHub will run an asynchronous process to mark notifications as "read." To check whether any "unread" notifications remain, you can use the List notifications for the authenticated user endpoint and pass the query parameter all=false.

See also

GitHub Developer Guide

pub fn list_watched_repos_for_authenticated_user(
&self
) -> ActivityListWatchedReposForAuthenticatedUserBuilder

List repositories watched by the authenticated user

Lists repositories the authenticated user is watching.

See also

GitHub Developer Guide

pub fn list_repo_events(&self) -> ActivityListRepoEventsBuilder

List repository events

See also

GitHub Developer Guide

pub fn list_public_events_for_user(
&self
) -> ActivityListPublicEventsForUserBuilder

List public events for a user

See also

GitHub Developer Guide

pub fn list_stargazers_for_repo(&self) -> ActivityListStargazersForRepoBuilder

List stargazers

Lists the people that have starred the repository.

You can also find out when stars were created by passing the following custom media type via the Accept header:

See also

GitHub Developer Guide

pub fn list_public_org_events(&self) -> ActivityListPublicOrgEventsBuilder

List public organization events

See also

GitHub Developer Guide

pub fn list_watchers_for_repo(&self) -> ActivityListWatchersForRepoBuilder

List watchers

Lists the people watching the specified repository.

See also

GitHub Developer Guide

pub fn list_events_for_authenticated_user(
&self
) -> ActivityListEventsForAuthenticatedUserBuilder

List events for the authenticated user

If you are authenticated as the given user, you will see your private events. Otherwise, you'll only see public events.

See also

GitHub Developer Guide

pub fn list_received_events_for_user(
&self
) -> ActivityListReceivedEventsForUserBuilder

List events received by the authenticated user

These are events that you've received by watching repos and following users. If you are authenticated as the given user, you will see private events. Otherwise, you'll only see public events.

See also

GitHub Developer Guide

pub fn get_feeds(&self) -> ActivityGetFeedsBuilder

Get feeds

GitHub provides several timeline resources in Atom format. The Feeds API lists all the feeds available to the authenticated user:

Timeline: The GitHub global public timeline

User: The public timeline for any user, using URI template

Current user public: The public timeline for the authenticated user

Current user: The private timeline for the authenticated user

Current user actor: The private timeline for activity created by the authenticated user

Current user organizations: The private timeline for the organizations the authenticated user is a member of.

Security advisories: A collection of public announcements that provide information about security-related vulnerabilities in software on GitHub.

Note: Private feeds are only returned when authenticating via Basic Auth since current feed URIs use the older, non revocable auth tokens.

See also

GitHub Developer Guide

pub fn list_repos_starred_by_authenticated_user(
&self
) -> ActivityListReposStarredByAuthenticatedUserBuilder

List repositories starred by the authenticated user

Lists repositories the authenticated user has starred.

You can also find out when stars were created by passing the following custom media type via the Accept header:

See also

GitHub Developer Guide

pub fn star_repo_for_authenticated_user(
&self
) -> ActivityStarRepoForAuthenticatedUserBuilder

Star a repository for the authenticated user

Note that you'll need to set Content-Length to zero when calling out to this endpoint. For more information, see "HTTP verbs."

See also

GitHub Developer Guide

pub fn unstar_repo_for_authenticated_user(
&self
) -> ActivityUnstarRepoForAuthenticatedUserBuilder

Unstar a repository for the authenticated user

See also

GitHub Developer Guide

pub fn check_repo_is_starred_by_authenticated_user(
&self
) -> ActivityCheckRepoIsStarredByAuthenticatedUserBuilder

Check if a repository is starred by the authenticated user

See also

GitHub Developer Guide

pub fn set_thread_subscription(&self) -> ActivitySetThreadSubscriptionBuilder

Set a thread subscription

If you are watching a repository, you receive notifications for all threads by default. Use this endpoint to ignore future notifications for threads until you comment on the thread or get an @mention.

You can also use this endpoint to subscribe to threads that you are currently not receiving notifications for or to subscribed to threads that you have previously ignored.

Unsubscribing from a conversation in a repository that you are not watching is functionally equivalent to the Delete a thread subscription endpoint.

See also

GitHub Developer Guide

pub fn get_thread_subscription_for_authenticated_user(
&self
) -> ActivityGetThreadSubscriptionForAuthenticatedUserBuilder

Get a thread subscription for the authenticated user

This checks to see if the current user is subscribed to a thread. You can also get a repository subscription.

Note that subscriptions are only generated if a user is participating in a conversation--for example, they've replied to the thread, were @mentioned, or manually subscribe to a thread.

See also

GitHub Developer Guide

pub fn delete_thread_subscription(
&self
) -> ActivityDeleteThreadSubscriptionBuilder

Delete a thread subscription

Mutes all future notifications for a conversation until you comment on the thread or get an @mention. If you are watching the repository of the thread, you will still receive notifications. To ignore future notifications for a repository you are watching, use the Set a thread subscription endpoint and set ignore to true.

See also

GitHub Developer Guide

pub fn list_org_events_for_authenticated_user(
&self
) -> ActivityListOrgEventsForAuthenticatedUserBuilder

List organization events for the authenticated user

This is the user's organization dashboard. You must be authenticated as the user to view this.

See also

GitHub Developer Guide

pub fn list_repos_watched_by_user(
&self
) -> ActivityListReposWatchedByUserBuilder

List repositories watched by a user

Lists repositories a user is watching.

See also

GitHub Developer Guide

pub fn list_received_public_events_for_user(
&self
) -> ActivityListReceivedPublicEventsForUserBuilder

List public events received by a user

See also

GitHub Developer Guide

pub fn list_public_events_for_repo_network(
&self
) -> ActivityListPublicEventsForRepoNetworkBuilder

List public events for a network of repositories

See also

GitHub Developer Guide

pub fn mark_thread_as_read(&self) -> ActivityMarkThreadAsReadBuilder

Mark a thread as read

See also

GitHub Developer Guide

pub fn get_thread(&self) -> ActivityGetThreadBuilder

Get a thread

See also

GitHub Developer Guide

pub fn set_repo_subscription(&self) -> ActivitySetRepoSubscriptionBuilder

Set a repository subscription

If you would like to watch a repository, set subscribed to true. If you would like to ignore notifications made within a repository, set ignored to true. If you would like to stop watching a repository, delete the repository's subscription completely.

See also

GitHub Developer Guide

pub fn delete_repo_subscription(&self) -> ActivityDeleteRepoSubscriptionBuilder

Delete a repository subscription

This endpoint should only be used to stop watching a repository. To control whether or not you wish to receive notifications from a repository, set the repository's subscription manually.

See also

GitHub Developer Guide

pub fn get_repo_subscription(&self) -> ActivityGetRepoSubscriptionBuilder

Get a repository subscription

See also

GitHub Developer Guide

pub fn list_repos_starred_by_user(
&self
) -> ActivityListReposStarredByUserBuilder

List repositories starred by a user

Lists repositories a user has starred.

You can also find out when stars were created by passing the following custom media type via the Accept header:

See also

GitHub Developer Guide