Struct octorest::apis::AppsApi

pub struct AppsApi<'t> { /* fields omitted */ }

This is supported on feature="gh-apps" only.

Apps API

Implementations

impl<'client> AppsApi<'client>

pub fn get_user_installation(&self) -> AppsGetUserInstallationBuilder

Get a user installation for the authenticated app

Enables an authenticated GitHub App to find the user’s installation information.

You must use a JWT to access this endpoint.

See also

GitHub Developer Guide

pub fn revoke_authorization_for_application(
&self
) -> AppsRevokeAuthorizationForApplicationBuilder

Revoke an authorization for an application

Deprecation Notice: GitHub will replace and discontinue OAuth endpoints containing access_token in the path parameter. We are introducing new endpoints that allow you to securely manage tokens for OAuth Apps by using access_token as an input parameter. The OAuth Application API will be removed on May 5, 2021. For more information, including scheduled brownouts, see the blog post.

OAuth application owners can revoke a single token for an OAuth application. You must use Basic Authentication when accessing this endpoint, using the OAuth application's client_id and client_secret as the username and password.

See also

GitHub Developer Guide

pub fn check_authorization(&self) -> AppsCheckAuthorizationBuilder

Check an authorization

Deprecation Notice: GitHub will replace and discontinue OAuth endpoints containing access_token in the path parameter. We are introducing new endpoints that allow you to securely manage tokens for OAuth Apps by using access_token as an input parameter. The OAuth Application API will be removed on May 5, 2021. For more information, including scheduled brownouts, see the blog post.

OAuth applications can use a special API method for checking OAuth token validity without exceeding the normal rate limits for failed login attempts. Authentication works differently with this particular endpoint. You must use Basic Authentication when accessing this endpoint, using the OAuth application's client_id and client_secret as the username and password. Invalid tokens will return 404 NOT FOUND.

See also

GitHub Developer Guide

pub fn reset_authorization(&self) -> AppsResetAuthorizationBuilder

Reset an authorization

Deprecation Notice: GitHub will replace and discontinue OAuth endpoints containing access_token in the path parameter. We are introducing new endpoints that allow you to securely manage tokens for OAuth Apps by using access_token as an input parameter. The OAuth Application API will be removed on May 5, 2021. For more information, including scheduled brownouts, see the blog post.

OAuth applications can use this API method to reset a valid OAuth token without end-user involvement. Applications must save the "token" property in the response because changes take effect immediately. You must use Basic Authentication when accessing this endpoint, using the OAuth application's client_id and client_secret as the username and password. Invalid tokens will return 404 NOT FOUND.

See also

GitHub Developer Guide

pub fn get_repo_installation(&self) -> AppsGetRepoInstallationBuilder

Get a repository installation for the authenticated app

Enables an authenticated GitHub App to find the repository's installation information. The installation's account type will be either an organization or a user account, depending which account the repository belongs to.

You must use a JWT to access this endpoint.

See also

GitHub Developer Guide

pub fn list_plans(&self) -> AppsListPlansBuilder

List plans

Lists all plans that are part of your GitHub Marketplace listing.

GitHub Apps must use a JWT to access this endpoint. OAuth Apps must use basic authentication with their client ID and client secret to access this endpoint.

See also

GitHub Developer Guide

pub fn get_by_slug(&self) -> AppsGetBySlugBuilder

Get an app

Note: The :app_slug is just the URL-friendly name of your GitHub App. You can find this on the settings page for your GitHub App (e.g., https://github.com/settings/apps/:app_slug).

If the GitHub App you specify is public, you can access this endpoint without authenticating. If the GitHub App you specify is private, you must authenticate with a personal access token or an installation access token to access this endpoint.

See also

GitHub Developer Guide

pub fn list_installation_repos_for_authenticated_user(
&self
) -> AppsListInstallationReposForAuthenticatedUserBuilder

List repositories accessible to the user access token

List repositories that the authenticated user has explicit permission (:read, :write, or :admin) to access for an installation.

The authenticated user has explicit permission to access repositories they own, repositories where they are a collaborator, and repositories that they can access through an organization membership.

You must use a user-to-server OAuth access token, created for a user who has authorized your GitHub App, to access this endpoint.

The access the user has to each repository is included in the hash under the permissions key.

See also

GitHub Developer Guide

pub fn remove_repo_from_installation(
&self
) -> AppsRemoveRepoFromInstallationBuilder

Remove a repository from an app installation

Remove a single repository from an installation. The authenticated user must have admin access to the repository.

You must use a personal access token (which you can create via the command line or the OAuth Authorizations API) or Basic Authentication to access this endpoint.

See also

GitHub Developer Guide

pub fn add_repo_to_installation(&self) -> AppsAddRepoToInstallationBuilder

Add a repository to an app installation

Add a single repository to an installation. The authenticated user must have admin access to the repository.

You must use a personal access token (which you can create via the command line or the OAuth Authorizations API) or Basic Authentication to access this endpoint.

See also

GitHub Developer Guide

pub fn list_accounts_for_plan(&self) -> AppsListAccountsForPlanBuilder

List accounts for a plan

Returns user and organization accounts associated with the specified plan, including free plans. For per-seat pricing, you see the list of accounts that have purchased the plan, including the number of seats purchased. When someone submits a plan change that won't be processed until the end of their billing cycle, you will also see the upcoming pending change.

GitHub Apps must use a JWT to access this endpoint. OAuth Apps must use basic authentication with their client ID and client secret to access this endpoint.

See also

GitHub Developer Guide

pub fn get_authenticated(&self) -> AppsGetAuthenticatedBuilder

Get the authenticated app

Returns the GitHub App associated with the authentication credentials used. To see how many app installations are associated with this GitHub App, see the installations_count in the response. For more details about your app's installations, see the "List installations for the authenticated app" endpoint.

You must use a JWT to access this endpoint.

See also

GitHub Developer Guide

pub fn revoke_grant_for_application(
&self
) -> AppsRevokeGrantForApplicationBuilder

Revoke a grant for an application

Deprecation Notice: GitHub will replace and discontinue OAuth endpoints containing access_token in the path parameter. We are introducing new endpoints that allow you to securely manage tokens for OAuth Apps by using access_token as an input parameter. The OAuth Application API will be removed on May 5, 2021. For more information, including scheduled brownouts, see the blog post.

OAuth application owners can revoke a grant for their OAuth application and a specific user. You must use Basic Authentication when accessing this endpoint, using the OAuth application's client_id and client_secret as the username and password. You must also provide a valid token as :access_token and the grant for the token's owner will be deleted.

Deleting an OAuth application's grant will also delete all OAuth tokens associated with the application for the user. Once deleted, the application will have no access to the user's account and will no longer be listed on the Applications settings page under "Authorized OAuth Apps" on GitHub.

See also

GitHub Developer Guide

pub fn list_installations_for_authenticated_user(
&self
) -> AppsListInstallationsForAuthenticatedUserBuilder

List app installations accessible to the user access token

Lists installations of your GitHub App that the authenticated user has explicit permission (:read, :write, or :admin) to access.

You must use a user-to-server OAuth access token, created for a user who has authorized your GitHub App, to access this endpoint.

The authenticated user has explicit permission to access repositories they own, repositories where they are a collaborator, and repositories that they can access through an organization membership.

You can find the permissions for the installation under the permissions key.

See also

GitHub Developer Guide

pub fn list_accounts_for_plan_stubbed(
&self
) -> AppsListAccountsForPlanStubbedBuilder

List accounts for a plan (stubbed)

Returns repository and organization accounts associated with the specified plan, including free plans. For per-seat pricing, you see the list of accounts that have purchased the plan, including the number of seats purchased. When someone submits a plan change that won't be processed until the end of their billing cycle, you will also see the upcoming pending change.

GitHub Apps must use a JWT to access this endpoint. OAuth Apps must use basic authentication with their client ID and client secret to access this endpoint.

See also

GitHub Developer Guide

pub fn list_subscriptions_for_authenticated_user_stubbed(
&self
) -> AppsListSubscriptionsForAuthenticatedUserStubbedBuilder

List subscriptions for the authenticated user (stubbed)

Lists the active subscriptions for the authenticated user. You must use a user-to-server OAuth access token, created for a user who has authorized your GitHub App, to access this endpoint. . OAuth Apps must authenticate using an OAuth token.

See also

GitHub Developer Guide

pub fn get_subscription_plan_for_account(
&self
) -> AppsGetSubscriptionPlanForAccountBuilder

Get a subscription plan for an account

Shows whether the user or organization account actively subscribes to a plan listed by the authenticated GitHub App. When someone submits a plan change that won't be processed until the end of their billing cycle, you will also see the upcoming pending change.

GitHub Apps must use a JWT to access this endpoint. OAuth Apps must use basic authentication with their client ID and client secret to access this endpoint.

See also

GitHub Developer Guide

pub fn revoke_installation_access_token(
&self
) -> AppsRevokeInstallationAccessTokenBuilder

Revoke an installation access token

Revokes the installation token you're using to authenticate as an installation and access this endpoint.

Once an installation token is revoked, the token is invalidated and cannot be used. Other endpoints that require the revoked installation token must have a new installation token to work. You can create a new token using the "Create an installation access token for an app" endpoint.

You must use an installation access token to access this endpoint.

See also

GitHub Developer Guide

pub fn create_content_attachment(&self) -> AppsCreateContentAttachmentBuilder

Create a content attachment

Creates an attachment under a content reference URL in the body or comment of an issue or pull request. Use the id of the content reference from the content_reference event to create an attachment.

The app must create a content attachment within six hours of the content reference URL being posted. See "Using content attachments" for details about content attachments.

You must use an installation access token to access this endpoint.

This example creates a content attachment for the domain https://errors.ai/.

See also

GitHub Developer Guide

pub fn list_plans_stubbed(&self) -> AppsListPlansStubbedBuilder

List plans (stubbed)

Lists all plans that are part of your GitHub Marketplace listing.

GitHub Apps must use a JWT to access this endpoint. OAuth Apps must use basic authentication with their client ID and client secret to access this endpoint.

See also

GitHub Developer Guide

pub fn get_installation(&self) -> AppsGetInstallationBuilder

Get an installation for the authenticated app

Enables an authenticated GitHub App to find an installation's information using the installation id. The installation's account type (target_type) will be either an organization or a user account, depending which account the repository belongs to.

You must use a JWT to access this endpoint.

See also

GitHub Developer Guide

pub fn delete_installation(&self) -> AppsDeleteInstallationBuilder

Delete an installation for the authenticated app

Uninstalls a GitHub App on a user, organization, or business account. If you prefer to temporarily suspend an app's access to your account's resources, then we recommend the "Suspend an app installation" endpoint.

You must use a JWT to access this endpoint.

See also

GitHub Developer Guide

pub fn list_installations(&self) -> AppsListInstallationsBuilder

List installations for the authenticated app

You must use a JWT to access this endpoint.

The permissions the installation has are included under the permissions key.

See also

GitHub Developer Guide

pub fn list_repos_accessible_to_installation(
&self
) -> AppsListReposAccessibleToInstallationBuilder

List repositories accessible to the app installation

List repositories that an app installation can access.

You must use an installation access token to access this endpoint.

See also

GitHub Developer Guide

pub fn list_subscriptions_for_authenticated_user(
&self
) -> AppsListSubscriptionsForAuthenticatedUserBuilder

List subscriptions for the authenticated user

Lists the active subscriptions for the authenticated user. You must use a user-to-server OAuth access token, created for a user who has authorized your GitHub App, to access this endpoint. . OAuth Apps must authenticate using an OAuth token.

See also

GitHub Developer Guide

pub fn create_installation_access_token(
&self
) -> AppsCreateInstallationAccessTokenBuilder

Create an installation access token for an app

Creates an installation access token that enables a GitHub App to make authenticated API requests for the app's installation on an organization or individual account. Installation tokens expire one hour from the time you create them. Using an expired token produces a status code of 401 - Unauthorized, and requires creating a new installation token. By default the installation token has access to all repositories that the installation can access. To restrict the access to specific repositories, you can provide the repository_ids when creating the token. When you omit repository_ids, the response does not contain the repositories key.

You must use a JWT to access this endpoint.

This example grants the token "Read and write" permission to issues and "Read" permission to contents, and restricts the token's access to the repository with an id of 1296269.

See also

GitHub Developer Guide

pub fn check_token(&self) -> AppsCheckTokenBuilder

Check a token

OAuth applications can use a special API method for checking OAuth token validity without exceeding the normal rate limits for failed login attempts. Authentication works differently with this particular endpoint. You must use Basic Authentication to use this endpoint, where the username is the OAuth application client_id and the password is its client_secret. Invalid tokens will return 404 NOT FOUND.

See also

GitHub Developer Guide

pub fn delete_token(&self) -> AppsDeleteTokenBuilder

Delete an app token

OAuth application owners can revoke a single token for an OAuth application. You must use Basic Authentication when accessing this endpoint, using the OAuth application's client_id and client_secret as the username and password.

See also

GitHub Developer Guide

pub fn reset_token(&self) -> AppsResetTokenBuilder

Reset a token

OAuth applications can use this API method to reset a valid OAuth token without end-user involvement. Applications must save the "token" property in the response because changes take effect immediately. You must use Basic Authentication when accessing this endpoint, using the OAuth application's client_id and client_secret as the username and password. Invalid tokens will return 404 NOT FOUND.

See also

GitHub Developer Guide

pub fn delete_authorization(&self) -> AppsDeleteAuthorizationBuilder

Delete an app authorization

OAuth application owners can revoke a grant for their OAuth application and a specific user. You must use Basic Authentication when accessing this endpoint, using the OAuth application's client_id and client_secret as the username and password. You must also provide a valid OAuth access_token as an input parameter and the grant for the token's owner will be deleted.

Deleting an OAuth application's grant will also delete all OAuth tokens associated with the application for the user. Once deleted, the application will have no access to the user's account and will no longer be listed on the application authorizations settings screen within GitHub.

See also

GitHub Developer Guide

pub fn create_from_manifest(&self) -> AppsCreateFromManifestBuilder

Create a GitHub App from a manifest

Use this endpoint to complete the handshake necessary when implementing the GitHub App Manifest flow. When you create a GitHub App with the manifest flow, you receive a temporary code used to retrieve the GitHub App's id, pem (private key), and webhook_secret.

See also

GitHub Developer Guide

pub fn get_org_installation(&self) -> AppsGetOrgInstallationBuilder

Get an organization installation for the authenticated app

Enables an authenticated GitHub App to find the organization's installation information.

You must use a JWT to access this endpoint.

See also

GitHub Developer Guide

pub fn unsuspend_installation(&self) -> AppsUnsuspendInstallationBuilder

Unsuspend an app installation

Note: Suspending a GitHub App installation is currently in beta and subject to change. Before you can suspend a GitHub App, the app owner must enable suspending installations for the app by opting-in to the beta. For more information, see "Suspending a GitHub App installation."

Removes a GitHub App installation suspension.

To unsuspend a GitHub App, you must be an account owner or have admin permissions in the repository or organization where the app is installed and suspended.

You must use a JWT to access this endpoint.

See also

GitHub Developer Guide

pub fn suspend_installation(&self) -> AppsSuspendInstallationBuilder

Suspend an app installation

Note: Suspending a GitHub App installation is currently in beta and subject to change. Before you can suspend a GitHub App, the app owner must enable suspending installations for the app by opting-in to the beta. For more information, see "Suspending a GitHub App installation."

Suspends a GitHub App on a user, organization, or business account, which blocks the app from accessing the account's resources. When a GitHub App is suspended, the app's access to the GitHub API or webhook events is blocked for that account.

To suspend a GitHub App, you must be an account owner or have admin permissions in the repository or organization where the app is installed.

You must use a JWT to access this endpoint.

See also

GitHub Developer Guide

pub fn get_subscription_plan_for_account_stubbed(
&self
) -> AppsGetSubscriptionPlanForAccountStubbedBuilder

Get a subscription plan for an account (stubbed)

Shows whether the user or organization account actively subscribes to a plan listed by the authenticated GitHub App. When someone submits a plan change that won't be processed until the end of their billing cycle, you will also see the upcoming pending change.

GitHub Apps must use a JWT to access this endpoint. OAuth Apps must use basic authentication with their client ID and client secret to access this endpoint.

See also

GitHub Developer Guide