Struct octorest::apis::InteractionsApi

pub struct InteractionsApi<'t> { /* fields omitted */ }

This is supported on feature="gh-interactions" only.

Interactions API

Implementations

impl<'client> InteractionsApi<'client>

pub fn get_restrictions_for_repo(
&self
) -> InteractionsGetRestrictionsForRepoBuilder

Get interaction restrictions for a repository

Shows which group of GitHub users can interact with this repository and when the restriction expires. If there are no restrictions, you will see an empty response.

See also

GitHub Developer Guide

pub fn set_restrictions_for_repo(
&self
) -> InteractionsSetRestrictionsForRepoBuilder

Set interaction restrictions for a repository

Temporarily restricts interactions to certain GitHub users within the given repository. You must have owner or admin access to set restrictions.

See also

GitHub Developer Guide

pub fn remove_restrictions_for_repo(
&self
) -> InteractionsRemoveRestrictionsForRepoBuilder

Remove interaction restrictions for a repository

Removes all interaction restrictions from the given repository. You must have owner or admin access to remove restrictions.

See also

GitHub Developer Guide

pub fn remove_restrictions_for_org(
&self
) -> InteractionsRemoveRestrictionsForOrgBuilder

Remove interaction restrictions for an organization

Removes all interaction restrictions from public repositories in the given organization. You must be an organization owner to remove restrictions.

See also

GitHub Developer Guide

pub fn get_restrictions_for_org(
&self
) -> InteractionsGetRestrictionsForOrgBuilder

Get interaction restrictions for an organization

Shows which group of GitHub users can interact with this organization and when the restriction expires. If there are no restrictions, you will see an empty response.

See also

GitHub Developer Guide

pub fn set_restrictions_for_org(
&self
) -> InteractionsSetRestrictionsForOrgBuilder

Set interaction restrictions for an organization

Temporarily restricts interactions to certain GitHub users in any public repository in the given organization. You must be an organization owner to set these restrictions.

See also

GitHub Developer Guide