Struct octorest::apis::OrgsApi

pub struct OrgsApi<'t> { /* fields omitted */ }

This is supported on feature="gh-orgs" only.

Orgs API

Implementations

impl<'client> OrgsApi<'client>

pub fn list(&self) -> OrgsListBuilder

List organizations

Lists all organizations, in the order that they were created on GitHub.

Note: Pagination is powered exclusively by the since parameter. Use the Link header to get the URL for the next page of organizations.

See also

GitHub Developer Guide

pub fn list_members(&self) -> OrgsListMembersBuilder

List organization members

List all users who are members of an organization. If the authenticated user is also a member of this organization then both concealed and public members will be returned.

See also

GitHub Developer Guide

pub fn list_app_installations(&self) -> OrgsListAppInstallationsBuilder

List app installations for an organization

Lists all GitHub Apps in an organization. The installation count includes all GitHub Apps installed on repositories in the organization. You must be an organization owner with admin:read scope to use this endpoint.

See also

GitHub Developer Guide

pub fn remove_saml_sso_authorization(
&self
) -> OrgsRemoveSamlSsoAuthorizationBuilder

Remove a SAML SSO authorization for an organization

Listing and deleting credential authorizations is available to organizations with GitHub Enterprise Cloud. For more information, see GitHub's products in the GitHub Help documentation.

An authenticated organization owner with the admin:org scope can remove a credential authorization for an organization that uses SAML SSO. Once you remove someone's credential authorization, they will need to create a new personal access token or SSH key and authorize it for the organization they want to access.

See also

GitHub Developer Guide

pub fn list_for_user(&self) -> OrgsListForUserBuilder

List organizations for a user

List public organization memberships for the specified user.

This method only lists public memberships, regardless of authentication. If you need to fetch all of the organization memberships (public and private) for the authenticated user, use the List organizations for the authenticated user API instead.

See also

GitHub Developer Guide

pub fn get_membership_for_user(&self) -> OrgsGetMembershipForUserBuilder

Get organization membership for a user

In order to get a user's membership with an organization, the authenticated user must be an organization member.

See also

GitHub Developer Guide

pub fn remove_membership_for_user(&self) -> OrgsRemoveMembershipForUserBuilder

Remove organization membership for a user

In order to remove a user's membership with an organization, the authenticated user must be an organization owner.

If the specified user is an active member of the organization, this will remove them from the organization. If the specified user has been invited to the organization, this will cancel their invitation. The specified user will receive an email notification in both cases.

See also

GitHub Developer Guide

pub fn set_membership_for_user(&self) -> OrgsSetMembershipForUserBuilder

Set organization membership for a user

Only authenticated organization owners can add a member to the organization or update the member's role.

If the authenticated user is adding a member to the organization, the invited user will receive an email inviting them to the organization. The user's membership status will be pending until they accept the invitation.

Authenticated users can update a user's membership by passing the role parameter. If the authenticated user changes a member's role to admin, the affected user will receive an email notifying them that they've been made an organization owner. If the authenticated user changes an owner's role to member, no email will be sent.

Rate limits

To prevent abuse, the authenticated user is limited to 50 organization invitations per 24 hour period. If the organization is more than one month old or on a paid plan, the limit is 500 invitations per 24 hour period.

See also

GitHub Developer Guide

pub fn list_saml_sso_authorizations(
&self
) -> OrgsListSamlSsoAuthorizationsBuilder

List SAML SSO authorizations for an organization

Listing and deleting credential authorizations is available to organizations with GitHub Enterprise Cloud. For more information, see GitHub's products in the GitHub Help documentation.

An authenticated organization owner with the read:org scope can list all credential authorizations for an organization that uses SAML single sign-on (SSO). The credentials are either personal access tokens or SSH keys that organization members have authorized for the organization. For more information, see About authentication with SAML single sign-on.

See also

GitHub Developer Guide

pub fn list_outside_collaborators(&self) -> OrgsListOutsideCollaboratorsBuilder

List outside collaborators for an organization

List all users who are outside collaborators of an organization.

See also

GitHub Developer Guide

pub fn get_membership_for_authenticated_user(
&self
) -> OrgsGetMembershipForAuthenticatedUserBuilder

Get an organization membership for the authenticated user

See also

GitHub Developer Guide

pub fn update_membership_for_authenticated_user(
&self
) -> OrgsUpdateMembershipForAuthenticatedUserBuilder

Update an organization membership for the authenticated user

See also

GitHub Developer Guide

pub fn list_for_authenticated_user(&self) -> OrgsListForAuthenticatedUserBuilder

List organizations for the authenticated user

List organizations for the authenticated user.

OAuth scope requirements

This only lists organizations that your authorization allows you to operate on in some way (e.g., you can list teams with read:org scope, you can publicize your organization membership with user scope, etc.). Therefore, this API requires at least user or read:org scope. OAuth requests with insufficient scope receive a 403 Forbidden response.

See also

GitHub Developer Guide

pub fn update(&self) -> OrgsUpdateBuilder

Update an organization

Parameter Deprecation Notice: GitHub will replace and discontinue members_allowed_repository_creation_type in favor of more granular permissions. The new input parameters are members_can_create_public_repositories, members_can_create_private_repositories for all organizations and members_can_create_internal_repositories for organizations associated with an enterprise account using GitHub Enterprise Cloud or GitHub Enterprise Server 2.20+. For more information, see the blog post.

Enables an authenticated organization owner with the admin:org scope to update the organization's profile and member privileges.

See also

GitHub Developer Guide

pub fn get(&self) -> OrgsGetBuilder

Get an organization

To see many of the organization response values, you need to be an authenticated organization owner with the admin:org scope. When the value of two_factor_requirement_enabled is true, the organization requires all members, billing managers, and outside collaborators to enable two-factor authentication.

GitHub Apps with the Organization plan permission can use this endpoint to retrieve information about an organization's GitHub plan. See "Authenticating with GitHub Apps" for details. For an example response, see "Response with GitHub plan information."

See also

GitHub Developer Guide

pub fn convert_member_to_outside_collaborator(
&self
) -> OrgsConvertMemberToOutsideCollaboratorBuilder

Convert an organization member to outside collaborator

When an organization member is converted to an outside collaborator, they'll only have access to the repositories that their current team membership allows. The user will no longer be a member of the organization. For more information, see "Converting an organization member to an outside collaborator".

See also

GitHub Developer Guide

pub fn remove_outside_collaborator(
&self
) -> OrgsRemoveOutsideCollaboratorBuilder

Remove outside collaborator from an organization

Removing a user from this list will remove them from all the organization's repositories.

See also

GitHub Developer Guide

pub fn block_user(&self) -> OrgsBlockUserBuilder

Block a user from an organization

See also

GitHub Developer Guide

pub fn check_blocked_user(&self) -> OrgsCheckBlockedUserBuilder

Check if a user is blocked by an organization

If the user is blocked:

If the user is not blocked:

See also

GitHub Developer Guide

pub fn unblock_user(&self) -> OrgsUnblockUserBuilder

Unblock a user from an organization

See also

GitHub Developer Guide

pub fn get_webhook(&self) -> OrgsGetWebhookBuilder

Get an organization webhook

See also

GitHub Developer Guide

pub fn update_webhook(&self) -> OrgsUpdateWebhookBuilder

Update an organization webhook

See also

GitHub Developer Guide

pub fn delete_webhook(&self) -> OrgsDeleteWebhookBuilder

Delete an organization webhook

See also

GitHub Developer Guide

pub fn ping_webhook(&self) -> OrgsPingWebhookBuilder

Ping an organization webhook

This will trigger a ping event to be sent to the hook.

See also

GitHub Developer Guide

pub fn list_pending_invitations(&self) -> OrgsListPendingInvitationsBuilder

List pending organization invitations

The return hash contains a role field which refers to the Organization Invitation role and will be one of the following values: direct_member, admin, billing_manager, hiring_manager, or reinstate. If the invitee is not a GitHub member, the login field in the return hash will be null.

See also

GitHub Developer Guide

pub fn create_invitation(&self) -> OrgsCreateInvitationBuilder

Create an organization invitation

Invite people to an organization by using their GitHub user ID or their email address. In order to create invitations in an organization, the authenticated user must be an organization owner.

This endpoint triggers notifications. Creating content too quickly using this endpoint may result in abuse rate limiting. See "Abuse rate limits" and "Dealing with abuse rate limits" for details.

See also

GitHub Developer Guide

pub fn check_membership_for_user(&self) -> OrgsCheckMembershipForUserBuilder

Check organization membership for a user

Check if a user is, publicly or privately, a member of the organization.

See also

GitHub Developer Guide

pub fn remove_member(&self) -> OrgsRemoveMemberBuilder

Remove an organization member

Removing a user from this list will remove them from all teams and they will no longer have any access to the organization's repositories.

See also

GitHub Developer Guide

pub fn remove_public_membership_for_authenticated_user(
&self
) -> OrgsRemovePublicMembershipForAuthenticatedUserBuilder

Remove public organization membership for the authenticated user

See also

GitHub Developer Guide

pub fn set_public_membership_for_authenticated_user(
&self
) -> OrgsSetPublicMembershipForAuthenticatedUserBuilder

Set public organization membership for the authenticated user

The user can publicize their own membership. (A user cannot publicize the membership for another user.)

Note that you'll need to set Content-Length to zero when calling out to this endpoint. For more information, see "HTTP verbs."

See also

GitHub Developer Guide

pub fn check_public_membership_for_user(
&self
) -> OrgsCheckPublicMembershipForUserBuilder

Check public organization membership for a user

See also

GitHub Developer Guide

pub fn list_public_members(&self) -> OrgsListPublicMembersBuilder

List public organization members

Members of an organization can choose to have their membership publicized or not.

See also

GitHub Developer Guide

pub fn list_webhooks(&self) -> OrgsListWebhooksBuilder

List organization webhooks

See also

GitHub Developer Guide

pub fn create_webhook(&self) -> OrgsCreateWebhookBuilder

Create an organization webhook

Here's how you can create a hook that posts payloads in JSON format:

See also

GitHub Developer Guide

pub fn list_memberships_for_authenticated_user(
&self
) -> OrgsListMembershipsForAuthenticatedUserBuilder

List organization memberships for the authenticated user

See also

GitHub Developer Guide

pub fn list_invitation_teams(&self) -> OrgsListInvitationTeamsBuilder

List organization invitation teams

List all teams associated with an invitation. In order to see invitations in an organization, the authenticated user must be an organization owner.

See also

GitHub Developer Guide

pub fn list_blocked_users(&self) -> OrgsListBlockedUsersBuilder

List users blocked by an organization

List the users blocked by an organization.

See also

GitHub Developer Guide