Struct octorest::apis::ProjectsApi

pub struct ProjectsApi<'t> { /* fields omitted */ }

This is supported on feature="gh-projects" only.

Projects API

Implementations

impl<'client> ProjectsApi<'client>

pub fn move_card(&self) -> ProjectsMoveCardBuilder

Move a project card

See also

GitHub Developer Guide

pub fn add_collaborator(&self) -> ProjectsAddCollaboratorBuilder

Add project collaborator

Adds a collaborator to an organization project and sets their permission level. You must be an organization owner or a project admin to add a collaborator.

See also

GitHub Developer Guide

pub fn remove_collaborator(&self) -> ProjectsRemoveCollaboratorBuilder

Remove project collaborator

Removes a collaborator from an organization project. You must be an organization owner or a project admin to remove a collaborator.

See also

GitHub Developer Guide

pub fn create_for_authenticated_user(
&self
) -> ProjectsCreateForAuthenticatedUserBuilder

Create a user project

See also

GitHub Developer Guide

pub fn move_column(&self) -> ProjectsMoveColumnBuilder

Move a project column

See also

GitHub Developer Guide

pub fn list_for_user(&self) -> ProjectsListForUserBuilder

List user projects

See also

GitHub Developer Guide

pub fn list_for_repo(&self) -> ProjectsListForRepoBuilder

List repository projects

Lists the projects in a repository. Returns a 404 Not Found status if projects are disabled in the repository. If you do not have sufficient privileges to perform this action, a 401 Unauthorized or 410 Gone status is returned.

See also

GitHub Developer Guide

pub fn create_for_repo(&self) -> ProjectsCreateForRepoBuilder

Create a repository project

Creates a repository project board. Returns a 404 Not Found status if projects are disabled in the repository. If you do not have sufficient privileges to perform this action, a 401 Unauthorized or 410 Gone status is returned.

See also

GitHub Developer Guide

pub fn create_card(&self) -> ProjectsCreateCardBuilder

Create a project card

Note: GitHub's REST API v3 considers every pull request an issue, but not every issue is a pull request. For this reason, "Issues" endpoints may return both issues and pull requests in the response. You can identify pull requests by the pull_request key.

Be aware that the id of a pull request returned from "Issues" endpoints will be an issue id. To find out the pull request id, use the "List pull requests" endpoint.

See also

GitHub Developer Guide

pub fn list_cards(&self) -> ProjectsListCardsBuilder

List project cards

See also

GitHub Developer Guide

pub fn update_column(&self) -> ProjectsUpdateColumnBuilder

Update a project column

See also

GitHub Developer Guide

pub fn get_column(&self) -> ProjectsGetColumnBuilder

Get a project column

See also

GitHub Developer Guide

pub fn delete_column(&self) -> ProjectsDeleteColumnBuilder

Delete a project column

See also

GitHub Developer Guide

pub fn delete_card(&self) -> ProjectsDeleteCardBuilder

Delete a project card

See also

GitHub Developer Guide

pub fn get_card(&self) -> ProjectsGetCardBuilder

Get a project card

See also

GitHub Developer Guide

pub fn update_card(&self) -> ProjectsUpdateCardBuilder

Update a project card

See also

GitHub Developer Guide

pub fn get_permission_for_user(&self) -> ProjectsGetPermissionForUserBuilder

Get project permission for a user

Returns the collaborator's permission level for an organization project. Possible values for the permission key: admin, write, read, none. You must be an organization owner or a project admin to review a user's permission level.

See also

GitHub Developer Guide

pub fn list_collaborators(&self) -> ProjectsListCollaboratorsBuilder

List project collaborators

Lists the collaborators for an organization project. For a project, the list of collaborators includes outside collaborators, organization members that are direct collaborators, organization members with access through team memberships, organization members with access through default organization permissions, and organization owners. You must be an organization owner or a project admin to list collaborators.

See also

GitHub Developer Guide

pub fn list_columns(&self) -> ProjectsListColumnsBuilder

List project columns

See also

GitHub Developer Guide

pub fn create_column(&self) -> ProjectsCreateColumnBuilder

Create a project column

See also

GitHub Developer Guide

pub fn list_for_org(&self) -> ProjectsListForOrgBuilder

List organization projects

Lists the projects in an organization. Returns a 404 Not Found status if projects are disabled in the organization. If you do not have sufficient privileges to perform this action, a 401 Unauthorized or 410 Gone status is returned.

s

See also

GitHub Developer Guide

pub fn create_for_org(&self) -> ProjectsCreateForOrgBuilder

Create an organization project

Creates an organization project board. Returns a 404 Not Found status if projects are disabled in the organization. If you do not have sufficient privileges to perform this action, a 401 Unauthorized or 410 Gone status is returned.

See also

GitHub Developer Guide

pub fn update(&self) -> ProjectsUpdateBuilder

Update a project

Updates a project board's information. Returns a 404 Not Found status if projects are disabled. If you do not have sufficient privileges to perform this action, a 401 Unauthorized or 410 Gone status is returned.

See also

GitHub Developer Guide

pub fn get(&self) -> ProjectsGetBuilder

Get a project

Gets a project by its id. Returns a 404 Not Found status if projects are disabled. If you do not have sufficient privileges to perform this action, a 401 Unauthorized or 410 Gone status is returned.

See also

GitHub Developer Guide

pub fn delete(&self) -> ProjectsDeleteBuilder

Delete a project

Deletes a project board. Returns a 404 Not Found status if projects are disabled.

See also

GitHub Developer Guide