Struct octorest::apis::PullsApi

pub struct PullsApi<'t> { /* fields omitted */ }

This is supported on feature="gh-pulls" only.

Pulls API

Implementations

impl<'client> PullsApi<'client>

pub fn list_comments_for_review(&self) -> PullsListCommentsForReviewBuilder

List comments for a pull request review

List comments for a specific pull request review.

See also

GitHub Developer Guide

pub fn update(&self) -> PullsUpdateBuilder

Update a pull request

Draft pull requests are available in public repositories with GitHub Free and GitHub Free for organizations, GitHub Pro, and legacy per-repository billing plans, and in public and private repositories with GitHub Team and GitHub Enterprise Cloud. For more information, see GitHub's products in the GitHub Help documentation.

To open or update a pull request in a public repository, you must have write access to the head or the source branch. For organization-owned repositories, you must be a member of the organization that owns the repository to open or update a pull request.

See also

GitHub Developer Guide

pub fn get(&self) -> PullsGetBuilder

Get a pull request

Draft pull requests are available in public repositories with GitHub Free and GitHub Free for organizations, GitHub Pro, and legacy per-repository billing plans, and in public and private repositories with GitHub Team and GitHub Enterprise Cloud. For more information, see GitHub's products in the GitHub Help documentation.

Lists details of a pull request by providing its number.

When you get, create, or edit a pull request, GitHub creates a merge commit to test whether the pull request can be automatically merged into the base branch. This test commit is not added to the base branch or the head branch. You can review the status of the test commit using the mergeable key. For more information, see "Checking mergeability of pull requests".

The value of the mergeable attribute can be true, false, or null. If the value is null, then GitHub has started a background job to compute the mergeability. After giving the job time to complete, resubmit the request. When the job finishes, you will see a non-null value for the mergeable attribute in the response. If mergeable is true, then merge_commit_sha will be the SHA of the test merge commit.

The value of the merge_commit_sha attribute changes depending on the state of the pull request. Before merging a pull request, the merge_commit_sha attribute holds the SHA of the test merge commit. After merging a pull request, the merge_commit_sha attribute changes depending on how you merged the pull request:

If merged as a merge commit, merge_commit_sha represents the SHA of the merge commit.

If merged via a squash, merge_commit_sha represents the SHA of the squashed commit on the base branch.

If rebased, merge_commit_sha represents the commit that the base branch was updated to.

Pass the appropriate media type to fetch diff and patch formats.

See also

GitHub Developer Guide

pub fn create_review(&self) -> PullsCreateReviewBuilder

Create a review for a pull request

This endpoint triggers notifications. Creating content too quickly using this endpoint may result in abuse rate limiting. See "Abuse rate limits" and "Dealing with abuse rate limits" for details.

Pull request reviews created in the PENDING state do not include the submitted_at property in the response.

Note: To comment on a specific line in a file, you need to first determine the position of that line in the diff. The GitHub REST API v3 offers the application/vnd.github.v3.diff media type. To see a pull request diff, add this media type to the Accept header of a call to the single pull request endpoint.

The position value equals the number of lines down from the first "@@" hunk header in the file you want to add a comment. The line just below the "@@" line is position 1, the next line is position 2, and so on. The position in the diff continues to increase through lines of whitespace and additional hunks until the beginning of a new file.

See also

GitHub Developer Guide

pub fn list_reviews(&self) -> PullsListReviewsBuilder

List reviews for a pull request

The list of reviews returns in chronological order.

See also

GitHub Developer Guide

pub fn remove_requested_reviewers(&self) -> PullsRemoveRequestedReviewersBuilder

Remove requested reviewers from a pull request

See also

GitHub Developer Guide

pub fn list_requested_reviewers(&self) -> PullsListRequestedReviewersBuilder

List requested reviewers for a pull request

See also

GitHub Developer Guide

pub fn request_reviewers(&self) -> PullsRequestReviewersBuilder

Request reviewers for a pull request

This endpoint triggers notifications. Creating content too quickly using this endpoint may result in abuse rate limiting. See "Abuse rate limits" and "Dealing with abuse rate limits" for details.

See also

GitHub Developer Guide

pub fn list_files(&self) -> PullsListFilesBuilder

List pull requests files

Note: Responses include a maximum of 3000 files. The paginated response returns 30 files per page by default.

See also

GitHub Developer Guide

pub fn get_review_comment(&self) -> PullsGetReviewCommentBuilder

Get a review comment for a pull request

Note: Multi-line comments on pull requests are currently in public beta and subject to change.

Provides details for a review comment.

Multi-line comment summary

Note: New parameters and response fields are available for developers to preview. During the preview period, these response fields may change without advance notice. Please see the blog post for full details.

Use the comfort-fade preview header and the line parameter to show multi-line comment-supported fields in the response.

If you use the comfort-fade preview header, your response will show:

For multi-line comments, values for start_line, original_start_line, start_side, line, original_line, and side.

For single-line comments, values for line, original_line, and side and a null value for start_line, original_start_line, and start_side.

If you don't use the comfort-fade preview header, multi-line and single-line comments will appear the same way in the response with a single position attribute. Your response will show:

For multi-line comments, the last line of the comment range for the position attribute.

For single-line comments, the diff-positioned way of referencing comments for the position attribute. For more information, see position in the input parameters table.

The reactions key will have the following payload where url can be used to construct the API location for listing and creating reactions.

See also

GitHub Developer Guide

pub fn update_review_comment(&self) -> PullsUpdateReviewCommentBuilder

Update a review comment for a pull request

Note: Multi-line comments on pull requests are currently in public beta and subject to change.

Enables you to edit a review comment.

Multi-line comment summary

Note: New parameters and response fields are available for developers to preview. During the preview period, these response fields may change without advance notice. Please see the blog post for full details.

Use the comfort-fade preview header and the line parameter to show multi-line comment-supported fields in the response.

If you use the comfort-fade preview header, your response will show:

For multi-line comments, values for start_line, original_start_line, start_side, line, original_line, and side.

For single-line comments, values for line, original_line, and side and a null value for start_line, original_start_line, and start_side.

If you don't use the comfort-fade preview header, multi-line and single-line comments will appear the same way in the response with a single position attribute. Your response will show:

For multi-line comments, the last line of the comment range for the position attribute.

For single-line comments, the diff-positioned way of referencing comments for the position attribute. For more information, see position in the input parameters table.

See also

GitHub Developer Guide

pub fn delete_review_comment(&self) -> PullsDeleteReviewCommentBuilder

Delete a review comment for a pull request

Deletes a review comment.

See also

GitHub Developer Guide

pub fn submit_review(&self) -> PullsSubmitReviewBuilder

Submit a review for a pull request

See also

GitHub Developer Guide

pub fn dismiss_review(&self) -> PullsDismissReviewBuilder

Dismiss a review for a pull request

Note: To dismiss a pull request review on a protected branch, you must be a repository administrator or be included in the list of people or teams who can dismiss pull request reviews.

See also

GitHub Developer Guide

pub fn update_branch(&self) -> PullsUpdateBranchBuilder

Update a pull request branch

Updates the pull request branch with the latest upstream changes by merging HEAD from the base branch into the pull request branch.

See also

GitHub Developer Guide

pub fn list_commits(&self) -> PullsListCommitsBuilder

List commits on a pull request

Lists a maximum of 250 commits for a pull request. To receive a complete commit list for pull requests with more than 250 commits, use the List commits endpoint.

See also

GitHub Developer Guide

pub fn list(&self) -> PullsListBuilder

List pull requests

Draft pull requests are available in public repositories with GitHub Free and GitHub Free for organizations, GitHub Pro, and legacy per-repository billing plans, and in public and private repositories with GitHub Team and GitHub Enterprise Cloud. For more information, see GitHub's products in the GitHub Help documentation.

See also

GitHub Developer Guide

pub fn create(&self) -> PullsCreateBuilder

Create a pull request

Draft pull requests are available in public repositories with GitHub Free and GitHub Free for organizations, GitHub Pro, and legacy per-repository billing plans, and in public and private repositories with GitHub Team and GitHub Enterprise Cloud. For more information, see GitHub's products in the GitHub Help documentation.

To open or update a pull request in a public repository, you must have write access to the head or the source branch. For organization-owned repositories, you must be a member of the organization that owns the repository to open or update a pull request.

You can create a new pull request.

This endpoint triggers notifications. Creating content too quickly using this endpoint may result in abuse rate limiting. See "Abuse rate limits" and "Dealing with abuse rate limits" for details.

See also

GitHub Developer Guide

pub fn create_reply_for_review_comment(
&self
) -> PullsCreateReplyForReviewCommentBuilder

Create a reply for a review comment

Creates a reply to a review comment for a pull request. For the comment_id, provide the ID of the review comment you are replying to. This must be the ID of a top-level review comment, not a reply to that comment. Replies to replies are not supported.

This endpoint triggers notifications. Creating content too quickly using this endpoint may result in abuse rate limiting. See "Abuse rate limits" and "Dealing with abuse rate limits" for details.

See also

GitHub Developer Guide

pub fn delete_pending_review(&self) -> PullsDeletePendingReviewBuilder

Delete a pending review for a pull request

See also

GitHub Developer Guide

pub fn update_review(&self) -> PullsUpdateReviewBuilder

Update a review for a pull request

Update the review summary comment with new text.

See also

GitHub Developer Guide

pub fn get_review(&self) -> PullsGetReviewBuilder

Get a review for a pull request

See also

GitHub Developer Guide

pub fn create_review_comment(&self) -> PullsCreateReviewCommentBuilder

Create a review comment for a pull request

Note: Multi-line comments on pull requests are currently in public beta and subject to change.

Creates a review comment in the pull request diff. To add a regular comment to a pull request timeline, see "Create an issue comment." We recommend creating a review comment using line, side, and optionally start_line and start_side if your comment applies to more than one line in the pull request diff.

You can still create a review comment using the position parameter. When you use position, the line, side, start_line, and start_side parameters are not required. For more information, see Multi-line comment summary.

Note: The position value equals the number of lines down from the first "@@" hunk header in the file you want to add a comment. The line just below the "@@" line is position 1, the next line is position 2, and so on. The position in the diff continues to increase through lines of whitespace and additional hunks until the beginning of a new file.

This endpoint triggers notifications. Creating content too quickly using this endpoint may result in abuse rate limiting. See "Abuse rate limits" and "Dealing with abuse rate limits" for details.

Multi-line comment summary

Note: New parameters and response fields are available for developers to preview. During the preview period, these response fields may change without advance notice. Please see the blog post for full details.

Use the comfort-fade preview header and the line parameter to show multi-line comment-supported fields in the response.

If you use the comfort-fade preview header, your response will show:

For multi-line comments, values for start_line, original_start_line, start_side, line, original_line, and side.

For single-line comments, values for line, original_line, and side and a null value for start_line, original_start_line, and start_side.

If you don't use the comfort-fade preview header, multi-line and single-line comments will appear the same way in the response with a single position attribute. Your response will show:

For multi-line comments, the last line of the comment range for the position attribute.

For single-line comments, the diff-positioned way of referencing comments for the position attribute. For more information, see position in the input parameters table.

See also

GitHub Developer Guide

pub fn list_review_comments(&self) -> PullsListReviewCommentsBuilder

List review comments on a pull request

Note: Multi-line comments on pull requests are currently in public beta and subject to change.

Lists all review comments for a pull request. By default, review comments are in ascending order by ID.

Multi-line comment summary

Note: New parameters and response fields are available for developers to preview. During the preview period, these response fields may change without advance notice. Please see the blog post for full details.

Use the comfort-fade preview header and the line parameter to show multi-line comment-supported fields in the response.

If you use the comfort-fade preview header, your response will show:

For multi-line comments, values for start_line, original_start_line, start_side, line, original_line, and side.

For single-line comments, values for line, original_line, and side and a null value for start_line, original_start_line, and start_side.

If you don't use the comfort-fade preview header, multi-line and single-line comments will appear the same way in the response with a single position attribute. Your response will show:

For multi-line comments, the last line of the comment range for the position attribute.

For single-line comments, the diff-positioned way of referencing comments for the position attribute. For more information, see position in the input parameters table.

The reactions key will have the following payload where url can be used to construct the API location for listing and creating reactions.

See also

GitHub Developer Guide

pub fn list_review_comments_for_repo(
&self
) -> PullsListReviewCommentsForRepoBuilder

List review comments in a repository

Note: Multi-line comments on pull requests are currently in public beta and subject to change.

Lists review comments for all pull requests in a repository. By default, review comments are in ascending order by ID.

Multi-line comment summary

Note: New parameters and response fields are available for developers to preview. During the preview period, these response fields may change without advance notice. Please see the blog post for full details.

Use the comfort-fade preview header and the line parameter to show multi-line comment-supported fields in the response.

If you use the comfort-fade preview header, your response will show:

For multi-line comments, values for start_line, original_start_line, start_side, line, original_line, and side.

For single-line comments, values for line, original_line, and side and a null value for start_line, original_start_line, and start_side.

If you don't use the comfort-fade preview header, multi-line and single-line comments will appear the same way in the response with a single position attribute. Your response will show:

For multi-line comments, the last line of the comment range for the position attribute.

For single-line comments, the diff-positioned way of referencing comments for the position attribute. For more information, see position in the input parameters table.

The reactions key will have the following payload where url can be used to construct the API location for listing and creating reactions.

See also

GitHub Developer Guide

pub fn merge(&self) -> PullsMergeBuilder

Merge a pull request

This endpoint triggers notifications. Creating content too quickly using this endpoint may result in abuse rate limiting. See "Abuse rate limits" and "Dealing with abuse rate limits" for details.

See also

GitHub Developer Guide

pub fn check_if_merged(&self) -> PullsCheckIfMergedBuilder

Check if a pull request has been merged

See also

GitHub Developer Guide