Struct octorest::apis::SearchApi

pub struct SearchApi<'t> { /* fields omitted */ }

This is supported on feature="gh-search" only.

Search API

Implementations

impl<'client> SearchApi<'client>

pub fn code(&self) -> SearchCodeBuilder

Search code

Find file contents via various criteria. This method returns up to 100 results per page.

When searching for code, you can get text match metadata for the file content and file path fields when you pass the text-match media type. For more details about how to receive highlighted search results, see Text match metadata.

Note: You must authenticate to search for code across all public repositories.

Considerations for code search

Due to the complexity of searching code, there are a few restrictions on how searches are performed:

Only the default branch is considered. In most cases, this will be the master branch.

Only files smaller than 384 KB are searchable.

You must always include at least one search term when searching source code. For example, searching for language:go is not valid, while amazing language:go is.

Suppose you want to find the definition of the addClass function inside jQuery. Your query would look something like this:

Here, we're searching for the keyword addClass within a file's contents. We're making sure that we're only looking in files where the language is JavaScript. And we're scoping the search to the repo:jquery/jquery repository.

See also

GitHub Developer Guide

pub fn labels(&self) -> SearchLabelsBuilder

Search labels

Find labels in a repository with names or descriptions that match search keywords. Returns up to 100 results per page.

When searching for labels, you can get text match metadata for the label name and description fields when you pass the text-match media type. For more details about how to receive highlighted search results, see Text match metadata.

Suppose you want to find labels in the linguist repository that match bug, defect, or enhancement. Your query might look like this:

The labels that best match for the query appear first in the search results.

See also

GitHub Developer Guide

pub fn users(&self) -> SearchUsersBuilder

Search users

Find users via various criteria. This method returns up to 100 results per page.

When searching for users, you can get text match metadata for the issue login, email, and name fields when you pass the text-match media type. For more details about highlighting search results, see Text match metadata. For more details about how to receive highlighted search results, see Text match metadata.

Imagine you're looking for a list of popular users. You might try out this query:

Here, we're looking at users with the name Tom. We're only interested in those with more than 42 repositories, and only if they have over 1,000 followers.

See also

GitHub Developer Guide

pub fn topics(&self) -> SearchTopicsBuilder

Search topics

Find topics via various criteria. Results are sorted by best match. This method returns up to 100 results per page.

When searching for topics, you can get text match metadata for the topic's short_description, description, name, or display_name field when you pass the text-match media type. For more details about how to receive highlighted search results, see Text match metadata.

See "Searching topics" for a detailed list of qualifiers.

Suppose you want to search for topics related to Ruby that are featured on https://github.com/topics. Your query might look like this:

In this request, we're searching for topics with the keyword ruby, and we're limiting the results to find only topics that are featured. The topics that are the best match for the query appear first in the search results.

Note: A search for featured Ruby topics only has 6 total results, so a Link header indicating pagination is not included in the response.

See also

GitHub Developer Guide

pub fn repos(&self) -> SearchReposBuilder

Search repositories

Find repositories via various criteria. This method returns up to 100 results per page.

When searching for repositories, you can get text match metadata for the name and description fields when you pass the text-match media type. For more details about how to receive highlighted search results, see Text match metadata.

Suppose you want to search for popular Tetris repositories written in Assembly. Your query might look like this.

You can search for multiple topics by adding more topic: instances, and including the mercy-preview header. For example:

In this request, we're searching for repositories with the word tetris in the name, the description, or the README. We're limiting the results to only find repositories where the primary language is Assembly. We're sorting by stars in descending order, so that the most popular repositories appear first in the search results.

See also

GitHub Developer Guide

pub fn commits(&self) -> SearchCommitsBuilder

Search commits

Find commits via various criteria. This method returns up to 100 results per page.

When searching for commits, you can get text match metadata for the message field when you provide the text-match media type. For more details about how to receive highlighted search results, see Text match metadata.

Considerations for commit search

Only the default branch is considered. In most cases, this will be the master branch.

Suppose you want to find commits related to CSS in the octocat/Spoon-Knife repository. Your query would look something like this:

See also

GitHub Developer Guide

pub fn issues_and_pull_requests(&self) -> SearchIssuesAndPullRequestsBuilder

Search issues and pull requests

Find issues by state and keyword. This method returns up to 100 results per page.

When searching for issues, you can get text match metadata for the issue title, issue body, and issue comment body fields when you pass the text-match media type. For more details about how to receive highlighted search results, see Text match metadata.

Let's say you want to find the oldest unresolved Python bugs on Windows. Your query might look something like this.

In this query, we're searching for the keyword windows, within any open issue that's labeled as bug. The search runs across repositories whose primary language is Python. We’re sorting by creation date in ascending order, so that the oldest issues appear first in the search results.

See also

GitHub Developer Guide