Struct octorest::apis::TeamsApi

pub struct TeamsApi<'t> { /* fields omitted */ }

This is supported on feature="gh-teams" only.

Teams API

Implementations

impl<'client> TeamsApi<'client>

pub fn list_child_in_org(&self) -> TeamsListChildInOrgBuilder

List child teams

Lists the child teams of the team requested by :team_slug.

Note: You can also specify a team by org_id and team_id using the route GET /organizations/:org_id/team/:team_id/teams.

See also

GitHub Developer Guide

pub fn update_discussion_comment_legacy(
&self
) -> TeamsUpdateDiscussionCommentLegacyBuilder

Update a discussion comment (Legacy)

Deprecation Notice: This endpoint route is deprecated and will be removed from the Teams API. We recommend migrating your existing code to use the new Update a discussion comment endpoint.

Edits the body text of a discussion comment. OAuth access tokens require the write:discussion scope.

See also

GitHub Developer Guide

pub fn get_discussion_comment_legacy(
&self
) -> TeamsGetDiscussionCommentLegacyBuilder

Get a discussion comment (Legacy)

Deprecation Notice: This endpoint route is deprecated and will be removed from the Teams API. We recommend migrating your existing code to use the new Get a discussion comment endpoint.

Get a specific comment on a team discussion. OAuth access tokens require the read:discussion scope.

See also

GitHub Developer Guide

pub fn delete_discussion_comment_legacy(
&self
) -> TeamsDeleteDiscussionCommentLegacyBuilder

Delete a discussion comment (Legacy)

Deprecation Notice: This endpoint route is deprecated and will be removed from the Teams API. We recommend migrating your existing code to use the new Delete a discussion comment endpoint.

Deletes a comment on a team discussion. OAuth access tokens require the write:discussion scope.

See also

GitHub Developer Guide

pub fn check_permissions_for_project_in_org(
&self
) -> TeamsCheckPermissionsForProjectInOrgBuilder

Check team permissions for a project

Checks whether a team has read, write, or admin permissions for an organization project. The response includes projects inherited from a parent team.

Note: You can also specify a team by org_id and team_id using the route GET /organizations/:org_id/team/:team_id/projects/:project_id.

See also

GitHub Developer Guide

pub fn remove_project_in_org(&self) -> TeamsRemoveProjectInOrgBuilder

Remove a project from a team

Removes an organization project from a team. An organization owner or a team maintainer can remove any project from the team. To remove a project from a team as an organization member, the authenticated user must have read access to both the team and project, or admin access to the team or project. This endpoint removes the project from the team, but does not delete the project.

Note: You can also specify a team by org_id and team_id using the route DELETE /organizations/:org_id/team/:team_id/projects/:project_id.

See also

GitHub Developer Guide

pub fn add_or_update_project_permissions_in_org(
&self
) -> TeamsAddOrUpdateProjectPermissionsInOrgBuilder

Add or update team project permissions

Adds an organization project to a team. To add a project to a team or update the team's permission on a project, the authenticated user must have admin permissions for the project. The project and team must be part of the same organization.

Note: You can also specify a team by org_id and team_id using the route PUT /organizations/:org_id/team/:team_id/projects/:project_id.

See also

GitHub Developer Guide

pub fn update_discussion_legacy(&self) -> TeamsUpdateDiscussionLegacyBuilder

Update a discussion (Legacy)

Deprecation Notice: This endpoint route is deprecated and will be removed from the Teams API. We recommend migrating your existing code to use the new Update a discussion endpoint.

Edits the title and body text of a discussion post. Only the parameters you provide are updated. OAuth access tokens require the write:discussion scope.

See also

GitHub Developer Guide

pub fn get_discussion_legacy(&self) -> TeamsGetDiscussionLegacyBuilder

Get a discussion (Legacy)

Deprecation Notice: This endpoint route is deprecated and will be removed from the Teams API. We recommend migrating your existing code to use the new Get a discussion endpoint.

Get a specific discussion on a team's page. OAuth access tokens require the read:discussion scope.

See also

GitHub Developer Guide

pub fn delete_discussion_legacy(&self) -> TeamsDeleteDiscussionLegacyBuilder

Delete a discussion (Legacy)

Deprecation Notice: This endpoint route is deprecated and will be removed from the Teams API. We recommend migrating your existing code to use the new Delete a discussion endpoint.

Delete a discussion from a team's page. OAuth access tokens require the write:discussion scope.

See also

GitHub Developer Guide

pub fn list_pending_invitations_legacy(
&self
) -> TeamsListPendingInvitationsLegacyBuilder

List pending team invitations (Legacy)

Deprecation Notice: This endpoint route is deprecated and will be removed from the Teams API. We recommend migrating your existing code to use the new List pending team invitations endpoint.

The return hash contains a role field which refers to the Organization Invitation role and will be one of the following values: direct_member, admin, billing_manager, hiring_manager, or reinstate. If the invitee is not a GitHub member, the login field in the return hash will be null.

See also

GitHub Developer Guide

pub fn add_or_update_repo_permissions_legacy(
&self
) -> TeamsAddOrUpdateRepoPermissionsLegacyBuilder

Add or update team repository permissions (Legacy)

Deprecation Notice: This endpoint route is deprecated and will be removed from the Teams API. We recommend migrating your existing code to use the new Add or update team repository permissions endpoint.

To add a repository to a team or update the team's permission on a repository, the authenticated user must have admin access to the repository, and must be able to see the team. The repository must be owned by the organization, or a direct fork of a repository owned by the organization. You will get a 422 Unprocessable Entity status if you attempt to add a repository to a team that is not owned by the organization.

Note that, if you choose not to pass any parameters, you'll need to set Content-Length to zero when calling out to this endpoint. For more information, see "HTTP verbs."

See also

GitHub Developer Guide

pub fn check_permissions_for_repo_legacy(
&self
) -> TeamsCheckPermissionsForRepoLegacyBuilder

Check team permissions for a repository (Legacy)

Note: Repositories inherited through a parent team will also be checked.

Deprecation Notice: This endpoint route is deprecated and will be removed from the Teams API. We recommend migrating your existing code to use the new Check team permissions for a repository endpoint.

You can also get information about the specified repository, including what permissions the team grants on it, by passing the following custom media type via the Accept header:

See also

GitHub Developer Guide

pub fn remove_repo_legacy(&self) -> TeamsRemoveRepoLegacyBuilder

Remove a repository from a team (Legacy)

Deprecation Notice: This endpoint route is deprecated and will be removed from the Teams API. We recommend migrating your existing code to use the new Remove a repository from a team endpoint.

If the authenticated user is an organization owner or a team maintainer, they can remove any repositories from the team. To remove a repository from a team as an organization member, the authenticated user must have admin access to the repository and must be able to see the team. NOTE: This does not delete the repository, it just removes it from the team.

See also

GitHub Developer Guide

pub fn create_discussion_legacy(&self) -> TeamsCreateDiscussionLegacyBuilder

Create a discussion (Legacy)

Deprecation Notice: This endpoint route is deprecated and will be removed from the Teams API. We recommend migrating your existing code to use the new Create a discussion endpoint.

Creates a new discussion post on a team's page. OAuth access tokens require the write:discussion scope.

This endpoint triggers notifications. Creating content too quickly using this endpoint may result in abuse rate limiting. See "Abuse rate limits" and "Dealing with abuse rate limits" for details.

See also

GitHub Developer Guide

pub fn list_discussions_legacy(&self) -> TeamsListDiscussionsLegacyBuilder

List discussions (Legacy)

Deprecation Notice: This endpoint route is deprecated and will be removed from the Teams API. We recommend migrating your existing code to use the new List discussions endpoint.

List all discussions on a team's page. OAuth access tokens require the read:discussion scope.

See also

GitHub Developer Guide

pub fn delete_discussion_comment_in_org(
&self
) -> TeamsDeleteDiscussionCommentInOrgBuilder

Delete a discussion comment

Deletes a comment on a team discussion. OAuth access tokens require the write:discussion scope.

Note: You can also specify a team by org_id and team_id using the route DELETE /organizations/:org_id/team/:team_id/discussions/:discussion_number/comments/:comment_number.

See also

GitHub Developer Guide

pub fn get_discussion_comment_in_org(
&self
) -> TeamsGetDiscussionCommentInOrgBuilder

Get a discussion comment

Get a specific comment on a team discussion. OAuth access tokens require the read:discussion scope.

Note: You can also specify a team by org_id and team_id using the route GET /organizations/:org_id/team/:team_id/discussions/:discussion_number/comments/:comment_number.

See also

GitHub Developer Guide

pub fn update_discussion_comment_in_org(
&self
) -> TeamsUpdateDiscussionCommentInOrgBuilder

Update a discussion comment

Edits the body text of a discussion comment. OAuth access tokens require the write:discussion scope.

Note: You can also specify a team by org_id and team_id using the route PATCH /organizations/:org_id/team/:team_id/discussions/:discussion_number/comments/:comment_number.

See also

GitHub Developer Guide

pub fn get_membership_for_user_in_org(
&self
) -> TeamsGetMembershipForUserInOrgBuilder

Get team membership for a user

Team members will include the members of child teams.

To get a user's membership with a team, the team must be visible to the authenticated user.

Note: You can also specify a team by org_id and team_id using the route GET /organizations/:org_id/team/:team_id/memberships/:username.

Note: The role for organization owners returns as maintainer. For more information about maintainer roles, see Create a team.

See also

GitHub Developer Guide

pub fn add_or_update_membership_for_user_in_org(
&self
) -> TeamsAddOrUpdateMembershipForUserInOrgBuilder

Add or update team membership for a user

Team synchronization is available for organizations using GitHub Enterprise Cloud. For more information, see GitHub's products in the GitHub Help documentation.

Adds an organization member to a team. An authenticated organization owner or team maintainer can add organization members to a team.

Note: When you have team synchronization set up for a team with your organization's identity provider (IdP), you will see an error if you attempt to use the API for making changes to the team's membership. If you have access to manage group membership in your IdP, you can manage GitHub team membership through your identity provider, which automatically adds and removes team members in an organization. For more information, see "Synchronizing teams between your identity provider and GitHub."

An organization owner can add someone who is not part of the team's organization to a team. When an organization owner adds someone to a team who is not an organization member, this endpoint will send an invitation to the person via email. This newly-created membership will be in the "pending" state until the person accepts the invitation, at which point the membership will transition to the "active" state and the user will be added as a member of the team.

If the user is already a member of the team, this endpoint will update the role of the team member's role. To update the membership of a team member, the authenticated user must be an organization owner or a team maintainer.

Note: You can also specify a team by org_id and team_id using the route PUT /organizations/:org_id/team/:team_id/memberships/:username.

See also

GitHub Developer Guide

pub fn remove_membership_for_user_in_org(
&self
) -> TeamsRemoveMembershipForUserInOrgBuilder

Remove team membership for a user

Team synchronization is available for organizations using GitHub Enterprise Cloud. For more information, see GitHub's products in the GitHub Help documentation.

To remove a membership between a user and a team, the authenticated user must have 'admin' permissions to the team or be an owner of the organization that the team is associated with. Removing team membership does not delete the user, it just removes their membership from the team.

Note: When you have team synchronization set up for a team with your organization's identity provider (IdP), you will see an error if you attempt to use the API for making changes to the team's membership. If you have access to manage group membership in your IdP, you can manage GitHub team membership through your identity provider, which automatically adds and removes team members in an organization. For more information, see "Synchronizing teams between your identity provider and GitHub."

Note: You can also specify a team by org_id and team_id using the route DELETE /organizations/:org_id/team/:team_id/memberships/:username.

See also

GitHub Developer Guide

pub fn list_repos_in_org(&self) -> TeamsListReposInOrgBuilder

List team repositories

Lists a team's repositories visible to the authenticated user.

Note: You can also specify a team by org_id and team_id using the route GET /organizations/:org_id/team/:team_id/repos.

See also

GitHub Developer Guide

pub fn list_id_p_groups_for_legacy(&self) -> TeamsListIdPGroupsForLegacyBuilder

List IdP groups for a team (Legacy)

Deprecation Notice: This endpoint route is deprecated and will be removed from the Teams API. We recommend migrating your existing code to use the new List IdP groups for a team endpoint.

Team synchronization is available for organizations using GitHub Enterprise Cloud. For more information, see GitHub's products in the GitHub Help documentation.

List IdP groups connected to a team on GitHub.

See also

GitHub Developer Guide

pub fn create_or_update_id_p_group_connections_legacy(
&self
) -> TeamsCreateOrUpdateIdPGroupConnectionsLegacyBuilder

Create or update IdP group connections (Legacy)

Deprecation Notice: This endpoint route is deprecated and will be removed from the Teams API. We recommend migrating your existing code to use the new Create or update IdP group connections endpoint.

Team synchronization is available for organizations using GitHub Enterprise Cloud. For more information, see GitHub's products in the GitHub Help documentation.

Creates, updates, or removes a connection between a team and an IdP group. When adding groups to a team, you must include all new and existing groups to avoid replacing existing groups with the new ones. Specifying an empty groups array will remove all connections for a team.

See also

GitHub Developer Guide

pub fn list_projects_legacy(&self) -> TeamsListProjectsLegacyBuilder

List team projects (Legacy)

Deprecation Notice: This endpoint route is deprecated and will be removed from the Teams API. We recommend migrating your existing code to use the new List team projects endpoint.

Lists the organization projects for a team.

See also

GitHub Developer Guide

pub fn list_for_authenticated_user(
&self
) -> TeamsListForAuthenticatedUserBuilder

List teams for the authenticated user

List all of the teams across all of the organizations to which the authenticated user belongs. This method requires user, repo, or read:org scope when authenticating via OAuth.

See also

GitHub Developer Guide

pub fn check_permissions_for_project_legacy(
&self
) -> TeamsCheckPermissionsForProjectLegacyBuilder

Check team permissions for a project (Legacy)

Deprecation Notice: This endpoint route is deprecated and will be removed from the Teams API. We recommend migrating your existing code to use the new Check team permissions for a project endpoint.

Checks whether a team has read, write, or admin permissions for an organization project. The response includes projects inherited from a parent team.

See also

GitHub Developer Guide

pub fn remove_project_legacy(&self) -> TeamsRemoveProjectLegacyBuilder

Remove a project from a team (Legacy)

Deprecation Notice: This endpoint route is deprecated and will be removed from the Teams API. We recommend migrating your existing code to use the new Remove a project from a team endpoint.

Removes an organization project from a team. An organization owner or a team maintainer can remove any project from the team. To remove a project from a team as an organization member, the authenticated user must have read access to both the team and project, or admin access to the team or project. Note: This endpoint removes the project from the team, but does not delete it.

See also

GitHub Developer Guide

pub fn add_or_update_project_permissions_legacy(
&self
) -> TeamsAddOrUpdateProjectPermissionsLegacyBuilder

Add or update team project permissions (Legacy)

Deprecation Notice: This endpoint route is deprecated and will be removed from the Teams API. We recommend migrating your existing code to use the new Add or update team project permissions endpoint.

Adds an organization project to a team. To add a project to a team or update the team's permission on a project, the authenticated user must have admin permissions for the project. The project and team must be part of the same organization.

See also

GitHub Developer Guide

pub fn get_legacy(&self) -> TeamsGetLegacyBuilder

Get a team (Legacy)

Deprecation Notice: This endpoint route is deprecated and will be removed from the Teams API. We recommend migrating your existing code to use the Get a team by name endpoint.

See also

GitHub Developer Guide

pub fn delete_legacy(&self) -> TeamsDeleteLegacyBuilder

Delete a team (Legacy)

Deprecation Notice: This endpoint route is deprecated and will be removed from the Teams API. We recommend migrating your existing code to use the new Delete a team endpoint.

To delete a team, the authenticated user must be an organization owner or team maintainer.

If you are an organization owner, deleting a parent team will delete all of its child teams as well.

See also

GitHub Developer Guide

pub fn update_legacy(&self) -> TeamsUpdateLegacyBuilder

Update a team (Legacy)

Deprecation Notice: This endpoint route is deprecated and will be removed from the Teams API. We recommend migrating your existing code to use the new Update a team endpoint.

To edit a team, the authenticated user must either be an organization owner or a team maintainer.

Note: With nested teams, the privacy for parent teams cannot be secret.

See also

GitHub Developer Guide

pub fn delete_in_org(&self) -> TeamsDeleteInOrgBuilder

Delete a team

To delete a team, the authenticated user must be an organization owner or team maintainer.

Note: You can also specify a team by org_id and team_id using the route DELETE /organizations/:org_id/team/:team_id.

If you are an organization owner, deleting a parent team will delete all of its child teams as well.

See also

GitHub Developer Guide

pub fn get_by_name(&self) -> TeamsGetByNameBuilder

Get a team by name

Gets a team using the team's slug. GitHub generates the slug from the team name.

Note: You can also specify a team by org_id and team_id using the route GET /organizations/:org_id/team/:team_id.

See also

GitHub Developer Guide

pub fn update_in_org(&self) -> TeamsUpdateInOrgBuilder

Update a team

To edit a team, the authenticated user must either be an organization owner or a team maintainer.

Note: You can also specify a team by org_id and team_id using the route PATCH /organizations/:org_id/team/:team_id.

See also

GitHub Developer Guide

pub fn list_id_p_groups_for_org(&self) -> TeamsListIdPGroupsForOrgBuilder

List IdP groups for an organization

Team synchronization is available for organizations using GitHub Enterprise Cloud. For more information, see GitHub's products in the GitHub Help documentation.

List IdP groups available in an organization. You can limit your page results using the per_page parameter. GitHub generates a url-encoded page token using a cursor value for where the next page begins. For more information on cursor pagination, see "Offset and Cursor Pagination explained."

The per_page parameter provides pagination for a list of IdP groups the authenticated user can access in an organization. For example, if the user octocat wants to see two groups per page in octo-org via cURL, it would look like this:

See also

GitHub Developer Guide

pub fn list_repos_legacy(&self) -> TeamsListReposLegacyBuilder

List team repositories (Legacy)

Deprecation Notice: This endpoint route is deprecated and will be removed from the Teams API. We recommend migrating your existing code to use the new List team repositories endpoint.

See also

GitHub Developer Guide

pub fn add_or_update_membership_for_user_legacy(
&self
) -> TeamsAddOrUpdateMembershipForUserLegacyBuilder

Add or update team membership for a user (Legacy)

Deprecation Notice: This endpoint route is deprecated and will be removed from the Teams API. We recommend migrating your existing code to use the new Add or update team membership for a user endpoint.

Team synchronization is available for organizations using GitHub Enterprise Cloud. For more information, see GitHub's products in the GitHub Help documentation.

If the user is already a member of the team's organization, this endpoint will add the user to the team. To add a membership between an organization member and a team, the authenticated user must be an organization owner or a team maintainer.

Note: When you have team synchronization set up for a team with your organization's identity provider (IdP), you will see an error if you attempt to use the API for making changes to the team's membership. If you have access to manage group membership in your IdP, you can manage GitHub team membership through your identity provider, which automatically adds and removes team members in an organization. For more information, see "Synchronizing teams between your identity provider and GitHub."

If the user is unaffiliated with the team's organization, this endpoint will send an invitation to the user via email. This newly-created membership will be in the "pending" state until the user accepts the invitation, at which point the membership will transition to the "active" state and the user will be added as a member of the team. To add a membership between an unaffiliated user and a team, the authenticated user must be an organization owner.

If the user is already a member of the team, this endpoint will update the role of the team member's role. To update the membership of a team member, the authenticated user must be an organization owner or a team maintainer.

See also

GitHub Developer Guide

pub fn remove_membership_for_user_legacy(
&self
) -> TeamsRemoveMembershipForUserLegacyBuilder

Remove team membership for a user (Legacy)

Deprecation Notice: This endpoint route is deprecated and will be removed from the Teams API. We recommend migrating your existing code to use the new Remove team membership for a user endpoint.

Team synchronization is available for organizations using GitHub Enterprise Cloud. For more information, see GitHub's products in the GitHub Help documentation.

To remove a membership between a user and a team, the authenticated user must have 'admin' permissions to the team or be an owner of the organization that the team is associated with. Removing team membership does not delete the user, it just removes their membership from the team.

Note: When you have team synchronization set up for a team with your organization's identity provider (IdP), you will see an error if you attempt to use the API for making changes to the team's membership. If you have access to manage group membership in your IdP, you can manage GitHub team membership through your identity provider, which automatically adds and removes team members in an organization. For more information, see "Synchronizing teams between your identity provider and GitHub."

See also

GitHub Developer Guide

pub fn get_membership_for_user_legacy(
&self
) -> TeamsGetMembershipForUserLegacyBuilder

Get team membership for a user (Legacy)

Deprecation Notice: This endpoint route is deprecated and will be removed from the Teams API. We recommend migrating your existing code to use the new Get team membership for a user endpoint.

Team members will include the members of child teams.

To get a user's membership with a team, the team must be visible to the authenticated user.

Note: The role for organization owners returns as maintainer. For more information about maintainer roles, see Create a team.

See also

GitHub Developer Guide

pub fn list_members_in_org(&self) -> TeamsListMembersInOrgBuilder

List team members

Team members will include the members of child teams.

To list members in a team, the team must be visible to the authenticated user.

See also

GitHub Developer Guide

pub fn list_discussions_in_org(&self) -> TeamsListDiscussionsInOrgBuilder

List discussions

List all discussions on a team's page. OAuth access tokens require the read:discussion scope.

Note: You can also specify a team by org_id and team_id using the route GET /organizations/:org_id/team/:team_id/discussions.

See also

GitHub Developer Guide

pub fn create_discussion_in_org(&self) -> TeamsCreateDiscussionInOrgBuilder

Create a discussion

Creates a new discussion post on a team's page. OAuth access tokens require the write:discussion scope.

This endpoint triggers notifications. Creating content too quickly using this endpoint may result in abuse rate limiting. See "Abuse rate limits" and "Dealing with abuse rate limits" for details.

Note: You can also specify a team by org_id and team_id using the route POST /organizations/:org_id/team/:team_id/discussions.

See also

GitHub Developer Guide

pub fn get_member_legacy(&self) -> TeamsGetMemberLegacyBuilder

Get team member (Legacy)

The "Get team member" endpoint (described below) is deprecated.

We recommend using the Get team membership for a user endpoint instead. It allows you to get both active and pending memberships.

To list members in a team, the team must be visible to the authenticated user.

See also

GitHub Developer Guide

pub fn remove_member_legacy(&self) -> TeamsRemoveMemberLegacyBuilder

Remove team member (Legacy)

The "Remove team member" endpoint (described below) is deprecated.

We recommend using the Remove team membership for a user endpoint instead. It allows you to remove both active and pending memberships.

Team synchronization is available for organizations using GitHub Enterprise Cloud. For more information, see GitHub's products in the GitHub Help documentation.

To remove a team member, the authenticated user must have 'admin' permissions to the team or be an owner of the org that the team is associated with. Removing a team member does not delete the user, it just removes them from the team.

Note: When you have team synchronization set up for a team with your organization's identity provider (IdP), you will see an error if you attempt to use the API for making changes to the team's membership. If you have access to manage group membership in your IdP, you can manage GitHub team membership through your identity provider, which automatically adds and removes team members in an organization. For more information, see "Synchronizing teams between your identity provider and GitHub."

See also

GitHub Developer Guide

pub fn add_member_legacy(&self) -> TeamsAddMemberLegacyBuilder

Add team member (Legacy)

The "Add team member" endpoint (described below) is deprecated.

We recommend using the Add or update team membership for a user endpoint instead. It allows you to invite new organization members to your teams.

Team synchronization is available for organizations using GitHub Enterprise Cloud. For more information, see GitHub's products in the GitHub Help documentation.

To add someone to a team, the authenticated user must be an organization owner or a team maintainer in the team they're changing. The person being added to the team must be a member of the team's organization.

Note: When you have team synchronization set up for a team with your organization's identity provider (IdP), you will see an error if you attempt to use the API for making changes to the team's membership. If you have access to manage group membership in your IdP, you can manage GitHub team membership through your identity provider, which automatically adds and removes team members in an organization. For more information, see "Synchronizing teams between your identity provider and GitHub."

Note that you'll need to set Content-Length to zero when calling out to this endpoint. For more information, see "HTTP verbs."

See also

GitHub Developer Guide

pub fn list_pending_invitations_in_org(
&self
) -> TeamsListPendingInvitationsInOrgBuilder

List pending team invitations

The return hash contains a role field which refers to the Organization Invitation role and will be one of the following values: direct_member, admin, billing_manager, hiring_manager, or reinstate. If the invitee is not a GitHub member, the login field in the return hash will be null.

Note: You can also specify a team by org_id and team_id using the route GET /organizations/:org_id/team/:team_id/invitations.

See also

GitHub Developer Guide

pub fn list_projects_in_org(&self) -> TeamsListProjectsInOrgBuilder

List team projects

Lists the organization projects for a team.

Note: You can also specify a team by org_id and team_id using the route GET /organizations/:org_id/team/:team_id/projects.

See also

GitHub Developer Guide

pub fn create_discussion_comment_legacy(
&self
) -> TeamsCreateDiscussionCommentLegacyBuilder

Create a discussion comment (Legacy)

Deprecation Notice: This endpoint route is deprecated and will be removed from the Teams API. We recommend migrating your existing code to use the new Create a discussion comment endpoint.

Creates a new comment on a team discussion. OAuth access tokens require the write:discussion scope.

This endpoint triggers notifications. Creating content too quickly using this endpoint may result in abuse rate limiting. See "Abuse rate limits" and "Dealing with abuse rate limits" for details.

See also

GitHub Developer Guide

pub fn list_discussion_comments_legacy(
&self
) -> TeamsListDiscussionCommentsLegacyBuilder

List discussion comments (Legacy)

Deprecation Notice: This endpoint route is deprecated and will be removed from the Teams API. We recommend migrating your existing code to use the new List discussion comments endpoint.

List all comments on a team discussion. OAuth access tokens require the read:discussion scope.

See also

GitHub Developer Guide

pub fn list(&self) -> TeamsListBuilder

List teams

Lists all teams in an organization that are visible to the authenticated user.

See also

GitHub Developer Guide

pub fn create(&self) -> TeamsCreateBuilder

Create a team

To create a team, the authenticated user must be a member or owner of :org. By default, organization members can create teams. Organization owners can limit team creation to organization owners. For more information, see "Setting team creation permissions."

When you create a new team, you automatically become a team maintainer without explicitly adding yourself to the optional array of maintainers. For more information, see "About teams" in the GitHub Help documentation.

See also

GitHub Developer Guide

pub fn create_or_update_id_p_group_connections_in_org(
&self
) -> TeamsCreateOrUpdateIdPGroupConnectionsInOrgBuilder

Create or update IdP group connections

Team synchronization is available for organizations using GitHub Enterprise Cloud. For more information, see GitHub's products in the GitHub Help documentation.

Creates, updates, or removes a connection between a team and an IdP group. When adding groups to a team, you must include all new and existing groups to avoid replacing existing groups with the new ones. Specifying an empty groups array will remove all connections for a team.

Note: You can also specify a team by org_id and team_id using the route PATCH /organizations/:org_id/team/:team_id/team-sync/group-mappings.

See also

GitHub Developer Guide

pub fn list_id_p_groups_in_org(&self) -> TeamsListIdPGroupsInOrgBuilder

List IdP groups for a team

Team synchronization is available for organizations using GitHub Enterprise Cloud. For more information, see GitHub's products in the GitHub Help documentation.

List IdP groups connected to a team on GitHub.

Note: You can also specify a team by org_id and team_id using the route GET /organizations/:org_id/team/:team_id/team-sync/group-mappings.

See also

GitHub Developer Guide

pub fn list_child_legacy(&self) -> TeamsListChildLegacyBuilder

List child teams (Legacy)

Deprecation Notice: This endpoint route is deprecated and will be removed from the Teams API. We recommend migrating your existing code to use the new List child teams endpoint.

See also

GitHub Developer Guide

pub fn get_discussion_in_org(&self) -> TeamsGetDiscussionInOrgBuilder

Get a discussion

Get a specific discussion on a team's page. OAuth access tokens require the read:discussion scope.

Note: You can also specify a team by org_id and team_id using the route GET /organizations/:org_id/team/:team_id/discussions/:discussion_number.

See also

GitHub Developer Guide

pub fn delete_discussion_in_org(&self) -> TeamsDeleteDiscussionInOrgBuilder

Delete a discussion

Delete a discussion from a team's page. OAuth access tokens require the write:discussion scope.

Note: You can also specify a team by org_id and team_id using the route DELETE /organizations/:org_id/team/:team_id/discussions/:discussion_number.

See also

GitHub Developer Guide

pub fn update_discussion_in_org(&self) -> TeamsUpdateDiscussionInOrgBuilder

Update a discussion

Edits the title and body text of a discussion post. Only the parameters you provide are updated. OAuth access tokens require the write:discussion scope.

Note: You can also specify a team by org_id and team_id using the route PATCH /organizations/:org_id/team/:team_id/discussions/:discussion_number.

See also

GitHub Developer Guide

pub fn list_discussion_comments_in_org(
&self
) -> TeamsListDiscussionCommentsInOrgBuilder

List discussion comments

List all comments on a team discussion. OAuth access tokens require the read:discussion scope.

Note: You can also specify a team by org_id and team_id using the route GET /organizations/:org_id/team/:team_id/discussions/:discussion_number/comments.

See also

GitHub Developer Guide

pub fn create_discussion_comment_in_org(
&self
) -> TeamsCreateDiscussionCommentInOrgBuilder

Create a discussion comment

Creates a new comment on a team discussion. OAuth access tokens require the write:discussion scope.

This endpoint triggers notifications. Creating content too quickly using this endpoint may result in abuse rate limiting. See "Abuse rate limits" and "Dealing with abuse rate limits" for details.

Note: You can also specify a team by org_id and team_id using the route POST /organizations/:org_id/team/:team_id/discussions/:discussion_number/comments.

See also

GitHub Developer Guide

pub fn add_or_update_repo_permissions_in_org(
&self
) -> TeamsAddOrUpdateRepoPermissionsInOrgBuilder

Add or update team repository permissions

To add a repository to a team or update the team's permission on a repository, the authenticated user must have admin access to the repository, and must be able to see the team. The repository must be owned by the organization, or a direct fork of a repository owned by the organization. You will get a 422 Unprocessable Entity status if you attempt to add a repository to a team that is not owned by the organization. Note that, if you choose not to pass any parameters, you'll need to set Content-Length to zero when calling out to this endpoint. For more information, see "HTTP verbs."

Note: You can also specify a team by org_id and team_id using the route PUT /organizations/:org_id/team/:team_id/repos/:owner/:repo.

For more information about the permission levels, see "Repository permission levels for an organization" in the GitHub Help documentation.

See also

GitHub Developer Guide

pub fn check_permissions_for_repo_in_org(
&self
) -> TeamsCheckPermissionsForRepoInOrgBuilder

Check team permissions for a repository

Checks whether a team has admin, push, maintain, triage, or pull permission for a repository. Repositories inherited through a parent team will also be checked.

Note: You can also specify a team by org_id and team_id using the route GET /organizations/:org_id/team/:team_id/repos/:owner/:repo.

You can also get information about the specified repository, including what permissions the team grants on it, by passing the following custom media type via the Accept header:

See also

GitHub Developer Guide

pub fn remove_repo_in_org(&self) -> TeamsRemoveRepoInOrgBuilder

Remove a repository from a team

If the authenticated user is an organization owner or a team maintainer, they can remove any repositories from the team. To remove a repository from a team as an organization member, the authenticated user must have admin access to the repository and must be able to see the team. This does not delete the repository, it just removes it from the team.

Note: You can also specify a team by org_id and team_id using the route DELETE /organizations/:org_id/team/:team_id/repos/:owner/:repo.

See also

GitHub Developer Guide

pub fn list_members_legacy(&self) -> TeamsListMembersLegacyBuilder

List team members (Legacy)

Deprecation Notice: This endpoint route is deprecated and will be removed from the Teams API. We recommend migrating your existing code to use the new List team members endpoint.

Team members will include the members of child teams.

See also

GitHub Developer Guide