Struct octorest::apis::UsersApi

pub struct UsersApi<'t> { /* fields omitted */ }

This is supported on feature="gh-users" only.

Users API

Implementations

impl<'client> UsersApi<'client>

pub fn list_following_for_user(&self) -> UsersListFollowingForUserBuilder

List the people a user follows

Lists the people who the specified user follows.

See also

GitHub Developer Guide

pub fn get_by_username(&self) -> UsersGetByUsernameBuilder

Get a user

Provides publicly available information about someone with a GitHub account.

GitHub Apps with the Plan user permission can use this endpoint to retrieve information about a user's GitHub plan. The GitHub App must be authenticated as a user. See "Identifying and authorizing users for GitHub Apps" for details about authentication. For an example response, see "Response with GitHub plan information."

The email key in the following response is the publicly visible email address from your GitHub profile page. When setting up your profile, you can select a primary email address to be “public” which provides an email entry for this endpoint. If you do not set a public email address for email, then it will have a value of null. You only see publicly visible email addresses when authenticated with GitHub. For more information, see Authentication.

The Emails API enables you to list all of your email addresses, and toggle a primary email to be visible publicly. For more information, see "Emails API".

See also

GitHub Developer Guide

pub fn add_email_for_authenticated(
&self
) -> UsersAddEmailForAuthenticatedBuilder

Add an email address for the authenticated user

This endpoint is accessible with the user scope.

See also

GitHub Developer Guide

pub fn list_emails_for_authenticated(
&self
) -> UsersListEmailsForAuthenticatedBuilder

List email addresses for the authenticated user

Lists all of your email addresses, and specifies which one is visible to the public. This endpoint is accessible with the user:email scope.

See also

GitHub Developer Guide

pub fn delete_email_for_authenticated(
&self
) -> UsersDeleteEmailForAuthenticatedBuilder

Delete an email address for the authenticated user

This endpoint is accessible with the user scope.

See also

GitHub Developer Guide

pub fn get_public_ssh_key_for_authenticated(
&self
) -> UsersGetPublicSshKeyForAuthenticatedBuilder

Get a public SSH key for the authenticated user

View extended details for a single public SSH key. Requires that you are authenticated via Basic Auth or via OAuth with at least read:public_key scope.

See also

GitHub Developer Guide

pub fn delete_public_ssh_key_for_authenticated(
&self
) -> UsersDeletePublicSshKeyForAuthenticatedBuilder

Delete a public SSH key for the authenticated user

Removes a public SSH key from the authenticated user's GitHub account. Requires that you are authenticated via Basic Auth or via OAuth with at least admin:public_key scope.

See also

GitHub Developer Guide

pub fn list_public_keys_for_user(&self) -> UsersListPublicKeysForUserBuilder

List public keys for a user

Lists the verified public SSH keys for a user. This is accessible by anyone.

See also

GitHub Developer Guide

pub fn list(&self) -> UsersListBuilder

List users

Lists all users, in the order that they signed up on GitHub. This list includes personal user accounts and organization accounts.

Note: Pagination is powered exclusively by the since parameter. Use the Link header to get the URL for the next page of users.

See also

GitHub Developer Guide

pub fn list_followed_by_authenticated(
&self
) -> UsersListFollowedByAuthenticatedBuilder

List the people the authenticated user follows

Lists the people who the authenticated user follows.

See also

GitHub Developer Guide

pub fn create_public_ssh_key_for_authenticated(
&self
) -> UsersCreatePublicSshKeyForAuthenticatedBuilder

Create a public SSH key for the authenticated user

Adds a public SSH key to the authenticated user's GitHub account. Requires that you are authenticated via Basic Auth, or OAuth with at least write:public_key scope.

See also

GitHub Developer Guide

pub fn list_public_ssh_keys_for_authenticated(
&self
) -> UsersListPublicSshKeysForAuthenticatedBuilder

List public SSH keys for the authenticated user

Lists the public SSH keys for the authenticated user's GitHub account. Requires that you are authenticated via Basic Auth or via OAuth with at least read:public_key scope.

See also

GitHub Developer Guide

pub fn list_followers_for_user(&self) -> UsersListFollowersForUserBuilder

List followers of a user

Lists the people following the specified user.

See also

GitHub Developer Guide

pub fn list_followers_for_authenticated_user(
&self
) -> UsersListFollowersForAuthenticatedUserBuilder

List followers of the authenticated user

Lists the people following the authenticated user.

See also

GitHub Developer Guide

pub fn follow(&self) -> UsersFollowBuilder

Follow a user

Note that you'll need to set Content-Length to zero when calling out to this endpoint. For more information, see "HTTP verbs."

Following a user requires the user to be logged in and authenticated with basic auth or OAuth with the user:follow scope.

See also

GitHub Developer Guide

pub fn unfollow(&self) -> UsersUnfollowBuilder

Unfollow a user

Unfollowing a user requires the user to be logged in and authenticated with basic auth or OAuth with the user:follow scope.

See also

GitHub Developer Guide

pub fn check_person_is_followed_by_authenticated(
&self
) -> UsersCheckPersonIsFollowedByAuthenticatedBuilder

Check if a person is followed by the authenticated user

See also

GitHub Developer Guide

pub fn list_gpg_keys_for_user(&self) -> UsersListGpgKeysForUserBuilder

List GPG keys for a user

Lists the GPG keys for a user. This information is accessible by anyone.

See also

GitHub Developer Guide

pub fn update_authenticated(&self) -> UsersUpdateAuthenticatedBuilder

Update the authenticated user

Note: If your email is set to private and you send an email parameter as part of this request to update your profile, your privacy settings are still enforced: the email address will not be displayed on your public profile or via the API.

See also

GitHub Developer Guide

pub fn get_authenticated(&self) -> UsersGetAuthenticatedBuilder

Get the authenticated user

Lists public and private profile information when authenticated through basic auth or OAuth with the user scope.

Lists public profile information when authenticated through OAuth without the user scope.

See also

GitHub Developer Guide

pub fn list_public_emails_for_authenticated(
&self
) -> UsersListPublicEmailsForAuthenticatedBuilder

List public email addresses for the authenticated user

Lists your publicly visible email address, which you can set with the Set primary email visibility for the authenticated user endpoint. This endpoint is accessible with the user:email scope.

See also

GitHub Developer Guide

pub fn set_primary_email_visibility_for_authenticated(
&self
) -> UsersSetPrimaryEmailVisibilityForAuthenticatedBuilder

Set primary email visibility for the authenticated user

Sets the visibility for your primary email addresses.

See also

GitHub Developer Guide

pub fn check_following_for_user(&self) -> UsersCheckFollowingForUserBuilder

Check if a user follows another user

See also

GitHub Developer Guide

pub fn list_blocked_by_authenticated(
&self
) -> UsersListBlockedByAuthenticatedBuilder

List users blocked by the authenticated user

List the users you've blocked on your personal account.

See also

GitHub Developer Guide

pub fn create_gpg_key_for_authenticated(
&self
) -> UsersCreateGpgKeyForAuthenticatedBuilder

Create a GPG key for the authenticated user

Adds a GPG key to the authenticated user's GitHub account. Requires that you are authenticated via Basic Auth, or OAuth with at least write:gpg_key scope.

See also

GitHub Developer Guide

pub fn list_gpg_keys_for_authenticated(
&self
) -> UsersListGpgKeysForAuthenticatedBuilder

List GPG keys for the authenticated user

Lists the current user's GPG keys. Requires that you are authenticated via Basic Auth or via OAuth with at least read:gpg_key scope.

See also

GitHub Developer Guide

pub fn get_context_for_user(&self) -> UsersGetContextForUserBuilder

Get contextual information for a user

Provides hovercard information when authenticated through basic auth or OAuth with the repo scope. You can find out more about someone in relation to their pull requests, issues, repositories, and organizations.

The subject_type and subject_id parameters provide context for the person's hovercard, which returns more information than without the parameters. For example, if you wanted to find out more about octocat who owns the Spoon-Knife repository via cURL, it would look like this:

See also

GitHub Developer Guide

pub fn block(&self) -> UsersBlockBuilder

Block a user

See also

GitHub Developer Guide

pub fn unblock(&self) -> UsersUnblockBuilder

Unblock a user

See also

GitHub Developer Guide

pub fn check_blocked(&self) -> UsersCheckBlockedBuilder

Check if a user is blocked by the authenticated user

If the user is blocked:

If the user is not blocked:

See also

GitHub Developer Guide

pub fn delete_gpg_key_for_authenticated(
&self
) -> UsersDeleteGpgKeyForAuthenticatedBuilder

Delete a GPG key for the authenticated user

Removes a GPG key from the authenticated user's GitHub account. Requires that you are authenticated via Basic Auth or via OAuth with at least admin:gpg_key scope.

See also

GitHub Developer Guide

pub fn get_gpg_key_for_authenticated(
&self
) -> UsersGetGpgKeyForAuthenticatedBuilder

Get a GPG key for the authenticated user

View extended details for a single GPG key. Requires that you are authenticated via Basic Auth or via OAuth with at least read:gpg_key scope.

See also

GitHub Developer Guide