Struct octorest::Client

pub struct Client { /* fields omitted */ }

An authenticated GitHub API client

Implementations

impl Client

pub fn reqwest(&self) -> &Client

pub fn auth(&self) -> &Mutex<Auth>

impl Client

pub fn create_unauthenticated(reqwest: Client) -> Self

Creates an unauthenticated API client.

Unauthenticated clients are limited to 60 requests per hour per IP address. Consider creating a token for a dummy user account and use from_token to increase rate limit to 5000 per hour.

pub fn create_from_token(reqwest: Client, token: &str) -> Self

Creates an API client associated with a static access token.

The token can be created from the GitHub settings page directly, obtained via OAuth web/non-web flow or obtained via integrations/installations API. However, there are dedicated methods for the latter methods.

pub async fn create_from_oauth_web_flow(
reqwest: Client,
client_id: impl Into<Cow<'static, str>>,
client_secret: impl Into<Cow<'static, str>>,
code: String
) -> Result<Self>

Creates an API client authenticated by web flow.

Usage

To authorize a web user, first redirect the user to https://github.com/login/oauth/authorize. When the user gets redirected back from GitHub, the request should contain a GET parameter code. Pass this parameter along with the app cient ID and secret to create the client.

pub async fn create_as_app<'_, '_>(
reqwest: Client,
app_id: &'_ str,
private_pem: &'_ [u8]
) -> Result<Self>

Creates an API client authenticated as a GitHub app (not an installation)

pub async fn get_auth_header<'_>(&'_ self) -> Result<Option<Arc<str>>>

impl Client

pub fn markdown(&self) -> MarkdownApi

This is supported on feature="gh-markdown" only.

Markdown API

pub fn users(&self) -> UsersApi

This is supported on feature="gh-users" only.

Users API

pub fn migrations(&self) -> MigrationsApi

This is supported on feature="gh-migrations" only.

Migrations API

pub fn repos(&self) -> ReposApi

This is supported on feature="gh-repos" only.

Repos API

pub fn pulls(&self) -> PullsApi

This is supported on feature="gh-pulls" only.

Pulls API

pub fn reactions(&self) -> ReactionsApi

This is supported on feature="gh-reactions" only.

Reactions API

pub fn actions(&self) -> ActionsApi

This is supported on feature="gh-actions" only.

Actions API

pub fn activity(&self) -> ActivityApi

This is supported on feature="gh-activity" only.

Activity API

pub fn git(&self) -> GitApi

This is supported on feature="gh-git" only.

Git API

pub fn checks(&self) -> ChecksApi

This is supported on feature="gh-checks" only.

Checks API

pub fn issues(&self) -> IssuesApi

This is supported on feature="gh-issues" only.

Issues API

pub fn gists(&self) -> GistsApi

This is supported on feature="gh-gists" only.

Gists API

pub fn search(&self) -> SearchApi

This is supported on feature="gh-search" only.

Search API

pub fn teams(&self) -> TeamsApi

This is supported on feature="gh-teams" only.

Teams API

pub fn apps(&self) -> AppsApi

This is supported on feature="gh-apps" only.

Apps API

pub fn oauth_authorizations(&self) -> OauthAuthorizationsApi

This is supported on feature="gh-oauth-authorizations" only.

Oauth Authorizations API

pub fn orgs(&self) -> OrgsApi

This is supported on feature="gh-orgs" only.

Orgs API

pub fn codes_of_conduct(&self) -> CodesOfConductApi

This is supported on feature="gh-codes-of-conduct" only.

Codes Of Conduct API

pub fn gitignore(&self) -> GitignoreApi

This is supported on feature="gh-gitignore" only.

Gitignore API

pub fn licenses(&self) -> LicensesApi

This is supported on feature="gh-licenses" only.

Licenses API

pub fn emojis(&self) -> EmojisApi

This is supported on feature="gh-emojis" only.

Emojis API

pub fn rate_limit(&self) -> RateLimitApi

This is supported on feature="gh-rate-limit" only.

Rate Limit API

pub fn projects(&self) -> ProjectsApi

This is supported on feature="gh-projects" only.

Projects API

pub fn code_scanning(&self) -> CodeScanningApi

This is supported on feature="gh-code-scanning" only.

Code Scanning API

pub fn interactions(&self) -> InteractionsApi

This is supported on feature="gh-interactions" only.

Interactions API

pub fn scim(&self) -> ScimApi

This is supported on feature="gh-scim" only.

Scim API

pub fn meta(&self) -> MetaApi

This is supported on feature="gh-meta" only.

Meta API