Module octorest::types

Miscellaneous data types used in the GitHub API.

The API is designed such that users shall not need to explicitly import types from this module.

Structs

AccessToken
Access Token

AccountId
Account Id

ActionsBillingUsage
Actions Billing Usage

ActionsPublicKey
The public key used for setting Actions Secrets.

ActionsSecret
Set secrets for GitHub Actions.

All
All

ApiOverview
Api Overview

ApplicationGrant
The authorization associated with an OAuth Access.

Artifact
An artifact

ArtifactId
Artifact Id

AssetId
Asset Id

AuthenticationToken
Authentication Token

Authorization
The authorization for an OAuth app, GitHub App, or a Personal Access Token.

AuthorizationId
Authorization Id

BaseGist
Base Gist

BasicError
Basic Error

Before
Before

Blob
Blob

Branch
Branch

BranchProtection
Branch Protection

BranchRestrictionPolicy
Branch Restriction Policy

BranchShort
Branch Short

BranchWithProtection
Branch With Protection

CardId
Card Id

CheckAnnotation
Check Annotation

CheckName
Check Name

CheckRun
A check performed on the code of a given code change

CheckRunId
Check Run Id

CheckSuite
A suite of checks performed on the code of a given code change

CheckSuiteId
Check Suite Id

CheckSuitePreference
Check suite configuration preferences for a repository.

ClientId
Client Id

CloneTraffic
Clone Traffic

CodeFrequencyStatItem
Code Frequency Stat Item

CodeOfConduct
Code Of Conduct

CodeOfConductSimple
Code of Conduct Simple

CodeScanningAlert
Code Scanning Alert

CodeSearchResultItem
Code Search Result Item

Collaborator
Collaborator

ColumnId
Column Id

CombinedBillingUsage
Combined Billing Usage

CombinedCommitStatus
Combined Commit Status

CommentId
Comment Id

CommentNumber
Comment Number

Commit
Commit

CommitActivity
Commit Activity

CommitComment
Commit Comment

CommitComparison
Commit Comparison

CommitSearchResultItem
Commit Search Result Item

CommitSha
Commit Sha

CommunityHealthFile
Community Health File

CommunityProfile
Community Profile

ContentDirectoryItem
Content Directory Item

ContentFile
Content File

ContentReferenceAttachment
Content Reference attachments allow you to provide context around URLs posted in comments

ContentSubmodule
An object describing a symlink

ContentSymlink
An object describing a symlink

ContentTraffic
Content Traffic

ContentTree
Content Tree

ContentType
Content Type

Contributor
Contributor

ContributorActivity
Contributor Activity

CredentialAuthorization
Credential Authorization

DeployKey
An SSH key granting access to a single repository.

Deployment
A request for a specific ref(branch,sha,tag) to be deployed

DeploymentId
Deployment Id

DeploymentStatus
The status of a deployment.

DiffEntry
Diff Entry

Direction
Direction

DiscussionNumber
Discussion Number

Email
Email

Enterprise
An enterprise account

EnterpriseId
Enterprise Id

Feed
Feed

FileCommit
File Commit

ForbiddenGist
Forbidden Gist

FullRepository
Full Repository

GistComment
A comment made to a gist.

GistCommit
Gist Commit

GistFull
Gist Full

GistId
Gist Id

GistSimple
Gist Simple

GitCommit
Low-level Git commit operations within a repository

GitRef
Git references within a repository

GitTag
Metadata for a Git tag

GitTree
The hierarchy between files in a Git repository.

GitUser
Metaproperties for Git author/committer information.

GitignoreTemplate
Gitignore Template

GpgKey
A unique encryption key

GpgKeyId
Gpg Key Id

GrantId
Grant Id

GroupMapping
External Groups to be mapped to a team for membership

HeaderLink
Header Link

Hook
Webhooks for repositories.

HookId
Hook Id

HookResponse
Hook Response

Hovercard
Hovercard

Import
A repository import from an external source.

Installation
Installation

InstallationId
Installation Id

InstallationToken
Authentication token for a GitHub App installed on a user or org.

Integration
GitHub apps are a new way to extend GitHub. They can be installed directly on organizations and user accounts and granted access to specific repositories. They come with granular permissions and built-in webhooks. GitHub apps are first class actors within GitHub.

InteractionLimit
Interaction limit settings.

InvitationId
Invitation Id

Issue
Issues are a great way to keep track of tasks, enhancements, and bugs for your projects.

IssueComment
Comments provide a way for people to collaborate on an issue.

IssueEvent
Issue Event

IssueEventDismissedReview
Issue Event Dismissed Review

IssueEventForIssue
Issue Event for Issue

IssueEventLabel
Issue Event Label

IssueEventMilestone
Issue Event Milestone

IssueEventProjectCard
Issue Event Project Card

IssueEventRename
Issue Event Rename

IssueNumber
Issue Number

IssueSearchResultItem
Issue Search Result Item

IssueSimple
Issue Simple

Job
Information of a job execution in a workflow run

JobId
Job Id

Key
Key

KeyId
Key Id

KeySimple
Key Simple

Label
Color-coded labels help you categorize and filter your issues (just like labels in Gmail).

LabelSearchResultItem
Label Search Result Item

Labels
Labels

Language
Language

License
License

LicenseContent
License Content

LicenseSimple
License Simple

LinkWithType
Hypermedia Link with Type

Location
Location

MarketplaceAccount
Marketplace Account

MarketplaceListingPlan
Marketplace Listing Plan

MarketplacePurchase
Marketplace Purchase

Migration
A migration.

MigrationId
Migration Id

Milestone
A collection of related issues and pull requests.

MilestoneNumber
Milestone Number

MinimalRepository
Minimal Repository

Order
Order

Org
Org

OrgHook
Org Hook

OrgMembership
Org Membership

Organization
GitHub account for managing multiple users, teams, and repositories

OrganizationActionsSecret
Secrets for GitHub Actions for an organization.

OrganizationFull
Organization Full

OrganizationInvitation
Organization Invitation

OrganizationSimple
Organization Simple

Owner
Owner

PackagesBillingUsage
Packages Billing Usage

PageBuild
Page Build

PageBuildStatus
Page Build Status

PagesSourceHash
Pages Source Hash

ParamActor
Param Actor

ParamEvent
Param Event

ParamPage
Param Page

ParamStatus
Param Status

Participating
Participating

ParticipationStats
Participation Stats

Per
Per

PerPage
Per Page

PlanId
Plan Id

PorterAuthor
Porter Author

PorterLargeFile
Porter Large File

PreviewHeaderMissing
Preview Header Missing

PrivateUser
Private User

Project
Projects are a way to organize columns and cards of work.

ProjectCard
Project cards represent a scope of work.

ProjectColumn
Project columns contain cards of work.

ProjectId
Project Id

ProtectedBranch
Branch protections protect branches

ProtectedBranchAdminEnforced
Protected Branch Admin Enforced

ProtectedBranchPullRequestReview
Protected Branch Pull Request Review

PublicUser
Public User

PullNumber
Pull Number

PullRequest
Pull requests let you tell others about changes you've pushed to a repository on GitHub. Once a pull request is sent, interested parties can review the set of changes, discuss potential modifications, and even push follow-up commits if necessary.

PullRequestMergeResult
Pull Request Merge Result

PullRequestMinimal
Pull Request Minimal

PullRequestReview
Pull Request Reviews are reviews on pull requests.

PullRequestReviewComment
Pull Request Review Comments are comments on a portion of the Pull Request's diff.

PullRequestReviewRequest
Pull Request Review Request

PullRequestSimple
Pull Request Simple

RateLimit
Rate Limit

RateLimitOverview
Rate Limit Overview

Reaction
Reactions to conversations provide a way to help people express their feelings more simply and effectively.

ReactionId
Reaction Id

ReactionRollup
Reaction Rollup

ReferrerTraffic
Referrer Traffic

Release
A release.

ReleaseAsset
Data related to a release.

ReleaseId
Release Id

Repo
Repo

RepoName
Repo Name

RepoSearchResultItem
Repo Search Result Item

Repository
A git repository

RepositoryCollaboratorPermission
Repository Collaborator Permission

RepositoryInvitation
Repository invitations let you manage who you collaborate with.

RepositorySubscription
Repository invitations let you manage who you collaborate with.

ReviewComment
Legacy Review Comment

ReviewId
Review Id

RunId
Run Id

Runner
A self hosted runner

RunnerApplication
Runner Application

RunnerId
Runner Id

SchemaActor
Actor

SchemaEvent
Event

SchemaLink
Hypermedia Link

SchemaPage
The configuration for GitHub Pages for a repository.

SchemaStatus
The status of a commit.

ScimError
Scim Error

ScimUser
SCIM /Users provisioning endpoints

ScimUserId
Scim User Id

ScimUserList
SCIM User List

ScopedInstallation
Scoped Installation

SearchResultTextMatchesItem
Search Result Text Matches Item

SecretName
Secret Name

ServiceUnavailable
Service Unavailable

ShortBlob
Short Blob

ShortBranch
Short Branch

SimpleCommit
Simple Commit

SimpleCommitStatus
Simple Commit Status

SimpleUser
Simple User

Since
Since

Sort
Sort

Stargazer
Stargazer

StarredRepository
Starred Repository

StatusCheckPolicy
Status Check Policy

Tag
Tag

Team
Groups of organization members that gives permissions on specified repositories.

TeamDiscussion
A team discussion is a persistent record of a free-form conversation within a team.

TeamDiscussionComment
A reply to a discussion within a team.

TeamFull
Groups of organization members that gives permissions on specified repositories.

TeamId
Team Id

TeamMembership
Team Membership

TeamProject
A team's access to a project.

TeamRepository
A team's access to a repository.

TeamSimple
Groups of organization members that gives permissions on specified repositories.

TeamSlug
Team Slug

Thread
Thread

ThreadId
Thread Id

ThreadSubscription
Thread Subscription

Topic
A topic aggregates entities that are related to a subject.

TopicSearchResultItem
Topic Search Result Item

Traffic
Traffic

UserMarketplacePurchase
User Marketplace Purchase

UserSearchResultItem
User Search Result Item

Username
Username

ValidationError
Validation Error

ValidationErrorSimple
Validation Error Simple

Verification
Verification

ViewTraffic
View Traffic

Workflow
A GitHub Actions workflow

WorkflowId
Workflow Id

WorkflowRun
An invocation of a workflow

WorkflowRunBranch
Workflow Run Branch

WorkflowRunStatus
Workflow Run Status

WorkflowRunUsage
Workflow Run Usage

WorkflowUsage
Workflow Usage

XCommonMarkerVersion
X Common Marker Version

XRateLimitLimit
X Rate Limit Limit

XRateLimitRemaining
X Rate Limit Remaining

XRateLimitReset
X Rate Limit Reset