Struct octorest::types::PullRequestSchemaArg

pub struct PullRequestSchemaArg<'ser> { /* fields omitted */ }

Pull requests let you tell others about changes you've pushed to a repository on GitHub. Once a pull request is sent, interested parties can review the set of changes, discuss potential modifications, and even push follow-up commits if necessary.

Implementations

impl<'ser> PullRequestSchemaArg<'ser>

pub fn with_links(self, value: PullRequestLinksArg<'ser>) -> Self

pub fn with_active_lock_reason(self, value: &'ser str) -> Self

pub fn with_additions(self, value: i64) -> Self

pub fn with_assignee(self, value: PullRequestAssigneeArg) -> Self

pub fn with_assignees(self, value: &'ser [SimpleUserArg<'ser>]) -> Self

pub fn with_author_association(self, value: &'ser str) -> Self

pub fn with_base(self, value: PullRequestBaseArg<'ser>) -> Self

pub fn with_body(self, value: &'ser str) -> Self

pub fn with_changed_files(self, value: i64) -> Self

pub fn with_closed_at(self, value: SystemTime) -> Self

pub fn with_comments(self, value: i64) -> Self

pub fn with_comments_url(self, value: &'ser str) -> Self

pub fn with_commits(self, value: i64) -> Self

pub fn with_commits_url(self, value: &'ser str) -> Self

pub fn with_created_at(self, value: SystemTime) -> Self

pub fn with_deletions(self, value: i64) -> Self

pub fn with_diff_url(self, value: &'ser str) -> Self

pub fn with_draft(self, value: bool) -> Self

pub fn with_head(self, value: PullRequestHeadArg<'ser>) -> Self

pub fn with_html_url(self, value: &'ser str) -> Self

pub fn with_id(self, value: i64) -> Self

pub fn with_issue_url(self, value: &'ser str) -> Self

pub fn with_labels(self, value: &'ser [PullRequestLabelsItemArg<'ser>]) -> Self

pub fn with_locked(self, value: bool) -> Self

pub fn with_maintainer_can_modify(self, value: bool) -> Self

pub fn with_merge_commit_sha(self, value: &'ser str) -> Self

pub fn with_mergeable(self, value: bool) -> Self

pub fn with_mergeable_state(self, value: &'ser str) -> Self

pub fn with_merged(self, value: bool) -> Self

pub fn with_merged_at(self, value: SystemTime) -> Self

pub fn with_merged_by(self, value: MergedByArg) -> Self

pub fn with_milestone(self, value: PullRequestMilestoneArg) -> Self

pub fn with_node_id(self, value: &'ser str) -> Self

pub fn with_number(self, value: i64) -> Self

pub fn with_patch_url(self, value: &'ser str) -> Self

pub fn with_rebaseable(self, value: bool) -> Self

pub fn with_requested_reviewers(
self,
value: &'ser [SimpleUserArg<'ser>]
) -> Self

pub fn with_requested_teams(self, value: &'ser [TeamSimpleArg<'ser>]) -> Self

pub fn with_review_comment_url(self, value: &'ser str) -> Self

pub fn with_review_comments(self, value: i64) -> Self

pub fn with_review_comments_url(self, value: &'ser str) -> Self

pub fn with_state(self, value: PullRequestState) -> Self

pub fn with_statuses_url(self, value: &'ser str) -> Self

pub fn with_title(self, value: &'ser str) -> Self

pub fn with_updated_at(self, value: SystemTime) -> Self

pub fn with_url(self, value: &'ser str) -> Self

pub fn with_user(self, value: PullRequestUserArg) -> Self