Struct octorest::types::PullRequestSchemaResponse

pub struct PullRequestSchemaResponse<'de> { /* fields omitted */ }

Pull requests let you tell others about changes you've pushed to a repository on GitHub. Once a pull request is sent, interested parties can review the set of changes, discuss potential modifications, and even push follow-up commits if necessary.

Implementations

impl<'de> PullRequestSchemaResponse<'de>

pub fn links(&self) -> &PullRequestLinksResponse<'de>

pub fn active_lock_reason(&self) -> &'de str

pub fn additions(&self) -> i64

pub fn assignee(&self) -> &PullRequestAssigneeResponse

pub fn assignees(&self) -> &Vec<SimpleUserResponse<'de>>

pub fn author_association(&self) -> &'de str

pub fn base(&self) -> &PullRequestBaseResponse<'de>

pub fn body(&self) -> &'de str

pub fn changed_files(&self) -> i64

pub fn closed_at(&self) -> DateTime<Utc>

pub fn comments(&self) -> i64

pub fn comments_url(&self) -> &'de str

pub fn commits(&self) -> i64

pub fn commits_url(&self) -> &'de str

pub fn created_at(&self) -> DateTime<Utc>

pub fn deletions(&self) -> i64

pub fn diff_url(&self) -> &'de str

pub fn draft(&self) -> bool

pub fn head(&self) -> &PullRequestHeadResponse<'de>

pub fn html_url(&self) -> &'de str

pub fn id(&self) -> i64

pub fn issue_url(&self) -> &'de str

pub fn labels(&self) -> &Vec<PullRequestLabelsItemResponse<'de>>

pub fn locked(&self) -> bool

pub fn maintainer_can_modify(&self) -> bool

pub fn merge_commit_sha(&self) -> &'de str

pub fn mergeable(&self) -> bool

pub fn mergeable_state(&self) -> &'de str

pub fn merged(&self) -> bool

pub fn merged_at(&self) -> DateTime<Utc>

pub fn merged_by(&self) -> &MergedByResponse

pub fn milestone(&self) -> &PullRequestMilestoneResponse

pub fn node_id(&self) -> &'de str

pub fn number(&self) -> i64

pub fn patch_url(&self) -> &'de str

pub fn rebaseable(&self) -> bool

pub fn requested_reviewers(&self) -> &Vec<SimpleUserResponse<'de>>

pub fn requested_teams(&self) -> &Vec<TeamSimpleResponse<'de>>

pub fn review_comment_url(&self) -> &'de str

pub fn review_comments(&self) -> i64

pub fn review_comments_url(&self) -> &'de str

pub fn state(&self) -> PullRequestState

pub fn statuses_url(&self) -> &'de str

pub fn title(&self) -> &'de str

pub fn updated_at(&self) -> DateTime<Utc>

pub fn url(&self) -> &'de str

pub fn user(&self) -> &PullRequestUserResponse