Module octorest::types

Miscellaneous data types used in the GitHub API.

The API is designed such that users shall not need to explicitly import types from this module.

Structs

AccountArg
Account

AccountResponse
Account

ActionsBillingUsageArg
Actions Billing Usage

ActionsBillingUsageResponse
Actions Billing Usage

ActionsPublicKeyArg
The public key used for setting Actions Secrets.

ActionsPublicKeyResponse
The public key used for setting Actions Secrets.

ActionsSecretArg
Set secrets for GitHub Actions.

ActionsSecretResponse
Set secrets for GitHub Actions.

ActorSchemaArg
Actor

ActorSchemaResponse
Actor

AliasesItemArg
Aliases Item

AliasesItemResponse
Aliases Item

AliasesItemTopicRelationArg
Aliases Item Topic Relation

AliasesItemTopicRelationResponse
Aliases Item Topic Relation

ApiOverviewArg
Api Overview

ApiOverviewResponse
Api Overview

ApplicationGrantAppArg
Application Grant App

ApplicationGrantAppResponse
Application Grant App

ApplicationGrantArg
The authorization associated with an OAuth Access.

ApplicationGrantResponse
The authorization associated with an OAuth Access.

ApplicationGrantUserArg
Application Grant User

ApplicationGrantUserResponse
Application Grant User

AppsItemArg
Apps Item

AppsItemOwnerArg
Apps Item Owner

AppsItemOwnerResponse
Apps Item Owner

AppsItemPermissionsArg
Apps Item Permissions

AppsItemPermissionsResponse
Apps Item Permissions

AppsItemResponse
Apps Item

ArtifactArg
An artifact

ArtifactResponse
An artifact

AssignerArg
Assigner

AssignerResponse
Assigner

AuthenticationTokenArg
Authentication Token

AuthenticationTokenPermissionsArg
Authentication Token Permissions

AuthenticationTokenPermissionsResponse
Authentication Token Permissions

AuthenticationTokenResponse
Authentication Token

AuthorizationAppArg
Authorization App

AuthorizationAppResponse
Authorization App

AuthorizationArg
The authorization for an OAuth app, GitHub App, or a Personal Access Token.

AuthorizationInstallationArg
Authorization Installation

AuthorizationInstallationResponse
Authorization Installation

AuthorizationResponse
The authorization for an OAuth app, GitHub App, or a Personal Access Token.

AuthorizationUserArg
Authorization User

AuthorizationUserResponse
Authorization User

AutoTriggerChecksItemArg
Auto Trigger Checks Item

AutoTriggerChecksItemResponse
Auto Trigger Checks Item

BaseGistArg
Base Gist

BaseGistFilesArg
Base Gist Files

BaseGistFilesResponse
Base Gist Files

BaseGistOwnerArg
Base Gist Owner

BaseGistOwnerResponse
Base Gist Owner

BaseGistResponse
Base Gist

BaseGistUserArg
Base Gist User

BaseGistUserResponse
Base Gist User

BaseRepoLicenseArg
Base Repo License

BaseRepoLicenseResponse
Base Repo License

BaseRepoOwnerArg
Base Repo Owner

BaseRepoOwnerResponse
Base Repo Owner

BaseRepoPermissionsArg
Base Repo Permissions

BaseRepoPermissionsResponse
Base Repo Permissions

BasicErrorArg
Basic Error

BasicErrorResponse
Basic Error

BlobArg
Blob

BlobResponse
Blob

BlockArg
Block

BlockResponse
Block

BranchProtectionAllowDeletionsArg
Branch Protection Allow Deletions

BranchProtectionAllowDeletionsResponse
Branch Protection Allow Deletions

BranchProtectionAllowForcePushesArg
Branch Protection Allow Force Pushes

BranchProtectionAllowForcePushesResponse
Branch Protection Allow Force Pushes

BranchProtectionArg
Branch Protection

BranchProtectionRequiredLinearHistoryArg
Branch Protection Required Linear History

BranchProtectionRequiredLinearHistoryResponse
Branch Protection Required Linear History

BranchProtectionResponse
Branch Protection

BranchRestrictionPolicyArg
Branch Restriction Policy

BranchRestrictionPolicyResponse
Branch Restriction Policy

BranchRestrictionPolicyTeamsItemArg
Branch Restriction Policy Teams Item

BranchRestrictionPolicyTeamsItemResponse
Branch Restriction Policy Teams Item

BranchRestrictionPolicyUsersItemArg
Branch Restriction Policy Users Item

BranchRestrictionPolicyUsersItemResponse
Branch Restriction Policy Users Item

BranchShortArg
Branch Short

BranchShortCommitArg
Branch Short Commit

BranchShortCommitResponse
Branch Short Commit

BranchShortResponse
Branch Short

BranchWithProtectionArg
Branch With Protection

BranchWithProtectionLinksArg
Branch With Protection Links

BranchWithProtectionLinksResponse
Branch With Protection Links

BranchWithProtectionResponse
Branch With Protection

ChangeStatusArg
Change Status

ChangeStatusResponse
Change Status

CheckAnnotationArg
Check Annotation

CheckAnnotationResponse
Check Annotation

CheckRunAppArg
Check Run App

CheckRunAppResponse
Check Run App

CheckRunArg
A check performed on the code of a given code change

CheckRunCheckSuiteArg
Check Run Check Suite

CheckRunCheckSuiteResponse
Check Run Check Suite

CheckRunResponse
A check performed on the code of a given code change

CheckSuiteAppArg
Check Suite App

CheckSuiteAppResponse
Check Suite App

CheckSuitePreferenceArg
Check suite configuration preferences for a repository.

CheckSuitePreferenceResponse
Check suite configuration preferences for a repository.

CheckSuiteSchemaArg
A suite of checks performed on the code of a given code change

CheckSuiteSchemaResponse
A suite of checks performed on the code of a given code change

CloneTrafficArg
Clone Traffic

CloneTrafficResponse
Clone Traffic

CodeOfConductSchemaArg
Code Of Conduct

CodeOfConductSchemaResponse
Code Of Conduct

CodeOfConductSimpleArg
Code of Conduct Simple

CodeOfConductSimpleResponse
Code of Conduct Simple

CodeScanningAlertArg
Code Scanning Alert

CodeScanningAlertClosedByArg
Code Scanning Alert Closed By

CodeScanningAlertClosedByResponse
Code Scanning Alert Closed By

CodeScanningAlertResponse
Code Scanning Alert

CodeSearchResultItemArg
Code Search Result Item

CodeSearchResultItemResponse
Code Search Result Item

CollaboratorArg
Collaborator

CollaboratorPermissionsArg
Collaborator Permissions

CollaboratorPermissionsResponse
Collaborator Permissions

CollaboratorResponse
Collaborator

CombinedBillingUsageArg
Combined Billing Usage

CombinedBillingUsageResponse
Combined Billing Usage

CombinedCommitStatusArg
Combined Commit Status

CombinedCommitStatusResponse
Combined Commit Status

CommitActivityArg
Commit Activity

CommitActivityResponse
Commit Activity

CommitAuthorSchemaArg
Commit Author Schema

CommitAuthorSchemaResponse
Commit Author Schema

CommitCommentArg
Commit Comment

CommitCommentResponse
Commit Comment

CommitCommentUserArg
Commit Comment User

CommitCommentUserResponse
Commit Comment User

CommitCommitArg
Commit Commit

CommitCommitAuthorArg
Commit Commit Author

CommitCommitAuthorResponse
Commit Commit Author

CommitCommitCommitterArg
Commit Commit Committer

CommitCommitCommitterResponse
Commit Commit Committer

CommitCommitResponse
Commit Commit

CommitCommitTreeArg
Commit Commit Tree

CommitCommitTreeResponse
Commit Commit Tree

CommitCommitterSchemaArg
Commit Committer Schema

CommitCommitterSchemaResponse
Commit Committer Schema

CommitComparisonArg
Commit Comparison

CommitComparisonResponse
Commit Comparison

CommitParentsItemSchemaArg
Commit Parents Item Schema

CommitParentsItemSchemaResponse
Commit Parents Item Schema

CommitSchemaArg
Commit

CommitSchemaResponse
Commit

CommitSearchResultItemArg
Commit Search Result Item

CommitSearchResultItemAuthorArg
Commit Search Result Item Author

CommitSearchResultItemAuthorResponse
Commit Search Result Item Author

CommitSearchResultItemCommitArg
Commit Search Result Item Commit

CommitSearchResultItemCommitAuthorArg
Commit Search Result Item Commit Author

CommitSearchResultItemCommitAuthorResponse
Commit Search Result Item Commit Author

CommitSearchResultItemCommitCommitterArg
Commit Search Result Item Commit Committer

CommitSearchResultItemCommitCommitterResponse
Commit Search Result Item Commit Committer

CommitSearchResultItemCommitResponse
Commit Search Result Item Commit

CommitSearchResultItemCommitTreeArg
Commit Search Result Item Commit Tree

CommitSearchResultItemCommitTreeResponse
Commit Search Result Item Commit Tree

CommitSearchResultItemCommitterArg
Commit Search Result Item Committer

CommitSearchResultItemCommitterResponse
Commit Search Result Item Committer

CommitSearchResultItemParentsItemArg
Commit Search Result Item Parents Item

CommitSearchResultItemParentsItemResponse
Commit Search Result Item Parents Item

CommitSearchResultItemResponse
Commit Search Result Item

CommitVerificationArg
Commit Verification

CommitVerificationResponse
Commit Verification

CommunityHealthFileArg
Community Health File

CommunityHealthFileResponse
Community Health File

CommunityProfileArg
Community Profile

CommunityProfileFilesArg
Community Profile Files

CommunityProfileFilesResponse
Community Profile Files

CommunityProfileResponse
Community Profile

ContentDirectoryItemArg
Content Directory Item

ContentDirectoryItemLinksArg
Content Directory Item Links

ContentDirectoryItemLinksResponse
Content Directory Item Links

ContentDirectoryItemResponse
Content Directory Item

ContentFileArg
Content File

ContentFileLinksArg
Content File Links

ContentFileLinksResponse
Content File Links

ContentFileResponse
Content File

ContentLinksArg
Content Links

ContentLinksResponse
Content Links

ContentReferenceAttachmentArg
Content Reference attachments allow you to provide context around URLs posted in comments

ContentReferenceAttachmentResponse
Content Reference attachments allow you to provide context around URLs posted in comments

ContentSubmoduleArg
An object describing a symlink

ContentSubmoduleLinksArg
Content Submodule Links

ContentSubmoduleLinksResponse
Content Submodule Links

ContentSubmoduleResponse
An object describing a symlink

ContentSymlinkArg
An object describing a symlink

ContentSymlinkLinksArg
Content Symlink Links

ContentSymlinkLinksResponse
Content Symlink Links

ContentSymlinkResponse
An object describing a symlink

ContentTrafficArg
Content Traffic

ContentTrafficResponse
Content Traffic

ContentTreeArg
Content Tree

ContentTreeLinksArg
Content Tree Links

ContentTreeLinksResponse
Content Tree Links

ContentTreeResponse
Content Tree

ContextsItemArg
Contexts Item

ContextsItemResponse
Contexts Item

ContributingArg
Contributing

ContributingResponse
Contributing

ContributorActivityArg
Contributor Activity

ContributorActivityAuthorArg
Contributor Activity Author

ContributorActivityAuthorResponse
Contributor Activity Author

ContributorActivityResponse
Contributor Activity

ContributorArg
Contributor

ContributorResponse
Contributor

CredentialAuthorizationArg
Credential Authorization

CredentialAuthorizationResponse
Credential Authorization

DeployKeyArg
An SSH key granting access to a single repository.

DeployKeyResponse
An SSH key granting access to a single repository.

DeploymentArg
A request for a specific ref(branch,sha,tag) to be deployed

DeploymentCreatorArg
Deployment Creator

DeploymentCreatorResponse
Deployment Creator

DeploymentPayloadArg
Deployment Payload

DeploymentPayloadResponse
Deployment Payload

DeploymentPerformedViaGithubAppArg
Deployment Performed Via Github App

DeploymentPerformedViaGithubAppResponse
Deployment Performed Via Github App

DeploymentResponse
A request for a specific ref(branch,sha,tag) to be deployed

DeploymentStatusArg
The status of a deployment.

DeploymentStatusCreatorArg
Deployment Status Creator

DeploymentStatusCreatorResponse
Deployment Status Creator

DeploymentStatusPerformedViaGithubAppArg
Deployment Status Performed Via Github App

DeploymentStatusPerformedViaGithubAppResponse
Deployment Status Performed Via Github App

DeploymentStatusResponse
The status of a deployment.

DiffEntryArg
Diff Entry

DiffEntryResponse
Diff Entry

EmailArg
Email

EmailResponse
Email

EnforceAdminsArg
Enforce Admins

EnforceAdminsResponse
Enforce Admins

EnterpriseArg
An enterprise account

EnterpriseResponse
An enterprise account

EntriesItemArg
Entries Item

EntriesItemLinksArg
Entries Item Links

EntriesItemLinksResponse
Entries Item Links

EntriesItemResponse
Entries Item

ErrorArg
Error

ErrorResponse
Error

ErrorsItemArg
Errors Item

ErrorsItemResponse
Errors Item

ErrorsItemValueArg
Errors Item Value

ErrorsItemValueResponse
Errors Item Value

EventArg
Event

EventPayloadArg
Event Payload

EventPayloadResponse
Event Payload

EventRepoArg
Event Repo

EventRepoResponse
Event Repo

EventResponse
Event

ExcludeItemArg
Exclude Item

ExcludeItemResponse
Exclude Item

FeedArg
Feed

FeedLinksArg
Feed Links

FeedLinksResponse
Feed Links

FeedResponse
Feed

FileCommitArg
File Commit

FileCommitCommitArg
File Commit Commit

FileCommitCommitAuthorArg
File Commit Commit Author

FileCommitCommitAuthorResponse
File Commit Commit Author

FileCommitCommitCommitterArg
File Commit Commit Committer

FileCommitCommitCommitterResponse
File Commit Commit Committer

FileCommitCommitParentsItemArg
File Commit Commit Parents Item

FileCommitCommitParentsItemResponse
File Commit Commit Parents Item

FileCommitCommitResponse
File Commit Commit

FileCommitCommitTreeArg
File Commit Commit Tree

FileCommitCommitTreeResponse
File Commit Commit Tree

FileCommitContentArg
File Commit Content

FileCommitContentResponse
File Commit Content

FileCommitResponse
File Commit

FilesCodeOfConductArg
Files Code Of Conduct

FilesCodeOfConductResponse
Files Code Of Conduct

FilesItemArg
Files Item

FilesItemResponse
Files Item

FilesLicenseArg
Files License

FilesLicenseResponse
Files License

ForbiddenGistArg
Forbidden Gist

ForbiddenGistResponse
Forbidden Gist

ForksItemArg
Forks Item

ForksItemResponse
Forks Item

FullRepositoryArg
Full Repository

FullRepositoryLicenseArg
Full Repository License

FullRepositoryLicenseResponse
Full Repository License

FullRepositoryOrganizationArg
Full Repository Organization

FullRepositoryOrganizationResponse
Full Repository Organization

FullRepositoryOwnerArg
Full Repository Owner

FullRepositoryOwnerResponse
Full Repository Owner

FullRepositoryPermissionsArg
Full Repository Permissions

FullRepositoryPermissionsResponse
Full Repository Permissions

FullRepositoryResponse
Full Repository

FullRepositoryTemplateRepositoryArg
Full Repository Template Repository

FullRepositoryTemplateRepositoryResponse
Full Repository Template Repository

GistCommentArg
A comment made to a gist.

GistCommentResponse
A comment made to a gist.

GistCommentUserArg
Gist Comment User

GistCommentUserResponse
Gist Comment User

GistCommitArg
Gist Commit

GistCommitResponse
Gist Commit

GistCommitUserArg
Gist Commit User

GistCommitUserResponse
Gist Commit User

GistFullArg
Gist Full

GistFullResponse
Gist Full

GistSimpleArg
Gist Simple

GistSimpleFilesArg
Gist Simple Files

GistSimpleFilesResponse
Gist Simple Files

GistSimpleResponse
Gist Simple

GitCommitArg
Low-level Git commit operations within a repository

GitCommitAuthorArg
Identifying information for the git-user

GitCommitAuthorResponse
Identifying information for the git-user

GitCommitCommitterArg
Identifying information for the git-user

GitCommitCommitterResponse
Identifying information for the git-user

GitCommitParentsItemArg
Git Commit Parents Item

GitCommitParentsItemResponse
Git Commit Parents Item

GitCommitResponse
Low-level Git commit operations within a repository

GitCommitTreeArg
Git Commit Tree

GitCommitTreeResponse
Git Commit Tree

GitCommitVerificationArg
Git Commit Verification

GitCommitVerificationResponse
Git Commit Verification

GitRefArg
Git references within a repository

GitRefObjectArg
Git Ref Object

GitRefObjectResponse
Git Ref Object

GitRefResponse
Git references within a repository

GitTagArg
Metadata for a Git tag

GitTagObjectArg
Git Tag Object

GitTagObjectResponse
Git Tag Object

GitTagResponse
Metadata for a Git tag

GitTreeArg
The hierarchy between files in a Git repository.

GitTreeResponse
The hierarchy between files in a Git repository.

GitUserArg
Metaproperties for Git author/committer information.

GitUserResponse
Metaproperties for Git author/committer information.

GitignoreTemplateArg
Gitignore Template

GitignoreTemplateResponse
Gitignore Template

GpgKeyArg
A unique encryption key

GpgKeyEmailsItemArg
Gpg Key Emails Item

GpgKeyEmailsItemResponse
Gpg Key Emails Item

GpgKeyResponse
A unique encryption key

GpgKeySubkeysItemArg
Gpg Key Subkeys Item

GpgKeySubkeysItemResponse
Gpg Key Subkeys Item

GroupMappingArg
External Groups to be mapped to a team for membership

GroupMappingGroupsItemArg
Group Mapping Groups Item

GroupMappingGroupsItemResponse
Group Mapping Groups Item

GroupMappingResponse
External Groups to be mapped to a team for membership

HeadRepoLicenseArg
Head Repo License

HeadRepoLicenseResponse
Head Repo License

HeadRepoOwnerArg
Head Repo Owner

HeadRepoOwnerResponse
Head Repo Owner

HeadRepoPermissionsArg
Head Repo Permissions

HeadRepoPermissionsResponse
Head Repo Permissions

HistoryItemArg
History Item

HistoryItemResponse
History Item

HookArg
Webhooks for repositories.

HookConfigArg
Hook Config

HookConfigResponse
Hook Config

HookResponse
Webhooks for repositories.

HookResponseArg
Hook Response

HookResponseResponse
Hook Response

HovercardArg
Hovercard

HovercardResponse
Hovercard

ImportArg
A repository import from an external source.

ImportResponse
A repository import from an external source.

InstallationPermissionsArg
Installation Permissions

InstallationPermissionsResponse
Installation Permissions

InstallationSchemaArg
Installation

InstallationSchemaResponse
Installation

InstallationTokenArg
Authentication token for a GitHub App installed on a user or org.

InstallationTokenPermissionsArg
Installation Token Permissions

InstallationTokenPermissionsResponse
Installation Token Permissions

InstallationTokenResponse
Authentication token for a GitHub App installed on a user or org.

IntegrationArg
GitHub apps are a new way to extend GitHub. They can be installed directly on organizations and user accounts and granted access to specific repositories. They come with granular permissions and built-in webhooks. GitHub apps are first class actors within GitHub.

IntegrationOwnerArg
Integration Owner

IntegrationOwnerResponse
Integration Owner

IntegrationPermissionsArg
The set of permissions for the GitHub app

IntegrationPermissionsResponse
The set of permissions for the GitHub app

IntegrationResponse
GitHub apps are a new way to extend GitHub. They can be installed directly on organizations and user accounts and granted access to specific repositories. They come with granular permissions and built-in webhooks. GitHub apps are first class actors within GitHub.

InteractionLimitArg
Interaction limit settings.

InteractionLimitResponse
Interaction limit settings.

InviteeArg
Invitee

InviteeResponse
Invitee

InviterArg
Inviter

InviterResponse
Inviter

IssueArg
Issues are a great way to keep track of tasks, enhancements, and bugs for your projects.

IssueAssigneeArg
Issue Assignee

IssueAssigneeResponse
Issue Assignee

IssueClosedByArg
Issue Closed By

IssueClosedByResponse
Issue Closed By

IssueCommentArg
Comments provide a way for people to collaborate on an issue.

IssueCommentPerformedViaGithubAppArg
Issue Comment Performed Via Github App

IssueCommentPerformedViaGithubAppResponse
Issue Comment Performed Via Github App

IssueCommentResponse
Comments provide a way for people to collaborate on an issue.

IssueCommentUserArg
Issue Comment User

IssueCommentUserResponse
Issue Comment User

IssueEventActorArg
Issue Event Actor

IssueEventActorResponse
Issue Event Actor

IssueEventArg
Issue Event

IssueEventAssigneeArg
Issue Event Assignee

IssueEventAssigneeResponse
Issue Event Assignee

IssueEventDismissedReviewArg
Issue Event Dismissed Review

IssueEventDismissedReviewResponse
Issue Event Dismissed Review

IssueEventForIssueArg
Issue Event for Issue

IssueEventForIssueResponse
Issue Event for Issue

IssueEventLabelArg
Issue Event Label

IssueEventLabelResponse
Issue Event Label

IssueEventMilestoneArg
Issue Event Milestone

IssueEventMilestoneResponse
Issue Event Milestone

IssueEventProjectCardArg
Issue Event Project Card

IssueEventProjectCardResponse
Issue Event Project Card

IssueEventRenameArg
Issue Event Rename

IssueEventRenameResponse
Issue Event Rename

IssueEventResponse
Issue Event

IssueLabelsItemArg
Issue Labels Item

IssueLabelsItemResponse
Issue Labels Item

IssueMilestoneArg
Issue Milestone

IssueMilestoneResponse
Issue Milestone

IssuePerformedViaGithubAppArg
Issue Performed Via Github App

IssuePerformedViaGithubAppResponse
Issue Performed Via Github App

IssuePullRequestArg
Issue Pull Request

IssuePullRequestResponse
Issue Pull Request

IssueResponse
Issues are a great way to keep track of tasks, enhancements, and bugs for your projects.

IssueSearchResultItemArg
Issue Search Result Item

IssueSearchResultItemAssigneeArg
Issue Search Result Item Assignee

IssueSearchResultItemAssigneeResponse
Issue Search Result Item Assignee

IssueSearchResultItemLabelsItemArg
Issue Search Result Item Labels Item

IssueSearchResultItemLabelsItemResponse
Issue Search Result Item Labels Item

IssueSearchResultItemMilestoneArg
Issue Search Result Item Milestone

IssueSearchResultItemMilestoneResponse
Issue Search Result Item Milestone

IssueSearchResultItemPerformedViaGithubAppArg
Issue Search Result Item Performed Via Github App

IssueSearchResultItemPerformedViaGithubAppResponse
Issue Search Result Item Performed Via Github App

IssueSearchResultItemPullRequestArg
Issue Search Result Item Pull Request

IssueSearchResultItemPullRequestResponse
Issue Search Result Item Pull Request

IssueSearchResultItemResponse
Issue Search Result Item

IssueSearchResultItemUserArg
Issue Search Result Item User

IssueSearchResultItemUserResponse
Issue Search Result Item User

IssueSimpleArg
Issue Simple

IssueSimpleAssigneeArg
Issue Simple Assignee

IssueSimpleAssigneeResponse
Issue Simple Assignee

IssueSimpleLabelsItemArg
Issue Simple Labels Item

IssueSimpleLabelsItemResponse
Issue Simple Labels Item

IssueSimpleMilestoneArg
Issue Simple Milestone

IssueSimpleMilestoneResponse
Issue Simple Milestone

IssueSimplePerformedViaGithubAppArg
Issue Simple Performed Via Github App

IssueSimplePerformedViaGithubAppResponse
Issue Simple Performed Via Github App

IssueSimplePullRequestArg
Issue Simple Pull Request

IssueSimplePullRequestResponse
Issue Simple Pull Request

IssueSimpleResponse
Issue Simple

IssueSimpleUserArg
Issue Simple User

IssueSimpleUserResponse
Issue Simple User

IssueTemplateArg
Issue Template

IssueTemplateResponse
Issue Template

IssueUserArg
Issue User

IssueUserResponse
Issue User

JobArg
Information of a job execution in a workflow run

JobResponse
Information of a job execution in a workflow run

KeyArg
Key

KeyResponse
Key

KeySimpleArg
Key Simple

KeySimpleResponse
Key Simple

LabelArg
Color-coded labels help you categorize and filter your issues (just like labels in Gmail).

LabelResponse
Color-coded labels help you categorize and filter your issues (just like labels in Gmail).

LabelSearchResultItemArg
Label Search Result Item

LabelSearchResultItemResponse
Label Search Result Item

LanguageArg
Language

LanguageResponse
Language

LicenseContentArg
License Content

LicenseContentLicenseArg
License Content License

LicenseContentLicenseResponse
License Content License

LicenseContentLinksArg
License Content Links

LicenseContentLinksResponse
License Content Links

LicenseContentResponse
License Content

LicenseSchemaArg
License

LicenseSchemaResponse
License

LicenseSimpleArg
License Simple

LicenseSimpleResponse
License Simple

LinkArg
Hypermedia Link

LinkResponse
Hypermedia Link

LinkWithTypeArg
Hypermedia Link with Type

LinkWithTypeResponse
Hypermedia Link with Type

MarketplaceAccountArg
Marketplace Account

MarketplaceAccountResponse
Marketplace Account

MarketplaceListingPlanArg
Marketplace Listing Plan

MarketplaceListingPlanResponse
Marketplace Listing Plan

MarketplacePendingChangeArg
Marketplace Pending Change

MarketplacePendingChangeResponse
Marketplace Pending Change

MarketplacePurchaseMarketplacePurchaseArg
Marketplace Purchase Marketplace Purchase

MarketplacePurchaseMarketplacePurchaseResponse
Marketplace Purchase Marketplace Purchase

MarketplacePurchaseSchemaArg
Marketplace Purchase

MarketplacePurchaseSchemaResponse
Marketplace Purchase

MatchesItemArg
Matches Item

MatchesItemResponse
Matches Item

MergedByArg
Merged By

MergedByResponse
Merged By

MetaArg
Meta

MetaResponse
Meta

MigrationArg
A migration.

MigrationOwnerArg
Migration Owner

MigrationOwnerResponse
Migration Owner

MigrationResponse
A migration.

MilestoneCreatorArg
Milestone Creator

MilestoneCreatorResponse
Milestone Creator

MilestoneSchemaArg
A collection of related issues and pull requests.

MilestoneSchemaResponse
A collection of related issues and pull requests.

MinimalRepositoryArg
Minimal Repository

MinimalRepositoryLicenseArg
Minimal Repository License

MinimalRepositoryLicenseResponse
Minimal Repository License

MinimalRepositoryOwnerArg
Minimal Repository Owner

MinimalRepositoryOwnerResponse
Minimal Repository Owner

MinimalRepositoryPermissionsArg
Minimal Repository Permissions

MinimalRepositoryPermissionsResponse
Minimal Repository Permissions

MinimalRepositoryResponse
Minimal Repository

MinutesUsedBreakdownArg
Minutes Used Breakdown

MinutesUsedBreakdownResponse
Minutes Used Breakdown

NameArg
Name

NameResponse
Name

OperationsItemArg
Operations Item

OperationsItemResponse
Operations Item

OperationsItemValueArg
Operations Item Value

OperationsItemValueResponse
Operations Item Value

OrgHookArg
Org Hook

OrgHookConfigArg
Org Hook Config

OrgHookConfigResponse
Org Hook Config

OrgHookResponse
Org Hook

OrgMembershipArg
Org Membership

OrgMembershipPermissionsArg
Org Membership Permissions

OrgMembershipPermissionsResponse
Org Membership Permissions

OrgMembershipResponse
Org Membership

OrgMembershipUserArg
Org Membership User

OrgMembershipUserResponse
Org Membership User

OrganizationActionsSecretArg
Secrets for GitHub Actions for an organization.

OrganizationActionsSecretResponse
Secrets for GitHub Actions for an organization.

OrganizationFullArg
Organization Full

OrganizationFullPlanArg
Organization Full Plan

OrganizationFullPlanResponse
Organization Full Plan

OrganizationFullResponse
Organization Full

OrganizationInvitationArg
Organization Invitation

OrganizationInvitationResponse
Organization Invitation

OrganizationPlanArg
Organization Plan

OrganizationPlanResponse
Organization Plan

OrganizationSchemaArg
GitHub account for managing multiple users, teams, and repositories

OrganizationSchemaResponse
GitHub account for managing multiple users, teams, and repositories

OrganizationSimpleArg
Organization Simple

OrganizationSimpleResponse
Organization Simple

OutputArg
Output

OutputResponse
Output

PackagesBillingUsageArg
Packages Billing Usage

PackagesBillingUsageResponse
Packages Billing Usage

PageArg
The configuration for GitHub Pages for a repository.

PageBuildArg
Page Build

PageBuildResponse
Page Build

PageBuildStatusArg
Page Build Status

PageBuildStatusResponse
Page Build Status

PageResponse
The configuration for GitHub Pages for a repository.

PagesItemArg
Pages Item

PagesItemResponse
Pages Item

PagesSourceHashArg
Pages Source Hash

PagesSourceHashResponse
Pages Source Hash

ParticipationStatsArg
Participation Stats

ParticipationStatsResponse
Participation Stats

PorterAuthorArg
Porter Author

PorterAuthorResponse
Porter Author

PorterLargeFileArg
Porter Large File

PorterLargeFileResponse
Porter Large File

PreferencesArg
Preferences

PreferencesResponse
Preferences

PreviewHeaderMissingArg
Preview Header Missing

PreviewHeaderMissingResponse
Preview Header Missing

PrivateUserArg
Private User

PrivateUserPlanArg
Private User Plan

PrivateUserPlanResponse
Private User Plan

PrivateUserResponse
Private User

ProjectArg
Projects are a way to organize columns and cards of work.

ProjectCardArg
Project cards represent a scope of work.

ProjectCardCreatorArg
Project Card Creator

ProjectCardCreatorResponse
Project Card Creator

ProjectCardResponse
Project cards represent a scope of work.

ProjectChoicesItemArg
Project Choices Item

ProjectChoicesItemResponse
Project Choices Item

ProjectColumnArg
Project columns contain cards of work.

ProjectColumnResponse
Project columns contain cards of work.

ProjectCreatorArg
Project Creator

ProjectCreatorResponse
Project Creator

ProjectPermissionsArg
Project Permissions

ProjectPermissionsResponse
Project Permissions

ProjectResponse
Projects are a way to organize columns and cards of work.

ProtectedBranchAdminEnforcedArg
Protected Branch Admin Enforced

ProtectedBranchAdminEnforcedResponse
Protected Branch Admin Enforced

ProtectedBranchAllowDeletionsArg
Protected Branch Allow Deletions

ProtectedBranchAllowDeletionsResponse
Protected Branch Allow Deletions

ProtectedBranchAllowForcePushesArg
Protected Branch Allow Force Pushes

ProtectedBranchAllowForcePushesResponse
Protected Branch Allow Force Pushes

ProtectedBranchArg
Branch protections protect branches

ProtectedBranchPullRequestReviewArg
Protected Branch Pull Request Review

ProtectedBranchPullRequestReviewDismissalRestrictionsArg
Protected Branch Pull Request Review Dismissal Restrictions

ProtectedBranchPullRequestReviewDismissalRestrictionsResponse
Protected Branch Pull Request Review Dismissal Restrictions

ProtectedBranchPullRequestReviewResponse
Protected Branch Pull Request Review

ProtectedBranchRequiredLinearHistoryArg
Protected Branch Required Linear History

ProtectedBranchRequiredLinearHistoryResponse
Protected Branch Required Linear History

ProtectedBranchResponse
Branch protections protect branches

PublicUserArg
Public User

PublicUserPlanArg
Public User Plan

PublicUserPlanResponse
Public User Plan

PublicUserResponse
Public User

PullRequestAssigneeArg
Pull Request Assignee

PullRequestAssigneeResponse
Pull Request Assignee

PullRequestBaseArg
Pull Request Base

PullRequestBaseRepoArg
Pull Request Base Repo

PullRequestBaseRepoResponse
Pull Request Base Repo

PullRequestBaseResponse
Pull Request Base

PullRequestBaseUserArg
Pull Request Base User

PullRequestBaseUserResponse
Pull Request Base User

PullRequestHeadArg
Pull Request Head

PullRequestHeadRepoArg
Pull Request Head Repo

PullRequestHeadRepoResponse
Pull Request Head Repo

PullRequestHeadResponse
Pull Request Head

PullRequestHeadUserArg
Pull Request Head User

PullRequestHeadUserResponse
Pull Request Head User

PullRequestLabelsItemArg
Pull Request Labels Item

PullRequestLabelsItemResponse
Pull Request Labels Item

PullRequestLinksArg
Pull Request Links

PullRequestLinksResponse
Pull Request Links

PullRequestMergeResultArg
Pull Request Merge Result

PullRequestMergeResultResponse
Pull Request Merge Result

PullRequestMilestoneArg
Pull Request Milestone

PullRequestMilestoneResponse
Pull Request Milestone

PullRequestMinimalArg
Pull Request Minimal

PullRequestMinimalBaseArg
Pull Request Minimal Base

PullRequestMinimalBaseRepoArg
Pull Request Minimal Base Repo

PullRequestMinimalBaseRepoResponse
Pull Request Minimal Base Repo

PullRequestMinimalBaseResponse
Pull Request Minimal Base

PullRequestMinimalHeadArg
Pull Request Minimal Head

PullRequestMinimalHeadRepoArg
Pull Request Minimal Head Repo

PullRequestMinimalHeadRepoResponse
Pull Request Minimal Head Repo

PullRequestMinimalHeadResponse
Pull Request Minimal Head

PullRequestMinimalResponse
Pull Request Minimal

PullRequestReviewArg
Pull Request Reviews are reviews on pull requests.

PullRequestReviewCommentArg
Pull Request Review Comments are comments on a portion of the Pull Request's diff.

PullRequestReviewCommentLinksArg
Pull Request Review Comment Links

PullRequestReviewCommentLinksHtmlArg
Pull Request Review Comment Links Html

PullRequestReviewCommentLinksHtmlResponse
Pull Request Review Comment Links Html

PullRequestReviewCommentLinksPullRequestArg
Pull Request Review Comment Links Pull Request

PullRequestReviewCommentLinksPullRequestResponse
Pull Request Review Comment Links Pull Request

PullRequestReviewCommentLinksResponse
Pull Request Review Comment Links

PullRequestReviewCommentResponse
Pull Request Review Comments are comments on a portion of the Pull Request's diff.

PullRequestReviewLinksArg
Pull Request Review Links

PullRequestReviewLinksHtmlArg
Pull Request Review Links Html

PullRequestReviewLinksHtmlResponse
Pull Request Review Links Html

PullRequestReviewLinksPullRequestArg
Pull Request Review Links Pull Request

PullRequestReviewLinksPullRequestResponse
Pull Request Review Links Pull Request

PullRequestReviewLinksResponse
Pull Request Review Links

PullRequestReviewRequestArg
Pull Request Review Request

PullRequestReviewRequestResponse
Pull Request Review Request

PullRequestReviewRequestTeamsItemArg
Pull Request Review Request Teams Item

PullRequestReviewRequestTeamsItemResponse
Pull Request Review Request Teams Item

PullRequestReviewRequestUsersItemArg
Pull Request Review Request Users Item

PullRequestReviewRequestUsersItemResponse
Pull Request Review Request Users Item

PullRequestReviewResponse
Pull Request Reviews are reviews on pull requests.

PullRequestReviewUserArg
Pull Request Review User

PullRequestReviewUserResponse
Pull Request Review User

PullRequestSchemaArg
Pull requests let you tell others about changes you've pushed to a repository on GitHub. Once a pull request is sent, interested parties can review the set of changes, discuss potential modifications, and even push follow-up commits if necessary.

PullRequestSchemaResponse
Pull requests let you tell others about changes you've pushed to a repository on GitHub. Once a pull request is sent, interested parties can review the set of changes, discuss potential modifications, and even push follow-up commits if necessary.

PullRequestSimpleArg
Pull Request Simple

PullRequestSimpleAssigneeArg
Pull Request Simple Assignee

PullRequestSimpleAssigneeResponse
Pull Request Simple Assignee

PullRequestSimpleBaseArg
Pull Request Simple Base

PullRequestSimpleBaseResponse
Pull Request Simple Base

PullRequestSimpleBaseUserArg
Pull Request Simple Base User

PullRequestSimpleBaseUserResponse
Pull Request Simple Base User

PullRequestSimpleHeadArg
Pull Request Simple Head

PullRequestSimpleHeadResponse
Pull Request Simple Head

PullRequestSimpleHeadUserArg
Pull Request Simple Head User

PullRequestSimpleHeadUserResponse
Pull Request Simple Head User

PullRequestSimpleLabelsItemArg
Pull Request Simple Labels Item

PullRequestSimpleLabelsItemResponse
Pull Request Simple Labels Item

PullRequestSimpleLinksArg
Pull Request Simple Links

PullRequestSimpleLinksResponse
Pull Request Simple Links

PullRequestSimpleMilestoneArg
Pull Request Simple Milestone

PullRequestSimpleMilestoneResponse
Pull Request Simple Milestone

PullRequestSimpleResponse
Pull Request Simple

PullRequestSimpleUserArg
Pull Request Simple User

PullRequestSimpleUserResponse
Pull Request Simple User

PullRequestTemplateArg
Pull Request Template

PullRequestTemplateResponse
Pull Request Template

PullRequestUserArg
Pull Request User

PullRequestUserResponse
Pull Request User

PusherArg
Pusher

PusherResponse
Pusher

RateLimitArg
Rate Limit

RateLimitOverviewArg
Rate Limit Overview

RateLimitOverviewResponse
Rate Limit Overview

RateLimitResponse
Rate Limit

ReactionArg
Reactions to conversations provide a way to help people express their feelings more simply and effectively.

ReactionResponse
Reactions to conversations provide a way to help people express their feelings more simply and effectively.

ReactionRollupArg
Reaction Rollup

ReactionRollupResponse
Reaction Rollup

ReactionUserArg
Reaction User

ReactionUserResponse
Reaction User

ReadmeArg
Readme

ReadmeResponse
Readme

ReferrerTrafficArg
Referrer Traffic

ReferrerTrafficResponse
Referrer Traffic

RelatedItemArg
Related Item

RelatedItemResponse
Related Item

RelatedItemTopicRelationArg
Related Item Topic Relation

RelatedItemTopicRelationResponse
Related Item Topic Relation

ReleaseArg
A release.

ReleaseAssetArg
Data related to a release.

ReleaseAssetResponse
Data related to a release.

ReleaseResponse
A release.

RepoSearchResultItemArg
Repo Search Result Item

RepoSearchResultItemLicenseArg
Repo Search Result Item License

RepoSearchResultItemLicenseResponse
Repo Search Result Item License

RepoSearchResultItemOwnerArg
Repo Search Result Item Owner

RepoSearchResultItemOwnerResponse
Repo Search Result Item Owner

RepoSearchResultItemPermissionsArg
Repo Search Result Item Permissions

RepoSearchResultItemPermissionsResponse
Repo Search Result Item Permissions

RepoSearchResultItemResponse
Repo Search Result Item

RepositoryArg
A git repository

RepositoryCollaboratorPermissionArg
Repository Collaborator Permission

RepositoryCollaboratorPermissionResponse
Repository Collaborator Permission

RepositoryCollaboratorPermissionUserArg
Repository Collaborator Permission User

RepositoryCollaboratorPermissionUserResponse
Repository Collaborator Permission User

RepositoryInvitationArg
Repository invitations let you manage who you collaborate with.

RepositoryInvitationResponse
Repository invitations let you manage who you collaborate with.

RepositoryLicenseArg
Repository License

RepositoryLicenseResponse
Repository License

RepositoryOwnerArg
Repository Owner

RepositoryOwnerResponse
Repository Owner

RepositoryPermissionsArg
Repository Permissions

RepositoryPermissionsResponse
Repository Permissions

RepositoryResponse
A git repository

RepositorySubscriptionArg
Repository invitations let you manage who you collaborate with.

RepositorySubscriptionResponse
Repository invitations let you manage who you collaborate with.

RepositoryTemplateRepositoryArg
Repository Template Repository

RepositoryTemplateRepositoryOwnerArg
Repository Template Repository Owner

RepositoryTemplateRepositoryOwnerResponse
Repository Template Repository Owner

RepositoryTemplateRepositoryPermissionsArg
Repository Template Repository Permissions

RepositoryTemplateRepositoryPermissionsResponse
Repository Template Repository Permissions

RepositoryTemplateRepositoryResponse
Repository Template Repository

RequestedReviewerArg
Requested Reviewer

RequestedReviewerResponse
Requested Reviewer

RequiredPullRequestReviewsArg
Required Pull Request Reviews

RequiredPullRequestReviewsDismissalRestrictionsArg
Required Pull Request Reviews Dismissal Restrictions

RequiredPullRequestReviewsDismissalRestrictionsResponse
Required Pull Request Reviews Dismissal Restrictions

RequiredPullRequestReviewsResponse
Required Pull Request Reviews

RequiredSignaturesArg
Required Signatures

RequiredSignaturesResponse
Required Signatures

RequiredStatusChecksArg
Required Status Checks

RequiredStatusChecksResponse
Required Status Checks

ResourcesArg
Resources

ResourcesResponse
Resources

ReviewCommentArg
Legacy Review Comment

ReviewCommentLinksArg
Review Comment Links

ReviewCommentLinksResponse
Review Comment Links

ReviewCommentResponse
Legacy Review Comment

ReviewCommentUserArg
Review Comment User

ReviewCommentUserResponse
Review Comment User

ReviewRequesterArg
Review Requester

ReviewRequesterResponse
Review Requester

RunnerApplicationArg
Runner Application

RunnerApplicationResponse
Runner Application

RunnerArg
A self hosted runner

RunnerResponse
A self hosted runner

ScimErrorArg
Scim Error

ScimErrorResponse
Scim Error

ScimUserArg
SCIM /Users provisioning endpoints

ScimUserEmailsItemArg
Scim User Emails Item

ScimUserEmailsItemResponse
Scim User Emails Item

ScimUserGroupsItemArg
Scim User Groups Item

ScimUserGroupsItemResponse
Scim User Groups Item

ScimUserListArg
SCIM User List

ScimUserListResponse
SCIM User List

ScimUserResponse
SCIM /Users provisioning endpoints

ScopedInstallationArg
Scoped Installation

ScopedInstallationPermissionsArg
Scoped Installation Permissions

ScopedInstallationPermissionsResponse
Scoped Installation Permissions

ScopedInstallationResponse
Scoped Installation

SearchResultTextMatchesItemArg
Search Result Text Matches Item

SearchResultTextMatchesItemResponse
Search Result Text Matches Item

SelfArg
Self

SelfResponse
Self

ServiceUnavailableArg
Service Unavailable

ServiceUnavailableResponse
Service Unavailable

ShortBlobArg
Short Blob

ShortBlobResponse
Short Blob

ShortBranchArg
Short Branch

ShortBranchCommitArg
Short Branch Commit

ShortBranchCommitResponse
Short Branch Commit

ShortBranchResponse
Short Branch

SimpleCommitArg
Simple Commit

SimpleCommitAuthorArg
Simple Commit Author

SimpleCommitAuthorResponse
Simple Commit Author

SimpleCommitCommitterArg
Simple Commit Committer

SimpleCommitCommitterResponse
Simple Commit Committer

SimpleCommitResponse
Simple Commit

SimpleCommitStatusArg
Simple Commit Status

SimpleCommitStatusResponse
Simple Commit Status

SimpleUserArg
Simple User

SimpleUserResponse
Simple User

SshKeyFingerprintsArg
Ssh Key Fingerprints

SshKeyFingerprintsResponse
Ssh Key Fingerprints

StargazerArg
Stargazer

StargazerResponse
Stargazer

StargazerUserArg
Stargazer User

StargazerUserResponse
Stargazer User

StarredRepositoryArg
Starred Repository

StarredRepositoryResponse
Starred Repository

StatsArg
Stats

StatsResponse
Stats

StatusCheckPolicyArg
Status Check Policy

StatusCheckPolicyResponse
Status Check Policy

StatusSchemaArg
The status of a commit.

StatusSchemaResponse
The status of a commit.

StepsItemArg
Steps Item

StepsItemResponse
Steps Item

SubjectArg
Subject

SubjectResponse
Subject

SubkeysItemEmailsItemArg
Subkeys Item Emails Item

SubkeysItemEmailsItemResponse
Subkeys Item Emails Item

SubkeysItemSubkeysItemArg
Subkeys Item Subkeys Item

SubkeysItemSubkeysItemResponse
Subkeys Item Subkeys Item

SuspendedByArg
Suspended By

SuspendedByResponse
Suspended By

TagArg
Tag

TagCommitArg
Tag Commit

TagCommitResponse
Tag Commit

TagResponse
Tag

TaggerArg
Tagger

TaggerResponse
Tagger

TeamArg
Groups of organization members that gives permissions on specified repositories.

TeamDiscussionArg
A team discussion is a persistent record of a free-form conversation within a team.

TeamDiscussionAuthorArg
Team Discussion Author

TeamDiscussionAuthorResponse
Team Discussion Author

TeamDiscussionCommentArg
A reply to a discussion within a team.

TeamDiscussionCommentAuthorArg
Team Discussion Comment Author

TeamDiscussionCommentAuthorResponse
Team Discussion Comment Author

TeamDiscussionCommentResponse
A reply to a discussion within a team.

TeamDiscussionResponse
A team discussion is a persistent record of a free-form conversation within a team.

TeamFullArg
Groups of organization members that gives permissions on specified repositories.

TeamFullParentArg
Team Full Parent

TeamFullParentResponse
Team Full Parent

TeamFullResponse
Groups of organization members that gives permissions on specified repositories.

TeamMembershipArg
Team Membership

TeamMembershipResponse
Team Membership

TeamParentArg
Team Parent

TeamParentResponse
Team Parent

TeamProjectArg
A team's access to a project.

TeamProjectPermissionsArg
Team Project Permissions

TeamProjectPermissionsResponse
Team Project Permissions

TeamProjectResponse
A team's access to a project.

TeamRepositoryArg
A team's access to a repository.

TeamRepositoryLicenseArg
Team Repository License

TeamRepositoryLicenseResponse
Team Repository License

TeamRepositoryOwnerArg
Team Repository Owner

TeamRepositoryOwnerResponse
Team Repository Owner

TeamRepositoryPermissionsArg
Team Repository Permissions

TeamRepositoryPermissionsResponse
Team Repository Permissions

TeamRepositoryResponse
A team's access to a repository.

TeamRepositoryTemplateRepositoryArg
Team Repository Template Repository

TeamRepositoryTemplateRepositoryOwnerArg
Team Repository Template Repository Owner

TeamRepositoryTemplateRepositoryOwnerResponse
Team Repository Template Repository Owner

TeamRepositoryTemplateRepositoryPermissionsArg
Team Repository Template Repository Permissions

TeamRepositoryTemplateRepositoryPermissionsResponse
Team Repository Template Repository Permissions

TeamRepositoryTemplateRepositoryResponse
Team Repository Template Repository

TeamResponse
Groups of organization members that gives permissions on specified repositories.

TeamSimpleArg
Groups of organization members that gives permissions on specified repositories.

TeamSimpleResponse
Groups of organization members that gives permissions on specified repositories.

ThreadArg
Thread

ThreadResponse
Thread

ThreadSubscriptionArg
Thread Subscription

ThreadSubscriptionResponse
Thread Subscription

TopicArg
A topic aggregates entities that are related to a subject.

TopicResponse
A topic aggregates entities that are related to a subject.

TopicSearchResultItemArg
Topic Search Result Item

TopicSearchResultItemResponse
Topic Search Result Item

TrafficArg
Traffic

TrafficResponse
Traffic

TreeItemArg
Tree Item

TreeItemResponse
Tree Item

UploaderArg
Uploader

UploaderResponse
Uploader

UserMarketplacePurchaseArg
User Marketplace Purchase

UserMarketplacePurchaseResponse
User Marketplace Purchase

UserSearchResultItemArg
User Search Result Item

UserSearchResultItemResponse
User Search Result Item

ValidationErrorArg
Validation Error

ValidationErrorResponse
Validation Error

ValidationErrorSimpleArg
Validation Error Simple

ValidationErrorSimpleResponse
Validation Error Simple

VerificationSchemaArg
Verification Schema

VerificationSchemaResponse
Verification Schema

ViewTrafficArg
View Traffic

ViewTrafficResponse
View Traffic

WeeksItemArg
Weeks Item

WeeksItemResponse
Weeks Item

WorkflowArg
A GitHub Actions workflow

WorkflowResponse
A GitHub Actions workflow

WorkflowRunArg
An invocation of a workflow

WorkflowRunResponse
An invocation of a workflow

WorkflowRunUsageArg
Workflow Run Usage

WorkflowRunUsageBillableArg
Workflow Run Usage Billable

WorkflowRunUsageBillableMacosArg
Workflow Run Usage Billable Macos

WorkflowRunUsageBillableMacosResponse
Workflow Run Usage Billable Macos

WorkflowRunUsageBillableResponse
Workflow Run Usage Billable

WorkflowRunUsageBillableUbuntuArg
Workflow Run Usage Billable Ubuntu

WorkflowRunUsageBillableUbuntuResponse
Workflow Run Usage Billable Ubuntu

WorkflowRunUsageBillableWindowsArg
Workflow Run Usage Billable Windows

WorkflowRunUsageBillableWindowsResponse
Workflow Run Usage Billable Windows

WorkflowRunUsageResponse
Workflow Run Usage

WorkflowUsageArg
Workflow Usage

WorkflowUsageBillableArg
Workflow Usage Billable

WorkflowUsageBillableMacosArg
Workflow Usage Billable Macos

WorkflowUsageBillableMacosResponse
Workflow Usage Billable Macos

WorkflowUsageBillableResponse
Workflow Usage Billable

WorkflowUsageBillableUbuntuArg
Workflow Usage Billable Ubuntu

WorkflowUsageBillableUbuntuResponse
Workflow Usage Billable Ubuntu

WorkflowUsageBillableWindowsArg
Workflow Usage Billable Windows

WorkflowUsageBillableWindowsResponse
Workflow Usage Billable Windows

WorkflowUsageResponse
Workflow Usage

Enums

AuthenticationTokenRepositorySelection
Describe whether all repositories have been selected or there's a selection involved

CheckRunStatus
The phase of the lifecycle that the check is currently in.

ClosedReason
Closed Reason

CommitComparisonStatus
Commit Comparison Status

DeploymentStatusState
The state of the status.

Direction
Direction

ImportStatus
Import Status

InstallationRepositorySelection
Describe whether all repositories have been selected or there's a selection involved

InstallationTokenRepositorySelection
Installation Token Repository Selection

JobStatus
The phase of the lifecycle that the job is currently in.

Limit
The interaction limit to enable.

MilestoneState
The state of the milestone.

Op
Op

Order
Order

OrganizationPermission
The baseline permission that all organization members have on this project

PageStatus
The status of the most recent build of the Page.

Per
Per

Privacy
The level of privacy this team should have

PullRequestReviewCommentSide
The side of the diff to which the comment applies. The side of the last line of the range for a multi-line comment

PullRequestReviewCommentStartSide
The side of the first line of the range for a multi-line comment.

PullRequestState
State of this Pull Request. Either open or closed.

ReactionContent
The reaction to use

ReleaseAssetState
State of the release asset.

RepositoryInvitationPermissions
The permission associated with the invitation.

ReviewCommentSide
The side of the first line of the range for a multi-line comment.

ReviewCommentStartSide
The side of the first line of the range for a multi-line comment.

Role
The role of the user in the team.

RuleSeverity
The severity of the alert.

ScopedInstallationRepositorySelection
Describe whether all repositories have been selected or there's a selection involved

Sort
Sort

StatusParam
Status Param

StepsItemStatus
The phase of the lifecycle that the job is currently in.

Visibility
Visibility of a secret

WorkflowRunStatus
Workflow Run Status

WorkflowState
Workflow State