Module octorest::types

Miscellaneous data types used in the GitHub API.

The API is designed such that users shall not need to explicitly import types from this module.

Structs

Account
Account

ActionsBillingUsage
Actions Billing Usage

ActionsPublicKey
The public key used for setting Actions Secrets.

ActionsSecret
Set secrets for GitHub Actions.

ActorSchema
Actor

AliasesItem
Aliases Item

AliasesItemTopicRelation
Aliases Item Topic Relation

ApiOverview
Api Overview

ApplicationGrant
The authorization associated with an OAuth Access.

ApplicationGrantApp
Application Grant App

ApplicationGrantUser
Application Grant User

AppsItem
Apps Item

AppsItemOwner
Apps Item Owner

AppsItemPermissions
Apps Item Permissions

Artifact
An artifact

Assigner
Assigner

AuthenticationToken
Authentication Token

AuthenticationTokenPermissions
Authentication Token Permissions

Authorization
The authorization for an OAuth app, GitHub App, or a Personal Access Token.

AuthorizationApp
Authorization App

AuthorizationInstallation
Authorization Installation

AuthorizationUser
Authorization User

AutoTriggerChecksItem
Auto Trigger Checks Item

BaseGist
Base Gist

BaseGistFiles
Base Gist Files

BaseGistOwner
Base Gist Owner

BaseGistUser
Base Gist User

BaseRepoLicense
Base Repo License

BaseRepoOwner
Base Repo Owner

BaseRepoPermissions
Base Repo Permissions

BasicError
Basic Error

Blob
Blob

Block
Block

BranchProtection
Branch Protection

BranchProtectionAllowDeletions
Branch Protection Allow Deletions

BranchProtectionAllowForcePushes
Branch Protection Allow Force Pushes

BranchProtectionRequiredLinearHistory
Branch Protection Required Linear History

BranchRestrictionPolicy
Branch Restriction Policy

BranchRestrictionPolicyTeamsItem
Branch Restriction Policy Teams Item

BranchRestrictionPolicyUsersItem
Branch Restriction Policy Users Item

BranchShort
Branch Short

BranchShortCommit
Branch Short Commit

BranchWithProtection
Branch With Protection

BranchWithProtectionLinks
Branch With Protection Links

ChangeStatus
Change Status

CheckAnnotation
Check Annotation

CheckRun
A check performed on the code of a given code change

CheckRunApp
Check Run App

CheckRunCheckSuite
Check Run Check Suite

CheckSuiteApp
Check Suite App

CheckSuitePreference
Check suite configuration preferences for a repository.

CheckSuiteSchema
A suite of checks performed on the code of a given code change

CloneTraffic
Clone Traffic

CodeOfConductSchema
Code Of Conduct

CodeOfConductSimple
Code of Conduct Simple

CodeScanningAlert
Code Scanning Alert

CodeScanningAlertClosedBy
Code Scanning Alert Closed By

CodeSearchResultItem
Code Search Result Item

Collaborator
Collaborator

CollaboratorPermissions
Collaborator Permissions

CombinedBillingUsage
Combined Billing Usage

CombinedCommitStatus
Combined Commit Status

CommitActivity
Commit Activity

CommitAuthorSchema
Commit Author Schema

CommitComment
Commit Comment

CommitCommentUser
Commit Comment User

CommitCommit
Commit Commit

CommitCommitAuthor
Commit Commit Author

CommitCommitCommitter
Commit Commit Committer

CommitCommitTree
Commit Commit Tree

CommitCommitterSchema
Commit Committer Schema

CommitComparison
Commit Comparison

CommitParentsItemSchema
Commit Parents Item Schema

CommitSchema
Commit

CommitSearchResultItem
Commit Search Result Item

CommitSearchResultItemAuthor
Commit Search Result Item Author

CommitSearchResultItemCommit
Commit Search Result Item Commit

CommitSearchResultItemCommitAuthor
Commit Search Result Item Commit Author

CommitSearchResultItemCommitCommitter
Commit Search Result Item Commit Committer

CommitSearchResultItemCommitTree
Commit Search Result Item Commit Tree

CommitSearchResultItemCommitter
Commit Search Result Item Committer

CommitSearchResultItemParentsItem
Commit Search Result Item Parents Item

CommitVerification
Commit Verification

CommunityHealthFile
Community Health File

CommunityProfile
Community Profile

CommunityProfileFiles
Community Profile Files

ContentDirectoryItem
Content Directory Item

ContentDirectoryItemLinks
Content Directory Item Links

ContentFile
Content File

ContentFileLinks
Content File Links

ContentLinks
Content Links

ContentReferenceAttachment
Content Reference attachments allow you to provide context around URLs posted in comments

ContentSubmodule
An object describing a symlink

ContentSubmoduleLinks
Content Submodule Links

ContentSymlink
An object describing a symlink

ContentSymlinkLinks
Content Symlink Links

ContentTraffic
Content Traffic

ContentTree
Content Tree

ContentTreeLinks
Content Tree Links

ContextsItem
Contexts Item

Contributing
Contributing

Contributor
Contributor

ContributorActivity
Contributor Activity

ContributorActivityAuthor
Contributor Activity Author

CredentialAuthorization
Credential Authorization

DeployKey
An SSH key granting access to a single repository.

Deployment
A request for a specific ref(branch,sha,tag) to be deployed

DeploymentCreator
Deployment Creator

DeploymentPayload
Deployment Payload

DeploymentPerformedViaGithubApp
Deployment Performed Via Github App

DeploymentStatus
The status of a deployment.

DeploymentStatusCreator
Deployment Status Creator

DeploymentStatusPerformedViaGithubApp
Deployment Status Performed Via Github App

DiffEntry
Diff Entry

Email
Email

EnforceAdmins
Enforce Admins

Enterprise
An enterprise account

EntriesItem
Entries Item

EntriesItemLinks
Entries Item Links

Error
Error

ErrorsItem
Errors Item

ErrorsItemValue
Errors Item Value

Event
Event

EventPayload
Event Payload

EventRepo
Event Repo

ExcludeItem
Exclude Item

Feed
Feed

FeedLinks
Feed Links

FileCommit
File Commit

FileCommitCommit
File Commit Commit

FileCommitCommitAuthor
File Commit Commit Author

FileCommitCommitCommitter
File Commit Commit Committer

FileCommitCommitParentsItem
File Commit Commit Parents Item

FileCommitCommitTree
File Commit Commit Tree

FileCommitContent
File Commit Content

FilesCodeOfConduct
Files Code Of Conduct

FilesItem
Files Item

FilesLicense
Files License

ForbiddenGist
Forbidden Gist

ForksItem
Forks Item

FullRepository
Full Repository

FullRepositoryLicense
Full Repository License

FullRepositoryOrganization
Full Repository Organization

FullRepositoryOwner
Full Repository Owner

FullRepositoryPermissions
Full Repository Permissions

FullRepositoryTemplateRepository
Full Repository Template Repository

GistComment
A comment made to a gist.

GistCommentUser
Gist Comment User

GistCommit
Gist Commit

GistCommitUser
Gist Commit User

GistFull
Gist Full

GistSimple
Gist Simple

GistSimpleFiles
Gist Simple Files

GitCommit
Low-level Git commit operations within a repository

GitCommitAuthor
Identifying information for the git-user

GitCommitCommitter
Identifying information for the git-user

GitCommitParentsItem
Git Commit Parents Item

GitCommitTree
Git Commit Tree

GitCommitVerification
Git Commit Verification

GitRef
Git references within a repository

GitRefObject
Git Ref Object

GitTag
Metadata for a Git tag

GitTagObject
Git Tag Object

GitTree
The hierarchy between files in a Git repository.

GitUser
Metaproperties for Git author/committer information.

GitignoreTemplate
Gitignore Template

GpgKey
A unique encryption key

GpgKeyEmailsItem
Gpg Key Emails Item

GpgKeySubkeysItem
Gpg Key Subkeys Item

GroupMapping
External Groups to be mapped to a team for membership

GroupMappingGroupsItem
Group Mapping Groups Item

HeadRepoLicense
Head Repo License

HeadRepoOwner
Head Repo Owner

HeadRepoPermissions
Head Repo Permissions

HistoryItem
History Item

Hook
Webhooks for repositories.

HookConfig
Hook Config

HookResponse
Hook Response

Hovercard
Hovercard

Import
A repository import from an external source.

InstallationPermissions
Installation Permissions

InstallationSchema
Installation

InstallationToken
Authentication token for a GitHub App installed on a user or org.

InstallationTokenPermissions
Installation Token Permissions

Integration
GitHub apps are a new way to extend GitHub. They can be installed directly on organizations and user accounts and granted access to specific repositories. They come with granular permissions and built-in webhooks. GitHub apps are first class actors within GitHub.

IntegrationOwner
Integration Owner

IntegrationPermissions
The set of permissions for the GitHub app

InteractionLimit
Interaction limit settings.

Invitee
Invitee

Inviter
Inviter

Issue
Issues are a great way to keep track of tasks, enhancements, and bugs for your projects.

IssueAssignee
Issue Assignee

IssueClosedBy
Issue Closed By

IssueComment
Comments provide a way for people to collaborate on an issue.

IssueCommentPerformedViaGithubApp
Issue Comment Performed Via Github App

IssueCommentUser
Issue Comment User

IssueEvent
Issue Event

IssueEventActor
Issue Event Actor

IssueEventAssignee
Issue Event Assignee

IssueEventDismissedReview
Issue Event Dismissed Review

IssueEventForIssue
Issue Event for Issue

IssueEventLabel
Issue Event Label

IssueEventMilestone
Issue Event Milestone

IssueEventProjectCard
Issue Event Project Card

IssueEventRename
Issue Event Rename

IssueLabelsItem
Issue Labels Item

IssueMilestone
Issue Milestone

IssuePerformedViaGithubApp
Issue Performed Via Github App

IssuePullRequest
Issue Pull Request

IssueSearchResultItem
Issue Search Result Item

IssueSearchResultItemAssignee
Issue Search Result Item Assignee

IssueSearchResultItemLabelsItem
Issue Search Result Item Labels Item

IssueSearchResultItemMilestone
Issue Search Result Item Milestone

IssueSearchResultItemPerformedViaGithubApp
Issue Search Result Item Performed Via Github App

IssueSearchResultItemPullRequest
Issue Search Result Item Pull Request

IssueSearchResultItemUser
Issue Search Result Item User

IssueSimple
Issue Simple

IssueSimpleAssignee
Issue Simple Assignee

IssueSimpleLabelsItem
Issue Simple Labels Item

IssueSimpleMilestone
Issue Simple Milestone

IssueSimplePerformedViaGithubApp
Issue Simple Performed Via Github App

IssueSimplePullRequest
Issue Simple Pull Request

IssueSimpleUser
Issue Simple User

IssueTemplate
Issue Template

IssueUser
Issue User

Job
Information of a job execution in a workflow run

Key
Key

KeySimple
Key Simple

Label
Color-coded labels help you categorize and filter your issues (just like labels in Gmail).

LabelSearchResultItem
Label Search Result Item

Language
Language

LicenseContent
License Content

LicenseContentLicense
License Content License

LicenseContentLinks
License Content Links

LicenseSchema
License

LicenseSimple
License Simple

Link
Hypermedia Link

LinkWithType
Hypermedia Link with Type

MarketplaceAccount
Marketplace Account

MarketplaceListingPlan
Marketplace Listing Plan

MarketplacePendingChange
Marketplace Pending Change

MarketplacePurchaseMarketplacePurchase
Marketplace Purchase Marketplace Purchase

MarketplacePurchaseSchema
Marketplace Purchase

MatchesItem
Matches Item

MergedBy
Merged By

Meta
Meta

Migration
A migration.

MigrationOwner
Migration Owner

MilestoneCreator
Milestone Creator

MilestoneSchema
A collection of related issues and pull requests.

MinimalRepository
Minimal Repository

MinimalRepositoryLicense
Minimal Repository License

MinimalRepositoryOwner
Minimal Repository Owner

MinimalRepositoryPermissions
Minimal Repository Permissions

MinutesUsedBreakdown
Minutes Used Breakdown

Name
Name

OperationsItem
Operations Item

OperationsItemValue
Operations Item Value

OrgHook
Org Hook

OrgHookConfig
Org Hook Config

OrgMembership
Org Membership

OrgMembershipPermissions
Org Membership Permissions

OrgMembershipUser
Org Membership User

OrganizationActionsSecret
Secrets for GitHub Actions for an organization.

OrganizationFull
Organization Full

OrganizationFullPlan
Organization Full Plan

OrganizationInvitation
Organization Invitation

OrganizationPlan
Organization Plan

OrganizationSchema
GitHub account for managing multiple users, teams, and repositories

OrganizationSimple
Organization Simple

Output
Output

PackagesBillingUsage
Packages Billing Usage

Page
The configuration for GitHub Pages for a repository.

PageBuild
Page Build

PageBuildStatus
Page Build Status

PagesItem
Pages Item

PagesSourceHash
Pages Source Hash

ParticipationStats
Participation Stats

PorterAuthor
Porter Author

PorterLargeFile
Porter Large File

Preferences
Preferences

PreviewHeaderMissing
Preview Header Missing

PrivateUser
Private User

PrivateUserPlan
Private User Plan

Project
Projects are a way to organize columns and cards of work.

ProjectCard
Project cards represent a scope of work.

ProjectCardCreator
Project Card Creator

ProjectChoicesItem
Project Choices Item

ProjectColumn
Project columns contain cards of work.

ProjectCreator
Project Creator

ProjectPermissions
Project Permissions

ProtectedBranch
Branch protections protect branches

ProtectedBranchAdminEnforced
Protected Branch Admin Enforced

ProtectedBranchAllowDeletions
Protected Branch Allow Deletions

ProtectedBranchAllowForcePushes
Protected Branch Allow Force Pushes

ProtectedBranchPullRequestReview
Protected Branch Pull Request Review

ProtectedBranchPullRequestReviewDismissalRestrictions
Protected Branch Pull Request Review Dismissal Restrictions

ProtectedBranchRequiredLinearHistory
Protected Branch Required Linear History

PublicUser
Public User

PublicUserPlan
Public User Plan

PullRequestAssignee
Pull Request Assignee

PullRequestBase
Pull Request Base

PullRequestBaseRepo
Pull Request Base Repo

PullRequestBaseUser
Pull Request Base User

PullRequestHead
Pull Request Head

PullRequestHeadRepo
Pull Request Head Repo

PullRequestHeadUser
Pull Request Head User

PullRequestLabelsItem
Pull Request Labels Item

PullRequestLinks
Pull Request Links

PullRequestMergeResult
Pull Request Merge Result

PullRequestMilestone
Pull Request Milestone

PullRequestMinimal
Pull Request Minimal

PullRequestMinimalBase
Pull Request Minimal Base

PullRequestMinimalBaseRepo
Pull Request Minimal Base Repo

PullRequestMinimalHead
Pull Request Minimal Head

PullRequestMinimalHeadRepo
Pull Request Minimal Head Repo

PullRequestReview
Pull Request Reviews are reviews on pull requests.

PullRequestReviewComment
Pull Request Review Comments are comments on a portion of the Pull Request's diff.

PullRequestReviewCommentLinks
Pull Request Review Comment Links

PullRequestReviewCommentLinksHtml
Pull Request Review Comment Links Html

PullRequestReviewCommentLinksPullRequest
Pull Request Review Comment Links Pull Request

PullRequestReviewLinks
Pull Request Review Links

PullRequestReviewLinksHtml
Pull Request Review Links Html

PullRequestReviewLinksPullRequest
Pull Request Review Links Pull Request

PullRequestReviewRequest
Pull Request Review Request

PullRequestReviewRequestTeamsItem
Pull Request Review Request Teams Item

PullRequestReviewRequestUsersItem
Pull Request Review Request Users Item

PullRequestReviewUser
Pull Request Review User

PullRequestSchema
Pull requests let you tell others about changes you've pushed to a repository on GitHub. Once a pull request is sent, interested parties can review the set of changes, discuss potential modifications, and even push follow-up commits if necessary.

PullRequestSimple
Pull Request Simple

PullRequestSimpleAssignee
Pull Request Simple Assignee

PullRequestSimpleBase
Pull Request Simple Base

PullRequestSimpleBaseUser
Pull Request Simple Base User

PullRequestSimpleHead
Pull Request Simple Head

PullRequestSimpleHeadUser
Pull Request Simple Head User

PullRequestSimpleLabelsItem
Pull Request Simple Labels Item

PullRequestSimpleLinks
Pull Request Simple Links

PullRequestSimpleMilestone
Pull Request Simple Milestone

PullRequestSimpleUser
Pull Request Simple User

PullRequestTemplate
Pull Request Template

PullRequestUser
Pull Request User

Pusher
Pusher

RateLimit
Rate Limit

RateLimitOverview
Rate Limit Overview

Reaction
Reactions to conversations provide a way to help people express their feelings more simply and effectively.

ReactionRollup
Reaction Rollup

ReactionUser
Reaction User

Readme
Readme

ReferrerTraffic
Referrer Traffic

RelatedItem
Related Item

RelatedItemTopicRelation
Related Item Topic Relation

Release
A release.

ReleaseAsset
Data related to a release.

RepoSearchResultItem
Repo Search Result Item

RepoSearchResultItemLicense
Repo Search Result Item License

RepoSearchResultItemOwner
Repo Search Result Item Owner

RepoSearchResultItemPermissions
Repo Search Result Item Permissions

Repository
A git repository

RepositoryCollaboratorPermission
Repository Collaborator Permission

RepositoryCollaboratorPermissionUser
Repository Collaborator Permission User

RepositoryInvitation
Repository invitations let you manage who you collaborate with.

RepositoryLicense
Repository License

RepositoryOwner
Repository Owner

RepositoryPermissions
Repository Permissions

RepositorySubscription
Repository invitations let you manage who you collaborate with.

RepositoryTemplateRepository
Repository Template Repository

RepositoryTemplateRepositoryOwner
Repository Template Repository Owner

RepositoryTemplateRepositoryPermissions
Repository Template Repository Permissions

RequestedReviewer
Requested Reviewer

RequiredPullRequestReviews
Required Pull Request Reviews

RequiredPullRequestReviewsDismissalRestrictions
Required Pull Request Reviews Dismissal Restrictions

RequiredSignatures
Required Signatures

RequiredStatusChecks
Required Status Checks

Resources
Resources

ReviewComment
Legacy Review Comment

ReviewCommentLinks
Review Comment Links

ReviewCommentUser
Review Comment User

ReviewRequester
Review Requester

Runner
A self hosted runner

RunnerApplication
Runner Application

ScimError
Scim Error

ScimUser
SCIM /Users provisioning endpoints

ScimUserEmailsItem
Scim User Emails Item

ScimUserGroupsItem
Scim User Groups Item

ScimUserList
SCIM User List

ScopedInstallation
Scoped Installation

ScopedInstallationPermissions
Scoped Installation Permissions

SearchResultTextMatchesItem
Search Result Text Matches Item

Self_
Self

ServiceUnavailable
Service Unavailable

ShortBlob
Short Blob

ShortBranch
Short Branch

ShortBranchCommit
Short Branch Commit

SimpleCommit
Simple Commit

SimpleCommitAuthor
Simple Commit Author

SimpleCommitCommitter
Simple Commit Committer

SimpleCommitStatus
Simple Commit Status

SimpleUser
Simple User

SshKeyFingerprints
Ssh Key Fingerprints

Stargazer
Stargazer

StargazerUser
Stargazer User

StarredRepository
Starred Repository

Stats
Stats

StatusCheckPolicy
Status Check Policy

StatusSchema
The status of a commit.

StepsItem
Steps Item

Subject
Subject

SubkeysItemEmailsItem
Subkeys Item Emails Item

SubkeysItemSubkeysItem
Subkeys Item Subkeys Item

SuspendedBy
Suspended By

Tag
Tag

TagCommit
Tag Commit

Tagger
Tagger

Team
Groups of organization members that gives permissions on specified repositories.

TeamDiscussion
A team discussion is a persistent record of a free-form conversation within a team.

TeamDiscussionAuthor
Team Discussion Author

TeamDiscussionComment
A reply to a discussion within a team.

TeamDiscussionCommentAuthor
Team Discussion Comment Author

TeamFull
Groups of organization members that gives permissions on specified repositories.

TeamFullParent
Team Full Parent

TeamMembership
Team Membership

TeamParent
Team Parent

TeamProject
A team's access to a project.

TeamProjectPermissions
Team Project Permissions

TeamRepository
A team's access to a repository.

TeamRepositoryLicense
Team Repository License

TeamRepositoryOwner
Team Repository Owner

TeamRepositoryPermissions
Team Repository Permissions

TeamRepositoryTemplateRepository
Team Repository Template Repository

TeamRepositoryTemplateRepositoryOwner
Team Repository Template Repository Owner

TeamRepositoryTemplateRepositoryPermissions
Team Repository Template Repository Permissions

TeamSimple
Groups of organization members that gives permissions on specified repositories.

Thread
Thread

ThreadSubscription
Thread Subscription

Topic
A topic aggregates entities that are related to a subject.

TopicSearchResultItem
Topic Search Result Item

Traffic
Traffic

TreeItem
Tree Item

Uploader
Uploader

UserMarketplacePurchase
User Marketplace Purchase

UserSearchResultItem
User Search Result Item

ValidationError
Validation Error

ValidationErrorSimple
Validation Error Simple

VerificationSchema
Verification Schema

ViewTraffic
View Traffic

WeeksItem
Weeks Item

Workflow
A GitHub Actions workflow

WorkflowRun
An invocation of a workflow

WorkflowRunUsage
Workflow Run Usage

WorkflowRunUsageBillable
Workflow Run Usage Billable

WorkflowRunUsageBillableMacos
Workflow Run Usage Billable Macos

WorkflowRunUsageBillableUbuntu
Workflow Run Usage Billable Ubuntu

WorkflowRunUsageBillableWindows
Workflow Run Usage Billable Windows

WorkflowUsage
Workflow Usage

WorkflowUsageBillable
Workflow Usage Billable

WorkflowUsageBillableMacos
Workflow Usage Billable Macos

WorkflowUsageBillableUbuntu
Workflow Usage Billable Ubuntu

WorkflowUsageBillableWindows
Workflow Usage Billable Windows

Enums

AuthenticationTokenRepositorySelection
Describe whether all repositories have been selected or there's a selection involved

CheckRunStatus
The phase of the lifecycle that the check is currently in.

ClosedReason
Closed Reason

CommitComparisonStatus
Commit Comparison Status

DeploymentStatusState
The state of the status.

Direction
Direction

ImportStatus
Import Status

InstallationRepositorySelection
Describe whether all repositories have been selected or there's a selection involved

InstallationTokenRepositorySelection
Installation Token Repository Selection

JobStatus
The phase of the lifecycle that the job is currently in.

Limit
The interaction limit to enable.

MilestoneState
The state of the milestone.

Op
Op

Order
Order

OrganizationPermission
The baseline permission that all organization members have on this project

PageStatus
The status of the most recent build of the Page.

Per
Per

Privacy
The level of privacy this team should have

PullRequestReviewCommentSide
The side of the diff to which the comment applies. The side of the last line of the range for a multi-line comment

PullRequestReviewCommentStartSide
The side of the first line of the range for a multi-line comment.

PullRequestState
State of this Pull Request. Either open or closed.

ReactionContent
The reaction to use

ReleaseAssetState
State of the release asset.

RepositoryInvitationPermissions
The permission associated with the invitation.

ReviewCommentSide
The side of the first line of the range for a multi-line comment.

ReviewCommentStartSide
The side of the first line of the range for a multi-line comment.

Role
The role of the user in the team.

RuleSeverity
The severity of the alert.

ScopedInstallationRepositorySelection
Describe whether all repositories have been selected or there's a selection involved

Sort
Sort

StatusParam
Status Param

StepsItemStatus
The phase of the lifecycle that the job is currently in.

Visibility
Visibility of a secret

WorkflowRunStatus
Workflow Run Status

WorkflowState
Workflow State